\name{berndt.dat}
\alias{berndt.dat}
\docType{data}
\title{
monthly returns data on 16 assets over the five year period January 1983 through December 1987
}
\description{
monthly returns data on 16 assets over the five year period January 1983 through December 1987
}
\usage{data(berndt.dat)}
\format{
  A data frame with 120 observations on the following 17 variables.
  \describe{
    \item{\code{CITCRP}}{a numeric vector}
    \item{\code{CONED}}{a numeric vector}
    \item{\code{CONTIL}}{a numeric vector}
    \item{\code{DATGEN}}{a numeric vector}
    \item{\code{DEC}}{a numeric vector}
    \item{\code{DELTA}}{a numeric vector}
    \item{\code{GENMIL}}{a numeric vector}
    \item{\code{GERBER}}{a numeric vector}
    \item{\code{IBM}}{a numeric vector}
    \item{\code{MARKET}}{a numeric vector}
    \item{\code{MOBIL}}{a numeric vector}
    \item{\code{PANAM}}{a numeric vector}
    \item{\code{PSNH}}{a numeric vector}
    \item{\code{TANDY}}{a numeric vector}
    \item{\code{TEXACO}}{a numeric vector}
    \item{\code{WEYER}}{a numeric vector}
    \item{\code{RKFREE}}{a numeric vector}
  }
}
\source{
S+ FinMetric "berndt.dat".
}
\references{
S+ FinMetric "berndt.dat".
}
\examples{
data(berndt.dat)
## maybe str(berndt.dat) ; plot(berndt.dat) ...
}
\keyword{datasets}
