% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{black}
\alias{black}
\title{Monthly returns on 16 assets from January 1983 through December 1992}
\format{A \samp{xts} object on Jan 1983/Dec 1992 containing:
\describe{
  \item{BOISE}{a numeric vector}
  \item{CITCRP}{a numeric vector}
  \item{CONED}{a numeric vector}
  \item{CONTIL}{a numeric vector}
  \item{DATGEN}{a numeric vector}
  \item{DEC}{a numeric vector}
  \item{DELTA}{a numeric vector}
  \item{GENMIL}{a numeric vector}
  \item{GERBER}{a numeric vector}
  \item{IBM}{a numeric vector}
  \item{MARKET}{a numeric vector}
  \item{MOBIL}{a numeric vector}
  \item{PANAM}{a numeric vector}
  \item{PSNH}{a numeric vector}
  \item{TANDY}{a numeric vector}
  \item{TEXACO}{a numeric vector}
  \item{WEYER}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
black
}
\description{
A \samp{xts} object of monthly returns on 16 assets and risk free rates from January 1983 through
December 1992.
}
\keyword{datasets}

