\name{black.ts}
\alias{black.ts}
\docType{data}
\title{
monthly data on 16 assets over the five year period January 1983 through
December 1987
}
\description{
monthly data on 16 assets over the five year period January 1983 through
December 1987
}
\usage{data(black.ts)}
\format{
  A data frame with 120 observations on the following 17 variables.
  \describe{
    \item{\code{BOISE}}{a numeric vector}
    \item{\code{CITCRP}}{a numeric vector}
    \item{\code{CONED}}{a numeric vector}
    \item{\code{CONTIL}}{a numeric vector}
    \item{\code{DATGEN}}{a numeric vector}
    \item{\code{DEC}}{a numeric vector}
    \item{\code{DELTA}}{a numeric vector}
    \item{\code{GENMIL}}{a numeric vector}
    \item{\code{GERBER}}{a numeric vector}
    \item{\code{IBM}}{a numeric vector}
    \item{\code{MARKET}}{a numeric vector}
    \item{\code{MOBIL}}{a numeric vector}
    \item{\code{PANAM}}{a numeric vector}
    \item{\code{PSNH}}{a numeric vector}
    \item{\code{TANDY}}{a numeric vector}
    \item{\code{TEXACO}}{a numeric vector}
    \item{\code{WEYER}}{a numeric vector}
  }
}

\source{
S+ FinMetric "black.ts"
}
\references{
S+ FinMetric "black.ts"
}
\examples{
data(black.ts)
## maybe str(black.ts) ; plot(black.ts) ...
}
\keyword{datasets}
