% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consump}
\alias{consump}
\title{Annual comsumption data from 1960 to 1995}
\format{A \samp{xts} object on Jan 1959/Jan 1995 containing:
\describe{
  \item{GC}{a numeric vector}
  \item{GY}{a numeric vector}
  \item{R3}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
consump
}
\description{
A \samp{xts} object of annual comsumption data from 1960 to 1995 taken from Woolridge (2002).
}
\keyword{datasets}

