\name{consump.ts}
\alias{consump.ts}
\docType{data}
\title{
annual comsumption data over the period 1960 to 1995 taken from Wooldridge (2002)
}
\description{
annual comsumption data over the period 1960 to 1995 taken from Wooldridge (2002)
}
\usage{data(consump.ts)}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{RowNames}}{a numeric vector}
    \item{\code{GC}}{a numeric vector}
    \item{\code{GY}}{a numeric vector}
    \item{\code{R3}}{a numeric vector}
  }
}

\source{
S+FinMetrics comsump.ts.
}
\references{
Wooldrigde(2002)
}
\examples{
data(consump.ts)
## maybe str(consump.ts) ; plot(consump.ts) ...
}
\keyword{datasets}
