\name{CPI.dat}
\alias{CPI.dat}
\docType{data}
\title{
The CPI data from December 1990
through January 2001
}
\description{
The CPI data from December 1990
through January 2001
}
\usage{data(CPI.dat)}
\format{
  A data frame with 1067 observations on the following 2 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{Apr 1913} \code{Apr 1914} \code{Apr 1915} \code{Apr 1916} \code{Apr 1917} \code{Apr 1918} \code{Apr 1919} \code{Apr 1920} \code{Apr 1921} \code{Apr 1922} \code{Apr 1923} \code{Apr 1924} \code{Apr 1925} \code{Apr 1926} \code{Apr 1927} \code{Apr 1928} \code{Apr 1929} \code{Apr 1930} \code{Apr 1931} \code{Apr 1932} \code{Apr 1933} \code{Apr 1934} \code{Apr 1935} \code{Apr 1936} \code{Apr 1937} \code{Apr 1938} \code{Apr 1939} \code{Apr 1940} \code{Apr 1941} \code{Apr 1942} \code{Apr 1943} \code{Apr 1944} \code{Apr 1945} \code{Apr 1946} \code{Apr 1947} \code{Apr 1948} \code{Apr 1949} \code{Apr 1950} \code{Apr 1951} \code{Apr 1952} \code{Apr 1953} \code{Apr 1954} \code{Apr 1955} \code{Apr 1956} \code{Apr 1957} \code{Apr 1958} \code{Apr 1959} \code{Apr 1960} \code{Apr 1961} \code{Apr 1962} \code{Apr 1963} \code{Apr 1964} \code{Apr 1965} \code{Apr 1966} \code{Apr 1967} \code{Apr 1968} \code{Apr 1969} \code{Apr 1970} \code{Apr 1971} \code{Apr 1972} \code{Apr 1973} \code{Apr 1974} \code{Apr 1975} \code{Apr 1976} \code{Apr 1977} \code{Apr 1978} \code{Apr 1979} \code{Apr 1980} \code{Apr 1981} \code{Apr 1982} \code{Apr 1983} \code{Apr 1984} \code{Apr 1985} \code{Apr 1986} \code{Apr 1987} \code{Apr 1988} \code{Apr 1989} \code{Apr 1990} \code{Apr 1991} \code{Apr 1992} \code{Apr 1993} \code{Apr 1994} \code{Apr 1995} \code{Apr 1996} \code{Apr 1997} \code{Apr 1998} \code{Apr 1999} \code{Apr 2000} \code{Apr 2001} \code{Aug 1913} \code{Aug 1914} \code{Aug 1915} \code{Aug 1916} \code{Aug 1917} \code{Aug 1918} \code{Aug 1919} \code{Aug 1920} \code{Aug 1921} \code{Aug 1922} \code{Aug 1923} \code{Aug 1924} \code{Aug 1925} \code{Aug 1926} \code{Aug 1927} \code{Aug 1928} \code{Aug 1929} \code{Aug 1930} \code{Aug 1931} \code{Aug 1932} \code{Aug 1933} \code{Aug 1934} \code{Aug 1935} \code{Aug 1936} \code{Aug 1937} \code{Aug 1938} \code{Aug 1939} \code{Aug 1940} \code{Aug 1941} \code{Aug 1942} \code{Aug 1943} \code{Aug 1944} \code{Aug 1945} \code{Aug 1946} \code{Aug 1947} \code{Aug 1948} \code{Aug 1949} \code{Aug 1950} \code{Aug 1951} \code{Aug 1952} \code{Aug 1953} \code{Aug 1954} \code{Aug 1955} \code{Aug 1956} \code{Aug 1957} \code{Aug 1958} \code{Aug 1959} \code{Aug 1960} \code{Aug 1961} \code{Aug 1962} \code{Aug 1963} \code{Aug 1964} \code{Aug 1965} \code{Aug 1966} \code{Aug 1967} \code{Aug 1968} \code{Aug 1969} \code{Aug 1970} \code{Aug 1971} \code{Aug 1972} \code{Aug 1973} \code{Aug 1974} \code{Aug 1975} \code{Aug 1976} \code{Aug 1977} \code{Aug 1978} \code{Aug 1979} \code{Aug 1980} \code{Aug 1981} \code{Aug 1982} \code{Aug 1983} \code{Aug 1984} \code{Aug 1985} \code{Aug 1986} \code{Aug 1987} \code{Aug 1988} \code{Aug 1989} \code{Aug 1990} \code{Aug 1991} \code{Aug 1992} \code{Aug 1993} \code{Aug 1994} \code{Aug 1995} \code{Aug 1996} \code{Aug 1997} \code{Aug 1998} \code{Aug 1999} \code{Aug 2000} \code{Aug 2001} \code{Dec 1913} \code{Dec 1914} \code{Dec 1915} \code{Dec 1916} \code{Dec 1917} \code{Dec 1918} \code{Dec 1919} \code{Dec 1920} \code{Dec 1921} \code{Dec 1922} \code{Dec 1923} \code{Dec 1924} \code{Dec 1925} \code{Dec 1926} \code{Dec 1927} \code{Dec 1928} \code{Dec 1929} \code{Dec 1930} \code{Dec 1931} \code{Dec 1932} \code{Dec 1933} \code{Dec 1934} \code{Dec 1935} \code{Dec 1936} \code{Dec 1937} \code{Dec 1938} \code{Dec 1939} \code{Dec 1940} \code{Dec 1941} \code{Dec 1942} \code{Dec 1943} \code{Dec 1944} \code{Dec 1945} \code{Dec 1946} \code{Dec 1947} \code{Dec 1948} \code{Dec 1949} \code{Dec 1950} \code{Dec 1951} \code{Dec 1952} \code{Dec 1953} \code{Dec 1954} \code{Dec 1955} \code{Dec 1956} \code{Dec 1957} \code{Dec 1958} \code{Dec 1959} \code{Dec 1960} \code{Dec 1961} \code{Dec 1962} \code{Dec 1963} \code{Dec 1964} \code{Dec 1965} \code{Dec 1966} \code{Dec 1967} \code{Dec 1968} \code{Dec 1969} \code{Dec 1970} \code{Dec 1971} \code{Dec 1972} \code{Dec 1973} \code{Dec 1974} \code{Dec 1975} \code{Dec 1976} \code{Dec 1977} \code{Dec 1978} \code{Dec 1979} \code{Dec 1980} \code{Dec 1981} \code{Dec 1982} \code{Dec 1983} \code{Dec 1984} \code{Dec 1985} \code{Dec 1986} \code{Dec 1987} \code{Dec 1988} \code{Dec 1989} \code{Dec 1990} \code{Dec 1991} \code{Dec 1992} \code{Dec 1993} \code{Dec 1994} \code{Dec 1995} \code{Dec 1996} \code{Dec 1997} \code{Dec 1998} \code{Dec 1999} \code{Dec 2000} \code{Feb 1913} \code{Feb 1914} \code{Feb 1915} \code{Feb 1916} \code{Feb 1917} \code{Feb 1918} \code{Feb 1919} \code{Feb 1920} \code{Feb 1921} \code{Feb 1922} \code{Feb 1923} \code{Feb 1924} \code{Feb 1925} \code{Feb 1926} \code{Feb 1927} \code{Feb 1928} \code{Feb 1929} \code{Feb 1930} \code{Feb 1931} \code{Feb 1932} \code{Feb 1933} \code{Feb 1934} \code{Feb 1935} \code{Feb 1936} \code{Feb 1937} \code{Feb 1938} \code{Feb 1939} \code{Feb 1940} \code{Feb 1941} \code{Feb 1942} \code{Feb 1943} \code{Feb 1944} \code{Feb 1945} \code{Feb 1946} \code{Feb 1947} \code{Feb 1948} \code{Feb 1949} \code{Feb 1950} \code{Feb 1951} \code{Feb 1952} \code{Feb 1953} \code{Feb 1954} \code{Feb 1955} \code{Feb 1956} \code{Feb 1957} \code{Feb 1958} \code{Feb 1959} \code{Feb 1960} \code{Feb 1961} \code{Feb 1962} \code{Feb 1963} \code{Feb 1964} \code{Feb 1965} \code{Feb 1966} \code{Feb 1967} \code{Feb 1968} \code{Feb 1969} \code{Feb 1970} \code{Feb 1971} \code{Feb 1972} \code{Feb 1973} \code{Feb 1974} \code{Feb 1975} \code{Feb 1976} \code{Feb 1977} \code{Feb 1978} \code{Feb 1979} \code{Feb 1980} \code{Feb 1981} \code{Feb 1982} \code{Feb 1983} \code{Feb 1984} \code{Feb 1985} \code{Feb 1986} \code{Feb 1987} \code{Feb 1988} \code{Feb 1989} \code{Feb 1990} \code{Feb 1991} \code{Feb 1992} \code{Feb 1993} \code{Feb 1994} \code{Feb 1995} \code{Feb 1996} \code{Feb 1997} \code{Feb 1998} \code{Feb 1999} \code{Feb 2000} \code{Feb 2001} \code{Jan 1913} \code{Jan 1914} \code{Jan 1915} \code{Jan 1916} \code{Jan 1917} \code{Jan 1918} \code{Jan 1919} \code{Jan 1920} \code{Jan 1921} \code{Jan 1922} \code{Jan 1923} \code{Jan 1924} \code{Jan 1925} \code{Jan 1926} \code{Jan 1927} \code{Jan 1928} \code{Jan 1929} \code{Jan 1930} \code{Jan 1931} \code{Jan 1932} \code{Jan 1933} \code{Jan 1934} \code{Jan 1935} \code{Jan 1936} \code{Jan 1937} \code{Jan 1938} \code{Jan 1939} \code{Jan 1940} \code{Jan 1941} \code{Jan 1942} \code{Jan 1943} \code{Jan 1944} \code{Jan 1945} \code{Jan 1946} \code{Jan 1947} \code{Jan 1948} \code{Jan 1949} \code{Jan 1950} \code{Jan 1951} \code{Jan 1952} \code{Jan 1953} \code{Jan 1954} \code{Jan 1955} \code{Jan 1956} \code{Jan 1957} \code{Jan 1958} \code{Jan 1959} \code{Jan 1960} \code{Jan 1961} \code{Jan 1962} \code{Jan 1963} \code{Jan 1964} \code{Jan 1965} \code{Jan 1966} \code{Jan 1967} \code{Jan 1968} \code{Jan 1969} \code{Jan 1970} \code{Jan 1971} \code{Jan 1972} \code{Jan 1973} \code{Jan 1974} \code{Jan 1975} \code{Jan 1976} \code{Jan 1977} \code{Jan 1978} \code{Jan 1979} \code{Jan 1980} \code{Jan 1981} \code{Jan 1982} \code{Jan 1983} \code{Jan 1984} \code{Jan 1985} \code{Jan 1986} \code{Jan 1987} \code{Jan 1988} \code{Jan 1989} \code{Jan 1990} \code{Jan 1991} \code{Jan 1992} \code{Jan 1993} \code{Jan 1994} \code{Jan 1995} \code{Jan 1996} \code{Jan 1997} \code{Jan 1998} \code{Jan 1999} \code{Jan 2000} \code{Jan 2001} \code{Jul 1913} \code{Jul 1914} \code{Jul 1915} \code{Jul 1916} \code{Jul 1917} \code{Jul 1918} \code{Jul 1919} \code{Jul 1920} \code{Jul 1921} \code{Jul 1922} \code{Jul 1923} \code{Jul 1924} \code{Jul 1925} \code{Jul 1926} \code{Jul 1927} \code{Jul 1928} \code{Jul 1929} \code{Jul 1930} \code{Jul 1931} \code{Jul 1932} \code{Jul 1933} \code{Jul 1934} \code{Jul 1935} \code{Jul 1936} \code{Jul 1937} \code{Jul 1938} \code{Jul 1939} \code{Jul 1940} \code{Jul 1941} \code{Jul 1942} \code{Jul 1943} \code{Jul 1944} \code{Jul 1945} \code{Jul 1946} \code{Jul 1947} \code{Jul 1948} \code{Jul 1949} \code{Jul 1950} \code{Jul 1951} \code{Jul 1952} \code{Jul 1953} \code{Jul 1954} \code{Jul 1955} \code{Jul 1956} \code{Jul 1957} \code{Jul 1958} \code{Jul 1959} \code{Jul 1960} \code{Jul 1961} \code{Jul 1962} \code{Jul 1963} \code{Jul 1964} \code{Jul 1965} \code{Jul 1966} \code{Jul 1967} \code{Jul 1968} \code{Jul 1969} \code{Jul 1970} \code{Jul 1971} \code{Jul 1972} \code{Jul 1973} \code{Jul 1974} \code{Jul 1975} \code{Jul 1976} \code{Jul 1977} \code{Jul 1978} \code{Jul 1979} \code{Jul 1980} \code{Jul 1981} \code{Jul 1982} \code{Jul 1983} \code{Jul 1984} \code{Jul 1985} \code{Jul 1986} \code{Jul 1987} \code{Jul 1988} \code{Jul 1989} \code{Jul 1990} \code{Jul 1991} \code{Jul 1992} \code{Jul 1993} \code{Jul 1994} \code{Jul 1995} \code{Jul 1996} \code{Jul 1997} \code{Jul 1998} \code{Jul 1999} \code{Jul 2000} \code{Jul 2001} \code{Jun 1913} \code{Jun 1914} \code{Jun 1915} \code{Jun 1916} \code{Jun 1917} \code{Jun 1918} \code{Jun 1919} \code{Jun 1920} \code{Jun 1921} \code{Jun 1922} \code{Jun 1923} \code{Jun 1924} \code{Jun 1925} \code{Jun 1926} \code{Jun 1927} \code{Jun 1928} \code{Jun 1929} \code{Jun 1930} \code{Jun 1931} \code{Jun 1932} \code{Jun 1933} \code{Jun 1934} \code{Jun 1935} \code{Jun 1936} \code{Jun 1937} \code{Jun 1938} \code{Jun 1939} \code{Jun 1940} \code{Jun 1941} \code{Jun 1942} \code{Jun 1943} \code{Jun 1944} \code{Jun 1945} \code{Jun 1946} \code{Jun 1947} \code{Jun 1948} \code{Jun 1949} \code{Jun 1950} \code{Jun 1951} \code{Jun 1952} \code{Jun 1953} \code{Jun 1954} \code{Jun 1955} \code{Jun 1956} \code{Jun 1957} \code{Jun 1958} \code{Jun 1959} \code{Jun 1960} \code{Jun 1961} \code{Jun 1962} \code{Jun 1963} \code{Jun 1964} \code{Jun 1965} \code{Jun 1966} \code{Jun 1967} \code{Jun 1968} \code{Jun 1969} \code{Jun 1970} \code{Jun 1971} \code{Jun 1972} \code{Jun 1973} \code{Jun 1974} \code{Jun 1975} \code{Jun 1976} \code{Jun 1977} \code{Jun 1978} \code{Jun 1979} \code{Jun 1980} \code{Jun 1981} \code{Jun 1982} \code{Jun 1983} \code{Jun 1984} \code{Jun 1985} \code{Jun 1986} \code{Jun 1987} \code{Jun 1988} \code{Jun 1989} \code{Jun 1990} \code{Jun 1991} \code{Jun 1992} \code{Jun 1993} \code{Jun 1994} \code{Jun 1995} \code{Jun 1996} \code{Jun 1997} \code{Jun 1998} \code{Jun 1999} \code{Jun 2000} \code{Jun 2001} \code{Mar 1913} \code{Mar 1914} \code{Mar 1915} \code{Mar 1916} \code{Mar 1917} \code{Mar 1918} \code{Mar 1919} \code{Mar 1920} \code{Mar 1921} \code{Mar 1922} \code{Mar 1923} \code{Mar 1924} \code{Mar 1925} \code{Mar 1926} \code{Mar 1927} \code{Mar 1928} \code{Mar 1929} \code{Mar 1930} \code{Mar 1931} \code{Mar 1932} \code{Mar 1933} \code{Mar 1934} \code{Mar 1935} \code{Mar 1936} \code{Mar 1937} \code{Mar 1938} \code{Mar 1939} \code{Mar 1940} \code{Mar 1941} \code{Mar 1942} \code{Mar 1943} \code{Mar 1944} \code{Mar 1945} \code{Mar 1946} \code{Mar 1947} \code{Mar 1948} \code{Mar 1949} \code{Mar 1950} \code{Mar 1951} \code{Mar 1952} \code{Mar 1953} \code{Mar 1954} \code{Mar 1955} \code{Mar 1956} \code{Mar 1957} \code{Mar 1958} \code{Mar 1959} \code{Mar 1960} \code{Mar 1961} \code{Mar 1962} \code{Mar 1963} \code{Mar 1964} \code{Mar 1965} \code{Mar 1966} \code{Mar 1967} \code{Mar 1968} \code{Mar 1969} \code{Mar 1970} \code{Mar 1971} \code{Mar 1972} \code{Mar 1973} \code{Mar 1974} \code{Mar 1975} \code{Mar 1976} \code{Mar 1977} \code{Mar 1978} \code{Mar 1979} \code{Mar 1980} \code{Mar 1981} \code{Mar 1982} \code{Mar 1983} \code{Mar 1984} \code{Mar 1985} \code{Mar 1986} \code{Mar 1987} \code{Mar 1988} \code{Mar 1989} \code{Mar 1990} \code{Mar 1991} \code{Mar 1992} \code{Mar 1993} \code{Mar 1994} \code{Mar 1995} \code{Mar 1996} \code{Mar 1997} \code{Mar 1998} \code{Mar 1999} \code{Mar 2000} \code{Mar 2001} \code{May 1913} \code{May 1914} \code{May 1915} \code{May 1916} \code{May 1917} \code{May 1918} \code{May 1919} \code{May 1920} \code{May 1921} \code{May 1922} \code{May 1923} \code{May 1924} \code{May 1925} \code{May 1926} \code{May 1927} \code{May 1928} \code{May 1929} \code{May 1930} \code{May 1931} \code{May 1932} \code{May 1933} \code{May 1934} \code{May 1935} \code{May 1936} \code{May 1937} \code{May 1938} \code{May 1939} \code{May 1940} \code{May 1941} \code{May 1942} \code{May 1943} \code{May 1944} \code{May 1945} \code{May 1946} \code{May 1947} \code{May 1948} \code{May 1949} \code{May 1950} \code{May 1951} \code{May 1952} \code{May 1953} \code{May 1954} \code{May 1955} \code{May 1956} \code{May 1957} \code{May 1958} \code{May 1959} \code{May 1960} \code{May 1961} \code{May 1962} \code{May 1963} \code{May 1964} \code{May 1965} \code{May 1966} \code{May 1967} \code{May 1968} \code{May 1969} \code{May 1970} \code{May 1971} \code{May 1972} \code{May 1973} \code{May 1974} \code{May 1975} \code{May 1976} \code{May 1977} \code{May 1978} \code{May 1979} \code{May 1980} \code{May 1981} \code{May 1982} \code{May 1983} \code{May 1984} \code{May 1985} \code{May 1986} \code{May 1987} \code{May 1988} \code{May 1989} \code{May 1990} \code{May 1991} \code{May 1992} \code{May 1993} \code{May 1994} \code{May 1995} \code{May 1996} \code{May 1997} \code{May 1998} \code{May 1999} \code{May 2000} \code{May 2001} \code{Nov 1913} \code{Nov 1914} \code{Nov 1915} \code{Nov 1916} \code{Nov 1917} \code{Nov 1918} \code{Nov 1919} \code{Nov 1920} \code{Nov 1921} \code{Nov 1922} \code{Nov 1923} \code{Nov 1924} \code{Nov 1925} \code{Nov 1926} \code{Nov 1927} \code{Nov 1928} \code{Nov 1929} \code{Nov 1930} \code{Nov 1931} \code{Nov 1932} \code{Nov 1933} \code{Nov 1934} \code{Nov 1935} \code{Nov 1936} \code{Nov 1937} \code{Nov 1938} \code{Nov 1939} \code{Nov 1940} \code{Nov 1941} \code{Nov 1942} \code{Nov 1943} \code{Nov 1944} \code{Nov 1945} \code{Nov 1946} \code{Nov 1947} \code{Nov 1948} \code{Nov 1949} \code{Nov 1950} \code{Nov 1951} \code{Nov 1952} \code{Nov 1953} \code{Nov 1954} \code{Nov 1955} \code{Nov 1956} \code{Nov 1957} \code{Nov 1958} \code{Nov 1959} \code{Nov 1960} \code{Nov 1961} \code{Nov 1962} \code{Nov 1963} \code{Nov 1964} \code{Nov 1965} \code{Nov 1966} \code{Nov 1967} \code{Nov 1968} \code{Nov 1969} \code{Nov 1970} \code{Nov 1971} \code{Nov 1972} \code{Nov 1973} \code{Nov 1974} \code{Nov 1975} \code{Nov 1976} \code{Nov 1977} \code{Nov 1978} \code{Nov 1979} \code{Nov 1980} \code{Nov 1981} \code{Nov 1982} \code{Nov 1983} \code{Nov 1984} \code{Nov 1985} \code{Nov 1986} \code{Nov 1987} \code{Nov 1988} \code{Nov 1989} \code{Nov 1990} \code{Nov 1991} \code{Nov 1992} \code{Nov 1993} \code{Nov 1994} \code{Nov 1995} \code{Nov 1996} \code{Nov 1997} \code{Nov 1998} \code{Nov 1999} \code{Nov 2000} \code{Nov 2001} \code{Oct 1913} \code{Oct 1914} \code{Oct 1915} \code{Oct 1916} \code{Oct 1917} \code{Oct 1918} \code{Oct 1919} \code{Oct 1920} \code{Oct 1921} \code{Oct 1922} \code{Oct 1923} \code{Oct 1924} \code{Oct 1925} \code{Oct 1926} \code{Oct 1927} \code{Oct 1928} \code{Oct 1929} \code{Oct 1930} \code{Oct 1931} \code{Oct 1932} \code{Oct 1933} \code{Oct 1934} \code{Oct 1935} \code{Oct 1936} \code{Oct 1937} \code{Oct 1938} \code{Oct 1939} \code{Oct 1940} \code{Oct 1941} \code{Oct 1942} \code{Oct 1943} \code{Oct 1944} \code{Oct 1945} \code{Oct 1946} \code{Oct 1947} \code{Oct 1948} \code{Oct 1949} \code{Oct 1950} \code{Oct 1951} \code{Oct 1952} \code{Oct 1953} \code{Oct 1954} \code{Oct 1955} \code{Oct 1956} \code{Oct 1957} \code{Oct 1958} \code{Oct 1959} \code{Oct 1960} \code{Oct 1961} \code{Oct 1962} \code{Oct 1963} \code{Oct 1964} \code{Oct 1965} \code{Oct 1966} \code{Oct 1967} \code{Oct 1968} \code{Oct 1969} \code{Oct 1970} \code{Oct 1971} \code{Oct 1972} \code{Oct 1973} \code{Oct 1974} \code{Oct 1975} \code{Oct 1976} \code{Oct 1977} \code{Oct 1978} \code{Oct 1979} \code{Oct 1980} \code{Oct 1981} \code{Oct 1982} \code{Oct 1983} \code{Oct 1984} \code{Oct 1985} \code{Oct 1986} \code{Oct 1987} \code{Oct 1988} \code{Oct 1989} \code{Oct 1990} \code{Oct 1991} \code{Oct 1992} \code{Oct 1993} \code{Oct 1994} \code{Oct 1995} \code{Oct 1996} \code{Oct 1997} \code{Oct 1998} \code{Oct 1999} \code{Oct 2000} \code{Oct 2001} \code{Sep 1913} \code{Sep 1914} \code{Sep 1915} \code{Sep 1916} \code{Sep 1917} \code{Sep 1918} \code{Sep 1919} \code{Sep 1920} \code{Sep 1921} \code{Sep 1922} \code{Sep 1923} \code{Sep 1924} \code{Sep 1925} \code{Sep 1926} \code{Sep 1927} \code{Sep 1928} \code{Sep 1929} \code{Sep 1930} \code{Sep 1931} \code{Sep 1932} \code{Sep 1933} \code{Sep 1934} \code{Sep 1935} \code{Sep 1936} \code{Sep 1937} \code{Sep 1938} \code{Sep 1939} \code{Sep 1940} \code{Sep 1941} \code{Sep 1942} \code{Sep 1943} \code{Sep 1944} \code{Sep 1945} \code{Sep 1946} \code{Sep 1947} \code{Sep 1948} \code{Sep 1949} \code{Sep 1950} \code{Sep 1951} \code{Sep 1952} \code{Sep 1953} \code{Sep 1954} \code{Sep 1955} \code{Sep 1956} \code{Sep 1957} \code{Sep 1958} \code{Sep 1959} \code{Sep 1960} \code{Sep 1961} \code{Sep 1962} \code{Sep 1963} \code{Sep 1964} \code{Sep 1965} \code{Sep 1966} \code{Sep 1967} \code{Sep 1968} \code{Sep 1969} \code{Sep 1970} \code{Sep 1971} \code{Sep 1972} \code{Sep 1973} \code{Sep 1974} \code{Sep 1975} \code{Sep 1976} \code{Sep 1977} \code{Sep 1978} \code{Sep 1979} \code{Sep 1980} \code{Sep 1981} \code{Sep 1982} \code{Sep 1983} \code{Sep 1984} \code{Sep 1985} \code{Sep 1986} \code{Sep 1987} \code{Sep 1988} \code{Sep 1989} \code{Sep 1990} \code{Sep 1991} \code{Sep 1992} \code{Sep 1993} \code{Sep 1994} \code{Sep 1995} \code{Sep 1996} \code{Sep 1997} \code{Sep 1998} \code{Sep 1999} \code{Sep 2000} \code{Sep 2001}}
    \item{\code{CPI}}{a numeric vector}
  }
}

\source{
S+ FinMetrics CPI.dat file.
}
\references{
S+ FinMetrics CPI.dat file.
}
\examples{
data(CPI.dat)
## maybe str(CPI.dat) ; plot(CPI.dat) ...
}
\keyword{datasets}
