% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CPI}
\alias{CPI}
\title{Monthly CPI data from January 1913 through November 2001}
\format{A \samp{xts} object on Jan 1913/Nov 2001 containing:
\describe{
  \item{CPI}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
CPI
}
\description{
A \samp{xts} object of monthly CPI data from January 1913 through November 2001.
}
\keyword{datasets}

