% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danish}
\alias{danish}
\title{Daily Danish fire loss data}
\format{A \samp{xts} object on 1980-01-03/1990-12-31 containing:
\describe{
  \item{danish}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
danish
}
\description{
A \samp{xts} object of daily Danish fire loss data.
}
\keyword{datasets}

