\name{dell.s}
\alias{dell.s}
\docType{data}
\title{
time series of daily stock returns of Dell Computer Corporation
}
\description{
time series of daily stock returns of Dell Computer Corporation from 1/10/1995 to 9/9/1997
}
\usage{data(dell.s)}
\format{
  A data frame with 1261 observations on the following 2 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{1/10/1994} \code{1/10/1995} \code{1/10/1996} \code{1/10/1997} \code{1/11/1994} \code{1/11/1995} \code{1/11/1996} \code{1/12/1994} \code{1/12/1995} \code{1/12/1996} \code{1/12/1998} \code{1/13/1994} \code{1/13/1995} \code{1/13/1997} \code{1/13/1998} \code{1/14/1994} \code{1/14/1997} \code{1/14/1998} \code{1/15/1996} \code{1/15/1997} \code{1/15/1998} \code{1/16/1995} \code{1/16/1996} \code{1/16/1997} \code{1/16/1998} \code{1/17/1994} \code{1/17/1995} \code{1/17/1996} \code{1/17/1997} \code{1/18/1994} \code{1/18/1995} \code{1/18/1996} \code{1/19/1994} \code{1/19/1995} \code{1/19/1996} \code{1/2/1996} \code{1/2/1997} \code{1/2/1998} \code{1/20/1994} \code{1/20/1995} \code{1/20/1997} \code{1/20/1998} \code{1/21/1994} \code{1/21/1997} \code{1/21/1998} \code{1/22/1996} \code{1/22/1997} \code{1/22/1998} \code{1/23/1995} \code{1/23/1996} \code{1/23/1997} \code{1/23/1998} \code{1/24/1994} \code{1/24/1995} \code{1/24/1996} \code{1/24/1997} \code{1/25/1994} \code{1/25/1995} \code{1/25/1996} \code{1/26/1994} \code{1/26/1995} \code{1/26/1996} \code{1/26/1998} \code{1/27/1994} \code{1/27/1995} \code{1/27/1997} \code{1/27/1998} \code{1/28/1994} \code{1/28/1997} \code{1/28/1998} \code{1/29/1996} \code{1/29/1997} \code{1/29/1998} \code{1/3/1994} \code{1/3/1995} \code{1/3/1996} \code{1/3/1997} \code{1/30/1995} \code{1/30/1996} \code{1/30/1997} \code{1/30/1998} \code{1/31/1994} \code{1/31/1995} \code{1/31/1996} \code{1/31/1997} \code{1/4/1994} \code{1/4/1995} \code{1/4/1996} \code{1/5/1994} \code{1/5/1995} \code{1/5/1996} \code{1/5/1998} \code{1/6/1994} \code{1/6/1995} \code{1/6/1997} \code{1/6/1998} \code{1/7/1994} \code{1/7/1997} \code{1/7/1998} \code{1/8/1996} \code{1/8/1997} \code{1/8/1998} \code{1/9/1995} \code{1/9/1996} \code{1/9/1997} \code{1/9/1998} \code{10/1/1993} \code{10/1/1996} \code{10/1/1997} \code{10/10/1994} \code{10/10/1995} \code{10/10/1996} \code{10/10/1997} \code{10/11/1993} \code{10/11/1994} \code{10/11/1995} \code{10/11/1996} \code{10/12/1993} \code{10/12/1994} \code{10/12/1995} \code{10/13/1993} \code{10/13/1994} \code{10/13/1995} \code{10/13/1997} \code{10/14/1993} \code{10/14/1994} \code{10/14/1996} \code{10/14/1997} \code{10/15/1993} \code{10/15/1996} \code{10/15/1997} \code{10/16/1995} \code{10/16/1996} \code{10/16/1997} \code{10/17/1994} \code{10/17/1995} \code{10/17/1996} \code{10/17/1997} \code{10/18/1993} \code{10/18/1994} \code{10/18/1995} \code{10/18/1996} \code{10/19/1993} \code{10/19/1994} \code{10/19/1995} \code{10/2/1995} \code{10/2/1996} \code{10/2/1997} \code{10/20/1993} \code{10/20/1994} \code{10/20/1995} \code{10/20/1997} \code{10/21/1993} \code{10/21/1994} \code{10/21/1996} \code{10/21/1997} \code{10/22/1993} \code{10/22/1996} \code{10/22/1997} \code{10/23/1995} \code{10/23/1996} \code{10/23/1997} \code{10/24/1994} \code{10/24/1995} \code{10/24/1996} \code{10/24/1997} \code{10/25/1993} \code{10/25/1994} \code{10/25/1995} \code{10/25/1996} \code{10/26/1993} \code{10/26/1994} \code{10/26/1995} \code{10/27/1993} \code{10/27/1994} \code{10/27/1995} \code{10/27/1997} \code{10/28/1993} \code{10/28/1994} \code{10/28/1996} \code{10/28/1997} \code{10/29/1993} \code{10/29/1996} \code{10/29/1997} \code{10/3/1994} \code{10/3/1995} \code{10/3/1996} \code{10/3/1997} \code{10/30/1995} \code{10/30/1996} \code{10/30/1997} \code{10/31/1994} \code{10/31/1995} \code{10/31/1996} \code{10/31/1997} \code{10/4/1993} \code{10/4/1994} \code{10/4/1995} \code{10/4/1996} \code{10/5/1993} \code{10/5/1994} \code{10/5/1995} \code{10/6/1993} \code{10/6/1994} \code{10/6/1995} \code{10/6/1997} \code{10/7/1993} \code{10/7/1994} \code{10/7/1996} \code{10/7/1997} \code{10/8/1993} \code{10/8/1996} \code{10/8/1997} \code{10/9/1995} \code{10/9/1996} \code{10/9/1997} \code{11/1/1993} \code{11/1/1994} \code{11/1/1995} \code{11/1/1996} \code{11/10/1993} \code{11/10/1994} \code{11/10/1995} \code{11/10/1997} \code{11/11/1993} \code{11/11/1994} \code{11/11/1996} \code{11/11/1997} \code{11/12/1993} \code{11/12/1996} \code{11/12/1997} \code{11/13/1995} \code{11/13/1996} \code{11/13/1997} \code{11/14/1994} \code{11/14/1995} \code{11/14/1996} \code{11/14/1997} \code{11/15/1993} \code{11/15/1994} \code{11/15/1995} \code{11/15/1996} \code{11/16/1993} \code{11/16/1994} \code{11/16/1995} \code{11/17/1993} \code{11/17/1994} \code{11/17/1995} \code{11/17/1997} \code{11/18/1993} \code{11/18/1994} \code{11/18/1996} \code{11/18/1997} \code{11/19/1993} \code{11/19/1996} \code{11/19/1997} \code{11/2/1993} \code{11/2/1994} \code{11/2/1995} \code{11/20/1995} \code{11/20/1996} \code{11/20/1997} \code{11/21/1994} \code{11/21/1995} \code{11/21/1996} \code{11/21/1997} \code{11/22/1993} \code{11/22/1994} \code{11/22/1995} \code{11/22/1996} \code{11/23/1993} \code{11/23/1994} \code{11/24/1993} \code{11/24/1995} \code{11/24/1997} \code{11/25/1994} \code{11/25/1996} \code{11/25/1997} \code{11/26/1993} \code{11/26/1996} \code{11/26/1997} \code{11/27/1995} \code{11/27/1996} \code{11/28/1994} \code{11/28/1995} \code{11/28/1997} \code{11/29/1993} \code{11/29/1994} \code{11/29/1995} \code{11/29/1996} \code{11/3/1993} \code{11/3/1994} \code{11/3/1995} \code{11/3/1997} \code{11/30/1993} \code{11/30/1994} \code{11/30/1995} \code{11/4/1993} \code{11/4/1994} \code{11/4/1996} \code{11/4/1997} \code{11/5/1993} \code{11/5/1996} \code{11/5/1997} \code{11/6/1995} \code{11/6/1996} \code{11/6/1997} \code{11/7/1994} \code{11/7/1995} \code{11/7/1996} \code{11/7/1997} \code{11/8/1993} \code{11/8/1994} \code{11/8/1995} \code{11/8/1996} \code{11/9/1993} \code{11/9/1994} \code{11/9/1995} \code{12/1/1993} \code{12/1/1994} \code{12/1/1995} \code{12/1/1997} \code{12/10/1993} \code{12/10/1996} \code{12/10/1997} \code{12/11/1995} \code{12/11/1996} \code{12/11/1997} \code{12/12/1994} \code{12/12/1995} \code{12/12/1996} \code{12/12/1997} \code{12/13/1993} \code{12/13/1994} \code{12/13/1995} \code{12/13/1996} \code{12/14/1993} \code{12/14/1994} \code{12/14/1995} \code{12/15/1993} \code{12/15/1994} \code{12/15/1995} \code{12/15/1997} \code{12/16/1993} \code{12/16/1994} \code{12/16/1996} \code{12/16/1997} \code{12/17/1993} \code{12/17/1996} \code{12/17/1997} \code{12/18/1995} \code{12/18/1996} \code{12/18/1997} \code{12/19/1994} \code{12/19/1995} \code{12/19/1996} \code{12/19/1997} \code{12/2/1993} \code{12/2/1994} \code{12/2/1996} \code{12/2/1997} \code{12/20/1993} \code{12/20/1994} \code{12/20/1995} \code{12/20/1996} \code{12/21/1993} \code{12/21/1994} \code{12/21/1995} \code{12/22/1993} \code{12/22/1994} \code{12/22/1995} \code{12/22/1997} \code{12/23/1993} \code{12/23/1994} \code{12/23/1996} \code{12/23/1997} \code{12/24/1996} \code{12/24/1997} \code{12/26/1995} \code{12/26/1996} \code{12/26/1997} \code{12/27/1993} \code{12/27/1994} \code{12/27/1995} \code{12/27/1996} \code{12/28/1993} \code{12/28/1994} \code{12/28/1995} \code{12/29/1993} \code{12/29/1994} \code{12/29/1995} \code{12/29/1997} \code{12/3/1993} \code{12/3/1996} \code{12/3/1997} \code{12/30/1993} \code{12/30/1994} \code{12/30/1996} \code{12/30/1997} \code{12/31/1993} \code{12/31/1996} \code{12/31/1997} \code{12/4/1995} \code{12/4/1996} \code{12/4/1997} \code{12/5/1994} \code{12/5/1995} \code{12/5/1996} \code{12/5/1997} \code{12/6/1993} \code{12/6/1994} \code{12/6/1995} \code{12/6/1996} \code{12/7/1993} \code{12/7/1994} \code{12/7/1995} \code{12/8/1993} \code{12/8/1994} \code{12/8/1995} \code{12/8/1997} \code{12/9/1993} \code{12/9/1994} \code{12/9/1996} \code{12/9/1997} \code{2/1/1994} \code{2/1/1995} \code{2/1/1996} \code{2/10/1994} \code{2/10/1995} \code{2/10/1997} \code{2/10/1998} \code{2/11/1994} \code{2/11/1997} \code{2/11/1998} \code{2/12/1996} \code{2/12/1997} \code{2/12/1998} \code{2/13/1995} \code{2/13/1996} \code{2/13/1997} \code{2/13/1998} \code{2/14/1994} \code{2/14/1995} \code{2/14/1996} \code{2/14/1997} \code{2/15/1994} \code{2/15/1995} \code{2/15/1996} \code{2/16/1994} \code{2/16/1995} \code{2/16/1996} \code{2/17/1994} \code{2/17/1995} \code{2/17/1998} \code{2/18/1994} \code{2/18/1997} \code{2/18/1998} \code{2/19/1997} \code{2/19/1998} \code{2/2/1994} \code{2/2/1995} \code{2/2/1996} \code{2/2/1998} \code{2/20/1996} \code{2/20/1997} \code{2/20/1998} \code{2/21/1995} \code{2/21/1996} \code{2/21/1997} \code{2/22/1994} \code{2/22/1995} \code{2/22/1996} \code{2/23/1994} \code{2/23/1995} \code{2/23/1996} \code{2/23/1998} \code{2/24/1994} \code{2/24/1995} \code{2/24/1997} \code{2/24/1998} \code{2/25/1994} \code{2/25/1997} \code{2/25/1998} \code{2/26/1996} \code{2/26/1997} \code{2/26/1998} \code{2/27/1995} \code{2/27/1996} \code{2/27/1997} \code{2/27/1998} \code{2/28/1994} \code{2/28/1995} \code{2/28/1996} \code{2/28/1997} \code{2/29/1996} \code{2/3/1994} \code{2/3/1995} \code{2/3/1997} \code{2/3/1998} \code{2/4/1994} \code{2/4/1997} \code{2/4/1998} \code{2/5/1996} \code{2/5/1997} \code{2/5/1998} \code{2/6/1995} \code{2/6/1996} \code{2/6/1997} \code{2/6/1998} \code{2/7/1994} \code{2/7/1995} \code{2/7/1996} \code{2/7/1997} \code{2/8/1994} \code{2/8/1995} \code{2/8/1996} \code{2/9/1994} \code{2/9/1995} \code{2/9/1996} \code{2/9/1998} \code{3/1/1994} \code{3/1/1995} \code{3/1/1996} \code{3/10/1994} \code{3/10/1995} \code{3/10/1997} \code{3/10/1998} \code{3/11/1994} \code{3/11/1996} \code{3/11/1997} \code{3/11/1998} \code{3/12/1996} \code{3/12/1997} \code{3/12/1998} \code{3/13/1995} \code{3/13/1996} \code{3/13/1997} \code{3/13/1998} \code{3/14/1994} \code{3/14/1995} \code{3/14/1996} \code{3/14/1997} \code{3/15/1994} \code{3/15/1995} \code{3/15/1996} \code{3/16/1994} \code{3/16/1995} \code{3/16/1998} \code{3/17/1994} \code{3/17/1995} \code{3/17/1997} \code{3/17/1998} \code{3/18/1994} \code{3/18/1996} \code{3/18/1997} \code{3/18/1998} \code{3/19/1996} \code{3/19/1997} \code{3/19/1998} \code{3/2/1994} \code{3/2/1995} \code{3/2/1998} \code{3/20/1995} \code{3/20/1996} \code{3/20/1997} \code{3/20/1998} \code{3/21/1994} \code{3/21/1995} \code{3/21/1996} \code{3/21/1997} \code{3/22/1994} \code{3/22/1995} \code{3/22/1996} \code{3/23/1994} \code{3/23/1995} \code{3/23/1998} \code{3/24/1994} \code{3/24/1995} \code{3/24/1997} \code{3/24/1998} \code{3/25/1994} \code{3/25/1996} \code{3/25/1997} \code{3/25/1998} \code{3/26/1996} \code{3/26/1997} \code{3/26/1998} \code{3/27/1995} \code{3/27/1996} \code{3/27/1997} \code{3/27/1998} \code{3/28/1994} \code{3/28/1995} \code{3/28/1996} \code{3/29/1994} \code{3/29/1995} \code{3/29/1996} \code{3/3/1994} \code{3/3/1995} \code{3/3/1997} \code{3/3/1998} \code{3/30/1994} \code{3/30/1995} \code{3/30/1998} \code{3/31/1994} \code{3/31/1995} \code{3/31/1997} \code{3/31/1998} \code{3/4/1994} \code{3/4/1996} \code{3/4/1997} \code{3/4/1998} \code{3/5/1996} \code{3/5/1997} \code{3/5/1998} \code{3/6/1995} \code{3/6/1996} \code{3/6/1997} \code{3/6/1998} \code{3/7/1994} \code{3/7/1995} \code{3/7/1996} \code{3/7/1997} \code{3/8/1994} \code{3/8/1995} \code{3/8/1996} \code{3/9/1994} \code{3/9/1995} \code{3/9/1998} \code{4/1/1996} \code{4/1/1997} \code{4/1/1998} \code{4/10/1995} \code{4/10/1996} \code{4/10/1997} \code{4/11/1994} \code{4/11/1995} \code{4/11/1996} \code{4/11/1997} \code{4/12/1994} \code{4/12/1995} \code{4/12/1996} \code{4/13/1994} \code{4/13/1995} \code{4/13/1998} \code{4/14/1994} \code{4/14/1997} \code{4/14/1998} \code{4/15/1994} \code{4/15/1996} \code{4/15/1997} \code{4/15/1998} \code{4/16/1996} \code{4/16/1997} \code{4/16/1998} \code{4/17/1995} \code{4/17/1996} \code{4/17/1997} \code{4/17/1998} \code{4/18/1994} \code{4/18/1995} \code{4/18/1996} \code{4/18/1997} \code{4/19/1994} \code{4/19/1995} \code{4/19/1996} \code{4/2/1996} \code{4/2/1997} \code{4/2/1998} \code{4/20/1994} \code{4/20/1995} \code{4/20/1998} \code{4/21/1994} \code{4/21/1995} \code{4/21/1997} \code{4/21/1998} \code{4/22/1994} \code{4/22/1996} \code{4/22/1997} \code{4/22/1998} \code{4/23/1996} \code{4/23/1997} \code{4/23/1998} \code{4/24/1995} \code{4/24/1996} \code{4/24/1997} \code{4/24/1998} \code{4/25/1994} \code{4/25/1995} \code{4/25/1996} \code{4/25/1997} \code{4/26/1994} \code{4/26/1995} \code{4/26/1996} \code{4/27/1995} \code{4/27/1998} \code{4/28/1994} \code{4/28/1995} \code{4/28/1997} \code{4/28/1998} \code{4/29/1994} \code{4/29/1996} \code{4/29/1997} \code{4/29/1998} \code{4/3/1995} \code{4/3/1996} \code{4/3/1997} \code{4/3/1998} \code{4/30/1996} \code{4/30/1997} \code{4/30/1998} \code{4/4/1994} \code{4/4/1995} \code{4/4/1996} \code{4/4/1997} \code{4/5/1994} \code{4/5/1995} \code{4/6/1994} \code{4/6/1995} \code{4/6/1998} \code{4/7/1994} \code{4/7/1995} \code{4/7/1997} \code{4/7/1998} \code{4/8/1994} \code{4/8/1996} \code{4/8/1997} \code{4/8/1998} \code{4/9/1996} \code{4/9/1997} \code{4/9/1998} \code{5/1/1995} \code{5/1/1996} \code{5/1/1997} \code{5/1/1998} \code{5/10/1994} \code{5/10/1995} \code{5/10/1996} \code{5/11/1994} \code{5/11/1995} \code{5/11/1998} \code{5/12/1994} \code{5/12/1995} \code{5/12/1997} \code{5/12/1998} \code{5/13/1994} \code{5/13/1996} \code{5/13/1997} \code{5/13/1998} \code{5/14/1996} \code{5/14/1997} \code{5/14/1998} \code{5/15/1995} \code{5/15/1996} \code{5/15/1997} \code{5/15/1998} \code{5/16/1994} \code{5/16/1995} \code{5/16/1996} \code{5/16/1997} \code{5/17/1994} \code{5/17/1995} \code{5/17/1996} \code{5/18/1994} \code{5/18/1995} \code{5/18/1998} \code{5/19/1994} \code{5/19/1995} \code{5/19/1997} \code{5/19/1998} \code{5/2/1994} \code{5/2/1995} \code{5/2/1996} \code{5/2/1997} \code{5/20/1994} \code{5/20/1996} \code{5/20/1997} \code{5/20/1998} \code{5/21/1996} \code{5/21/1997} \code{5/21/1998} \code{5/22/1995} \code{5/22/1996} \code{5/22/1997} \code{5/22/1998} \code{5/23/1994} \code{5/23/1995} \code{5/23/1996} \code{5/23/1997} \code{5/24/1994} \code{5/24/1995} \code{5/24/1996} \code{5/25/1994} \code{5/25/1995} \code{5/26/1994} \code{5/26/1995} \code{5/26/1998} \code{5/27/1994} \code{5/27/1997} \code{5/27/1998} \code{5/28/1996} \code{5/28/1997} \code{5/28/1998} \code{5/29/1996} \code{5/29/1997} \code{5/29/1998} \code{5/3/1994} \code{5/3/1995} \code{5/3/1996} \code{5/30/1995} \code{5/30/1996} \code{5/30/1997} \code{5/31/1994} \code{5/31/1995} \code{5/31/1996} \code{5/4/1994} \code{5/4/1995} \code{5/4/1998} \code{5/5/1994} \code{5/5/1995} \code{5/5/1997} \code{5/5/1998} \code{5/6/1994} \code{5/6/1996} \code{5/6/1997} \code{5/6/1998} \code{5/7/1996} \code{5/7/1997} \code{5/7/1998} \code{5/8/1995} \code{5/8/1996} \code{5/8/1997} \code{5/8/1998} \code{5/9/1994} \code{5/9/1995} \code{5/9/1996} \code{5/9/1997} \code{6/1/1994} \code{6/1/1995} \code{6/1/1998} \code{6/10/1994} \code{6/10/1996} \code{6/10/1997} \code{6/10/1998} \code{6/11/1996} \code{6/11/1997} \code{6/11/1998} \code{6/12/1995} \code{6/12/1996} \code{6/12/1997} \code{6/12/1998} \code{6/13/1994} \code{6/13/1995} \code{6/13/1996} \code{6/13/1997} \code{6/14/1994} \code{6/14/1995} \code{6/14/1996} \code{6/15/1994} \code{6/15/1995} \code{6/15/1998} \code{6/16/1994} \code{6/16/1995} \code{6/16/1997} \code{6/16/1998} \code{6/17/1994} \code{6/17/1996} \code{6/17/1997} \code{6/17/1998} \code{6/18/1996} \code{6/18/1997} \code{6/18/1998} \code{6/19/1995} \code{6/19/1996} \code{6/19/1997} \code{6/19/1998} \code{6/2/1994} \code{6/2/1995} \code{6/2/1997} \code{6/2/1998} \code{6/20/1994} \code{6/20/1995} \code{6/20/1996} \code{6/20/1997} \code{6/21/1994} \code{6/21/1995} \code{6/21/1996} \code{6/22/1994} \code{6/22/1995} \code{6/22/1998} \code{6/23/1994} \code{6/23/1995} \code{6/23/1997} \code{6/23/1998} \code{6/24/1994} \code{6/24/1996} \code{6/24/1997} \code{6/24/1998} \code{6/25/1996} \code{6/25/1997} \code{6/25/1998} \code{6/26/1995} \code{6/26/1996} \code{6/26/1997} \code{6/26/1998} \code{6/27/1994} \code{6/27/1995} \code{6/27/1996} \code{6/27/1997} \code{6/28/1994} \code{6/28/1995} \code{6/28/1996} \code{6/29/1994} \code{6/29/1995} \code{6/29/1998} \code{6/3/1994} \code{6/3/1996} \code{6/3/1997} \code{6/3/1998} \code{6/30/1994} \code{6/30/1995} \code{6/30/1997} \code{6/30/1998} \code{6/4/1996} \code{6/4/1997} \code{6/4/1998} \code{6/5/1995} \code{6/5/1996} \code{6/5/1997} \code{6/5/1998} \code{6/6/1994} \code{6/6/1995} \code{6/6/1996} \code{6/6/1997} \code{6/7/1994} \code{6/7/1995} \code{6/7/1996} \code{6/8/1994} \code{6/8/1995} \code{6/8/1998} \code{6/9/1994} \code{6/9/1995} \code{6/9/1997} \code{6/9/1998} \code{7/1/1994} \code{7/1/1996} \code{7/1/1997} \code{7/1/1998} \code{7/10/1995} \code{7/10/1996} \code{7/10/1997} \code{7/10/1998} \code{7/11/1994} \code{7/11/1995} \code{7/11/1996} \code{7/11/1997} \code{7/12/1994} \code{7/12/1995} \code{7/12/1996} \code{7/13/1994} \code{7/13/1995} \code{7/13/1998} \code{7/14/1994} \code{7/14/1995} \code{7/14/1997} \code{7/14/1998} \code{7/15/1994} \code{7/15/1996} \code{7/15/1997} \code{7/15/1998} \code{7/16/1996} \code{7/16/1997} \code{7/16/1998} \code{7/17/1995} \code{7/17/1996} \code{7/17/1997} \code{7/17/1998} \code{7/18/1994} \code{7/18/1995} \code{7/18/1996} \code{7/18/1997} \code{7/19/1994} \code{7/19/1995} \code{7/19/1996} \code{7/2/1996} \code{7/2/1997} \code{7/2/1998} \code{7/20/1994} \code{7/20/1995} \code{7/20/1998} \code{7/21/1994} \code{7/21/1995} \code{7/21/1997} \code{7/21/1998} \code{7/22/1994} \code{7/22/1996} \code{7/22/1997} \code{7/22/1998} \code{7/23/1996} \code{7/23/1997} \code{7/23/1998} \code{7/24/1995} \code{7/24/1996} \code{7/24/1997} \code{7/24/1998} \code{7/25/1994} \code{7/25/1995} \code{7/25/1996} \code{7/25/1997} \code{7/26/1994} \code{7/26/1995} \code{7/26/1996} \code{7/27/1994} \code{7/27/1995} \code{7/27/1998} \code{7/28/1994} \code{7/28/1995} \code{7/28/1997} \code{7/28/1998} \code{7/29/1994} \code{7/29/1996} \code{7/29/1997} \code{7/29/1998} \code{7/3/1995} \code{7/3/1996} \code{7/3/1997} \code{7/30/1996} \code{7/30/1997} \code{7/30/1998} \code{7/31/1995} \code{7/31/1996} \code{7/31/1997} \code{7/31/1998} \code{7/5/1994} \code{7/5/1995} \code{7/5/1996} \code{7/6/1994} \code{7/6/1995} \code{7/6/1998} \code{7/7/1994} \code{7/7/1995} \code{7/7/1997} \code{7/7/1998} \code{7/8/1994} \code{7/8/1996} \code{7/8/1997} \code{7/8/1998} \code{7/9/1996} \code{7/9/1997} \code{7/9/1998} \code{8/1/1994} \code{8/1/1995} \code{8/1/1996} \code{8/1/1997} \code{8/10/1994} \code{8/10/1995} \code{8/10/1998} \code{8/11/1994} \code{8/11/1995} \code{8/11/1997} \code{8/11/1998} \code{8/12/1994} \code{8/12/1996} \code{8/12/1997} \code{8/12/1998} \code{8/13/1996} \code{8/13/1997} \code{8/13/1998} \code{8/14/1995} \code{8/14/1996} \code{8/14/1997} \code{8/14/1998} \code{8/15/1994} \code{8/15/1995} \code{8/15/1996} \code{8/15/1997} \code{8/16/1994} \code{8/16/1995} \code{8/16/1996} \code{8/17/1994} \code{8/17/1995} \code{8/17/1998} \code{8/18/1994} \code{8/18/1995} \code{8/18/1997} \code{8/18/1998} \code{8/19/1994} \code{8/19/1996} \code{8/19/1997} \code{8/19/1998} \code{8/2/1994} \code{8/2/1995} \code{8/2/1996} \code{8/20/1996} \code{8/20/1997} \code{8/21/1995} \code{8/21/1996} \code{8/21/1997} \code{8/22/1994} \code{8/22/1995} \code{8/22/1996} \code{8/22/1997} \code{8/23/1994} \code{8/23/1995} \code{8/23/1996} \code{8/24/1993} \code{8/24/1994} \code{8/24/1995} \code{8/25/1993} \code{8/25/1994} \code{8/25/1995} \code{8/25/1997} \code{8/26/1993} \code{8/26/1994} \code{8/26/1996} \code{8/26/1997} \code{8/27/1993} \code{8/27/1996} \code{8/27/1997} \code{8/28/1995} \code{8/28/1996} \code{8/28/1997} \code{8/29/1994} \code{8/29/1995} \code{8/29/1996} \code{8/29/1997} \code{8/3/1994} \code{8/3/1995} \code{8/3/1998} \code{8/30/1993} \code{8/30/1994} \code{8/30/1995} \code{8/30/1996} \code{8/31/1993} \code{8/31/1994} \code{8/31/1995} \code{8/4/1994} \code{8/4/1995} \code{8/4/1997} \code{8/4/1998} \code{8/5/1994} \code{8/5/1996} \code{8/5/1997} \code{8/5/1998} \code{8/6/1996} \code{8/6/1997} \code{8/6/1998} \code{8/7/1995} \code{8/7/1996} \code{8/7/1997} \code{8/7/1998} \code{8/8/1994} \code{8/8/1995} \code{8/8/1996} \code{8/8/1997} \code{8/9/1994} \code{8/9/1995} \code{8/9/1996} \code{9/1/1993} \code{9/1/1994} \code{9/1/1995} \code{9/10/1993} \code{9/10/1996} \code{9/10/1997} \code{9/11/1995} \code{9/11/1996} \code{9/11/1997} \code{9/12/1994} \code{9/12/1995} \code{9/12/1996} \code{9/12/1997} \code{9/13/1993} \code{9/13/1994} \code{9/13/1995} \code{9/13/1996} \code{9/14/1993} \code{9/14/1994} \code{9/14/1995} \code{9/15/1993} \code{9/15/1994} \code{9/15/1995} \code{9/15/1997} \code{9/16/1993} \code{9/16/1994} \code{9/16/1996} \code{9/16/1997} \code{9/17/1993} \code{9/17/1996} \code{9/17/1997} \code{9/18/1995} \code{9/18/1996} \code{9/18/1997} \code{9/19/1994} \code{9/19/1995} \code{9/19/1996} \code{9/19/1997} \code{9/2/1993} \code{9/2/1994} \code{9/2/1997} \code{9/20/1993} \code{9/20/1994} \code{9/20/1995} \code{9/20/1996} \code{9/21/1993} \code{9/21/1994} \code{9/21/1995} \code{9/22/1993} \code{9/22/1994} \code{9/22/1995} \code{9/22/1997} \code{9/23/1993} \code{9/23/1994} \code{9/23/1996} \code{9/23/1997} \code{9/24/1993} \code{9/24/1996} \code{9/24/1997} \code{9/25/1995} \code{9/25/1996} \code{9/25/1997} \code{9/26/1994} \code{9/26/1995} \code{9/26/1996} \code{9/26/1997} \code{9/27/1993} \code{9/27/1994} \code{9/27/1995} \code{9/27/1996} \code{9/28/1993} \code{9/28/1994} \code{9/28/1995} \code{9/29/1993} \code{9/29/1994} \code{9/29/1995} \code{9/29/1997} \code{9/3/1993} \code{9/3/1996} \code{9/3/1997} \code{9/30/1993} \code{9/30/1994} \code{9/30/1996} \code{9/30/1997} \code{9/4/1996} \code{9/4/1997} \code{9/5/1995} \code{9/5/1996} \code{9/5/1997} \code{9/6/1994} \code{9/6/1995} \code{9/6/1996} \code{9/7/1993} \code{9/7/1994} \code{9/7/1995} \code{9/8/1993} \code{9/8/1994} \code{9/8/1995} \code{9/8/1997} \code{9/9/1993} \code{9/9/1994} \code{9/9/1996} \code{9/9/1997}}
    \item{\code{X}}{a numeric vector}
  }
}
\source{
S+ FinMetrics dell.s.
}
\references{
S+ FinMetrics dell.s.
}
\examples{
data(dell.s)
## maybe str(dell.s) ; plot(dell.s) ...
}
\keyword{datasets}
