\name{dell.v}
\alias{dell.v}
\docType{data}
\title{
daily trading volume of Dell stocks
}
\description{
daily trading volume of Dell stocks from 1/10/1995 to 9/9/1997
}
\usage{data(dell.v)}
\format{
  A data frame with 1261 observations on the following 2 variables.
}
\source{
S+ FinMetrics dell.s.
}
\references{
S+ FinMetrics dell.s.
}
\examples{
data(dell.v)
## maybe str(dell.v) ; plot(dell.v) ...
}
\keyword{datasets}
