% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dellReturns}
\alias{dellReturns}
\title{Daily stock returns of Dell Computer Corporation from October 1993 to October 1998}
\format{A \samp{xts} object on 1993-08-24/1998-08-19 containing:
\describe{
  \item{returns}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
dellReturns
}
\description{
A \samp{xts} object of daily stock returns of Dell Computer Corporation from October 1993 to
October 1998.
}
\keyword{datasets}

