% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DFX}
\alias{DFX}
\title{Weekly returns on the Dutch Guilder spot exchange rate from January 1980 to December 1998}
\format{A \samp{xts} object on 1980-01-08/1998-12-29 containing:
\describe{
  \item{DFX}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
DFX
}
\description{
A \samp{xts} object of weekly returns on the Dutch Guilder spot exchange rate from January 1980
to December 1998.
}
\keyword{datasets}

