\name{DFX.ts}
\alias{DFX.ts}
\docType{data}
\title{
weekly returns on the Dutch Guilder spot exchange rate from January 1980 to December 1998
}
\description{
weekly returns on the Dutch Guilder spot exchange rate from January 1980 to December 1998
}
\usage{data(DFX.ts)}
\format{
  A data frame with 1112 observations on the following 2 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{01/01/1981 00:00:00.000} \code{01/01/1982 00:00:00.000} \code{01/01/1983 00:00:00.000} \code{01/01/1984 00:00:00.000} \code{01/01/1985 00:00:00.000} \code{01/01/1986 00:00:00.000} \code{01/01/1987 00:00:00.000} \code{01/01/1988 00:00:00.000} \code{01/01/1989 00:00:00.000} \code{01/01/1990 00:00:00.000} \code{01/01/1991 00:00:00.000} \code{01/01/1992 00:00:00.000} \code{01/01/1993 00:00:00.000} \code{01/01/1994 00:00:00.000} \code{01/01/1995 00:00:00.000} \code{01/01/1996 00:00:00.000} \code{01/01/1997 00:00:00.000} \code{01/01/1998 00:00:00.000} \code{01/08/1980 00:00:00.000} \code{01/08/1981 00:00:00.000} \code{01/08/1982 00:00:00.000} \code{01/08/1983 00:00:00.000} \code{01/08/1984 00:00:00.000} \code{01/08/1985 00:00:00.000} \code{01/08/1986 00:00:00.000} \code{01/08/1987 00:00:00.000} \code{01/08/1988 00:00:00.000} \code{01/08/1989 00:00:00.000} \code{01/08/1990 00:00:00.000} \code{01/08/1991 00:00:00.000} \code{01/08/1992 00:00:00.000} \code{01/08/1993 00:00:00.000} \code{01/08/1994 00:00:00.000} \code{01/08/1995 00:00:00.000} \code{01/08/1996 00:00:00.000} \code{01/08/1997 00:00:00.000} \code{01/08/1998 00:00:00.000} \code{01/15/1980 00:00:00.000} \code{01/15/1981 00:00:00.000} \code{01/15/1982 00:00:00.000} \code{01/15/1983 00:00:00.000} \code{01/15/1984 00:00:00.000} \code{01/15/1985 00:00:00.000} \code{01/15/1986 00:00:00.000} \code{01/15/1987 00:00:00.000} \code{01/15/1988 00:00:00.000} \code{01/15/1989 00:00:00.000} \code{01/15/1990 00:00:00.000} \code{01/15/1991 00:00:00.000} \code{01/15/1992 00:00:00.000} \code{01/15/1993 00:00:00.000} \code{01/15/1994 00:00:00.000} \code{01/15/1995 00:00:00.000} \code{01/15/1996 00:00:00.000} \code{01/15/1997 00:00:00.000} \code{01/15/1998 00:00:00.000} \code{01/22/1980 00:00:00.000} \code{01/22/1981 00:00:00.000} \code{01/22/1982 00:00:00.000} \code{01/22/1983 00:00:00.000} \code{01/22/1984 00:00:00.000} \code{01/22/1985 00:00:00.000} \code{01/22/1986 00:00:00.000} \code{01/22/1987 00:00:00.000} \code{01/22/1988 00:00:00.000} \code{01/22/1989 00:00:00.000} \code{01/22/1990 00:00:00.000} \code{01/22/1991 00:00:00.000} \code{01/22/1992 00:00:00.000} \code{01/22/1993 00:00:00.000} \code{01/22/1994 00:00:00.000} \code{01/22/1995 00:00:00.000} \code{01/22/1996 00:00:00.000} \code{01/22/1997 00:00:00.000} \code{01/22/1998 00:00:00.000} \code{01/29/1980 00:00:00.000} \code{01/29/1981 00:00:00.000} \code{01/29/1982 00:00:00.000} \code{01/29/1983 00:00:00.000} \code{01/29/1984 00:00:00.000} \code{01/29/1985 00:00:00.000} \code{01/29/1986 00:00:00.000} \code{01/29/1987 00:00:00.000} \code{01/29/1988 00:00:00.000} \code{01/29/1989 00:00:00.000} \code{01/29/1990 00:00:00.000} \code{01/29/1991 00:00:00.000} \code{01/29/1992 00:00:00.000} \code{01/29/1993 00:00:00.000} \code{01/29/1994 00:00:00.000} \code{01/29/1995 00:00:00.000} \code{01/29/1996 00:00:00.000} \code{01/29/1997 00:00:00.000} \code{01/29/1998 00:00:00.000} \code{02/01/1980 00:00:00.000} \code{02/01/1981 00:00:00.000} \code{02/01/1982 00:00:00.000} \code{02/01/1983 00:00:00.000} \code{02/01/1984 00:00:00.000} \code{02/01/1985 00:00:00.000} \code{02/01/1986 00:00:00.000} \code{02/01/1987 00:00:00.000} \code{02/01/1988 00:00:00.000} \code{02/01/1989 00:00:00.000} \code{02/01/1990 00:00:00.000} \code{02/01/1991 00:00:00.000} \code{02/01/1992 00:00:00.000} \code{02/01/1993 00:00:00.000} \code{02/01/1994 00:00:00.000} \code{02/01/1995 00:00:00.000} \code{02/01/1996 00:00:00.000} \code{02/01/1997 00:00:00.000} \code{02/01/1998 00:00:00.000} \code{02/08/1980 00:00:00.000} \code{02/08/1981 00:00:00.000} \code{02/08/1982 00:00:00.000} \code{02/08/1983 00:00:00.000} \code{02/08/1984 00:00:00.000} \code{02/08/1985 00:00:00.000} \code{02/08/1986 00:00:00.000} \code{02/08/1987 00:00:00.000} \code{02/08/1988 00:00:00.000} \code{02/08/1989 00:00:00.000} \code{02/08/1990 00:00:00.000} \code{02/08/1991 00:00:00.000} \code{02/08/1992 00:00:00.000} \code{02/08/1993 00:00:00.000} \code{02/08/1994 00:00:00.000} \code{02/08/1995 00:00:00.000} \code{02/08/1996 00:00:00.000} \code{02/08/1997 00:00:00.000} \code{02/08/1998 00:00:00.000} \code{02/15/1980 00:00:00.000} \code{02/15/1981 00:00:00.000} \code{02/15/1982 00:00:00.000} \code{02/15/1983 00:00:00.000} \code{02/15/1984 00:00:00.000} \code{02/15/1985 00:00:00.000} \code{02/15/1986 00:00:00.000} \code{02/15/1987 00:00:00.000} \code{02/15/1988 00:00:00.000} \code{02/15/1989 00:00:00.000} \code{02/15/1990 00:00:00.000} \code{02/15/1991 00:00:00.000} \code{02/15/1992 00:00:00.000} \code{02/15/1993 00:00:00.000} \code{02/15/1994 00:00:00.000} \code{02/15/1995 00:00:00.000} \code{02/15/1996 00:00:00.000} \code{02/15/1997 00:00:00.000} \code{02/15/1998 00:00:00.000} \code{02/22/1980 00:00:00.000} \code{02/22/1981 00:00:00.000} \code{02/22/1982 00:00:00.000} \code{02/22/1983 00:00:00.000} \code{02/22/1984 00:00:00.000} \code{02/22/1985 00:00:00.000} \code{02/22/1986 00:00:00.000} \code{02/22/1987 00:00:00.000} \code{02/22/1988 00:00:00.000} \code{02/22/1989 00:00:00.000} \code{02/22/1990 00:00:00.000} \code{02/22/1991 00:00:00.000} \code{02/22/1992 00:00:00.000} \code{02/22/1993 00:00:00.000} \code{02/22/1994 00:00:00.000} \code{02/22/1995 00:00:00.000} \code{02/22/1996 00:00:00.000} \code{02/22/1997 00:00:00.000} \code{02/22/1998 00:00:00.000} \code{02/29/1980 00:00:00.000} \code{02/29/1984 00:00:00.000} \code{02/29/1988 00:00:00.000} \code{02/29/1996 00:00:00.000} \code{03/01/1980 00:00:00.000} \code{03/01/1981 00:00:00.000} \code{03/01/1982 00:00:00.000} \code{03/01/1983 00:00:00.000} \code{03/01/1984 00:00:00.000} \code{03/01/1985 00:00:00.000} \code{03/01/1986 00:00:00.000} \code{03/01/1987 00:00:00.000} \code{03/01/1988 00:00:00.000} \code{03/01/1989 00:00:00.000} \code{03/01/1990 00:00:00.000} \code{03/01/1991 00:00:00.000} \code{03/01/1992 00:00:00.000} \code{03/01/1993 00:00:00.000} \code{03/01/1994 00:00:00.000} \code{03/01/1995 00:00:00.000} \code{03/01/1996 00:00:00.000} \code{03/01/1997 00:00:00.000} \code{03/01/1998 00:00:00.000} \code{03/08/1980 00:00:00.000} \code{03/08/1981 00:00:00.000} \code{03/08/1982 00:00:00.000} \code{03/08/1983 00:00:00.000} \code{03/08/1984 00:00:00.000} \code{03/08/1985 00:00:00.000} \code{03/08/1986 00:00:00.000} \code{03/08/1987 00:00:00.000} \code{03/08/1988 00:00:00.000} \code{03/08/1989 00:00:00.000} \code{03/08/1990 00:00:00.000} \code{03/08/1991 00:00:00.000} \code{03/08/1992 00:00:00.000} \code{03/08/1993 00:00:00.000} \code{03/08/1994 00:00:00.000} \code{03/08/1995 00:00:00.000} \code{03/08/1996 00:00:00.000} \code{03/08/1997 00:00:00.000} \code{03/08/1998 00:00:00.000} \code{03/15/1980 00:00:00.000} \code{03/15/1981 00:00:00.000} \code{03/15/1982 00:00:00.000} \code{03/15/1983 00:00:00.000} \code{03/15/1984 00:00:00.000} \code{03/15/1985 00:00:00.000} \code{03/15/1986 00:00:00.000} \code{03/15/1987 00:00:00.000} \code{03/15/1988 00:00:00.000} \code{03/15/1989 00:00:00.000} \code{03/15/1990 00:00:00.000} \code{03/15/1991 00:00:00.000} \code{03/15/1992 00:00:00.000} \code{03/15/1993 00:00:00.000} \code{03/15/1994 00:00:00.000} \code{03/15/1995 00:00:00.000} \code{03/15/1996 00:00:00.000} \code{03/15/1997 00:00:00.000} \code{03/15/1998 00:00:00.000} \code{03/22/1980 00:00:00.000} \code{03/22/1981 00:00:00.000} \code{03/22/1982 00:00:00.000} \code{03/22/1983 00:00:00.000} \code{03/22/1984 00:00:00.000} \code{03/22/1985 00:00:00.000} \code{03/22/1986 00:00:00.000} \code{03/22/1987 00:00:00.000} \code{03/22/1988 00:00:00.000} \code{03/22/1989 00:00:00.000} \code{03/22/1990 00:00:00.000} \code{03/22/1991 00:00:00.000} \code{03/22/1992 00:00:00.000} \code{03/22/1993 00:00:00.000} \code{03/22/1994 00:00:00.000} \code{03/22/1995 00:00:00.000} \code{03/22/1996 00:00:00.000} \code{03/22/1997 00:00:00.000} \code{03/22/1998 00:00:00.000} \code{03/29/1980 00:00:00.000} \code{03/29/1981 00:00:00.000} \code{03/29/1982 00:00:00.000} \code{03/29/1983 00:00:00.000} \code{03/29/1984 00:00:00.000} \code{03/29/1985 00:00:00.000} \code{03/29/1986 00:00:00.000} \code{03/29/1987 00:00:00.000} \code{03/29/1988 00:00:00.000} \code{03/29/1989 00:00:00.000} \code{03/29/1990 00:00:00.000} \code{03/29/1991 00:00:00.000} \code{03/29/1992 00:00:00.000} \code{03/29/1993 00:00:00.000} \code{03/29/1994 00:00:00.000} \code{03/29/1995 00:00:00.000} \code{03/29/1996 00:00:00.000} \code{03/29/1997 00:00:00.000} \code{03/29/1998 00:00:00.000} \code{04/01/1980 00:00:00.000} \code{04/01/1981 00:00:00.000} \code{04/01/1982 00:00:00.000} \code{04/01/1983 00:00:00.000} \code{04/01/1984 00:00:00.000} \code{04/01/1985 00:00:00.000} \code{04/01/1986 00:00:00.000} \code{04/01/1987 00:00:00.000} \code{04/01/1988 00:00:00.000} \code{04/01/1989 00:00:00.000} \code{04/01/1990 00:00:00.000} \code{04/01/1991 00:00:00.000} \code{04/01/1992 00:00:00.000} \code{04/01/1993 00:00:00.000} \code{04/01/1994 00:00:00.000} \code{04/01/1995 00:00:00.000} \code{04/01/1996 00:00:00.000} \code{04/01/1997 00:00:00.000} \code{04/01/1998 00:00:00.000} \code{04/08/1980 00:00:00.000} \code{04/08/1981 00:00:00.000} \code{04/08/1982 00:00:00.000} \code{04/08/1983 00:00:00.000} \code{04/08/1984 00:00:00.000} \code{04/08/1985 00:00:00.000} \code{04/08/1986 00:00:00.000} \code{04/08/1987 00:00:00.000} \code{04/08/1988 00:00:00.000} \code{04/08/1989 00:00:00.000} \code{04/08/1990 00:00:00.000} \code{04/08/1991 00:00:00.000} \code{04/08/1992 00:00:00.000} \code{04/08/1993 00:00:00.000} \code{04/08/1994 00:00:00.000} \code{04/08/1995 00:00:00.000} \code{04/08/1996 00:00:00.000} \code{04/08/1997 00:00:00.000} \code{04/08/1998 00:00:00.000} \code{04/15/1980 00:00:00.000} \code{04/15/1981 00:00:00.000} \code{04/15/1982 00:00:00.000} \code{04/15/1983 00:00:00.000} \code{04/15/1984 00:00:00.000} \code{04/15/1985 00:00:00.000} \code{04/15/1986 00:00:00.000} \code{04/15/1987 00:00:00.000} \code{04/15/1988 00:00:00.000} \code{04/15/1989 00:00:00.000} \code{04/15/1990 00:00:00.000} \code{04/15/1991 00:00:00.000} \code{04/15/1992 00:00:00.000} \code{04/15/1993 00:00:00.000} \code{04/15/1994 00:00:00.000} \code{04/15/1995 00:00:00.000} \code{04/15/1996 00:00:00.000} \code{04/15/1997 00:00:00.000} \code{04/15/1998 00:00:00.000} \code{04/22/1980 00:00:00.000} \code{04/22/1981 00:00:00.000} \code{04/22/1982 00:00:00.000} \code{04/22/1983 00:00:00.000} \code{04/22/1984 00:00:00.000} \code{04/22/1985 00:00:00.000} \code{04/22/1986 00:00:00.000} \code{04/22/1987 00:00:00.000} \code{04/22/1988 00:00:00.000} \code{04/22/1989 00:00:00.000} \code{04/22/1990 00:00:00.000} \code{04/22/1991 00:00:00.000} \code{04/22/1992 00:00:00.000} \code{04/22/1993 00:00:00.000} \code{04/22/1994 00:00:00.000} \code{04/22/1995 00:00:00.000} \code{04/22/1996 00:00:00.000} \code{04/22/1997 00:00:00.000} \code{04/22/1998 00:00:00.000} \code{04/29/1980 00:00:00.000} \code{04/29/1981 00:00:00.000} \code{04/29/1982 00:00:00.000} \code{04/29/1983 00:00:00.000} \code{04/29/1984 00:00:00.000} \code{04/29/1985 00:00:00.000} \code{04/29/1986 00:00:00.000} \code{04/29/1987 00:00:00.000} \code{04/29/1988 00:00:00.000} \code{04/29/1990 00:00:00.000} \code{04/29/1991 00:00:00.000} \code{04/29/1992 00:00:00.000} \code{04/29/1993 00:00:00.000} \code{04/29/1994 00:00:00.000} \code{04/29/1996 00:00:00.000} \code{04/29/1997 00:00:00.000} \code{04/29/1998 00:00:00.000} \code{05/01/1980 00:00:00.000} \code{05/01/1981 00:00:00.000} \code{05/01/1982 00:00:00.000} \code{05/01/1983 00:00:00.000} \code{05/01/1984 00:00:00.000} \code{05/01/1985 00:00:00.000} \code{05/01/1986 00:00:00.000} \code{05/01/1987 00:00:00.000} \code{05/01/1988 00:00:00.000} \code{05/01/1989 00:00:00.000} \code{05/01/1990 00:00:00.000} \code{05/01/1991 00:00:00.000} \code{05/01/1992 00:00:00.000} \code{05/01/1993 00:00:00.000} \code{05/01/1994 00:00:00.000} \code{05/01/1995 00:00:00.000} \code{05/01/1996 00:00:00.000} \code{05/01/1997 00:00:00.000} \code{05/01/1998 00:00:00.000} \code{05/08/1980 00:00:00.000} \code{05/08/1981 00:00:00.000} \code{05/08/1982 00:00:00.000} \code{05/08/1983 00:00:00.000} \code{05/08/1984 00:00:00.000} \code{05/08/1985 00:00:00.000} \code{05/08/1986 00:00:00.000} \code{05/08/1987 00:00:00.000} \code{05/08/1988 00:00:00.000} \code{05/08/1989 00:00:00.000} \code{05/08/1990 00:00:00.000} \code{05/08/1991 00:00:00.000} \code{05/08/1992 00:00:00.000} \code{05/08/1993 00:00:00.000} \code{05/08/1994 00:00:00.000} \code{05/08/1995 00:00:00.000} \code{05/08/1996 00:00:00.000} \code{05/08/1997 00:00:00.000} \code{05/08/1998 00:00:00.000} \code{05/15/1980 00:00:00.000} \code{05/15/1981 00:00:00.000} \code{05/15/1982 00:00:00.000} \code{05/15/1983 00:00:00.000} \code{05/15/1984 00:00:00.000} \code{05/15/1985 00:00:00.000} \code{05/15/1986 00:00:00.000} \code{05/15/1987 00:00:00.000} \code{05/15/1988 00:00:00.000} \code{05/15/1989 00:00:00.000} \code{05/15/1990 00:00:00.000} \code{05/15/1991 00:00:00.000} \code{05/15/1992 00:00:00.000} \code{05/15/1993 00:00:00.000} \code{05/15/1994 00:00:00.000} \code{05/15/1995 00:00:00.000} \code{05/15/1996 00:00:00.000} \code{05/15/1997 00:00:00.000} \code{05/15/1998 00:00:00.000} \code{05/22/1980 00:00:00.000} \code{05/22/1981 00:00:00.000} \code{05/22/1982 00:00:00.000} \code{05/22/1983 00:00:00.000} \code{05/22/1984 00:00:00.000} \code{05/22/1985 00:00:00.000} \code{05/22/1986 00:00:00.000} \code{05/22/1987 00:00:00.000} \code{05/22/1988 00:00:00.000} \code{05/22/1989 00:00:00.000} \code{05/22/1990 00:00:00.000} \code{05/22/1991 00:00:00.000} \code{05/22/1992 00:00:00.000} \code{05/22/1993 00:00:00.000} \code{05/22/1994 00:00:00.000} \code{05/22/1995 00:00:00.000} \code{05/22/1996 00:00:00.000} \code{05/22/1997 00:00:00.000} \code{05/22/1998 00:00:00.000} \code{05/29/1980 00:00:00.000} \code{05/29/1981 00:00:00.000} \code{05/29/1983 00:00:00.000} \code{05/29/1984 00:00:00.000} \code{05/29/1985 00:00:00.000} \code{05/29/1986 00:00:00.000} \code{05/29/1987 00:00:00.000} \code{05/29/1988 00:00:00.000} \code{05/29/1989 00:00:00.000} \code{05/29/1990 00:00:00.000} \code{05/29/1991 00:00:00.000} \code{05/29/1992 00:00:00.000} \code{05/29/1994 00:00:00.000} \code{05/29/1995 00:00:00.000} \code{05/29/1996 00:00:00.000} \code{05/29/1997 00:00:00.000} \code{05/29/1998 00:00:00.000} \code{06/01/1980 00:00:00.000} \code{06/01/1981 00:00:00.000} \code{06/01/1982 00:00:00.000} \code{06/01/1983 00:00:00.000} \code{06/01/1984 00:00:00.000} \code{06/01/1985 00:00:00.000} \code{06/01/1986 00:00:00.000} \code{06/01/1987 00:00:00.000} \code{06/01/1988 00:00:00.000} \code{06/01/1989 00:00:00.000} \code{06/01/1990 00:00:00.000} \code{06/01/1991 00:00:00.000} \code{06/01/1992 00:00:00.000} \code{06/01/1993 00:00:00.000} \code{06/01/1994 00:00:00.000} \code{06/01/1995 00:00:00.000} \code{06/01/1996 00:00:00.000} \code{06/01/1997 00:00:00.000} \code{06/01/1998 00:00:00.000} \code{06/08/1980 00:00:00.000} \code{06/08/1981 00:00:00.000} \code{06/08/1982 00:00:00.000} \code{06/08/1983 00:00:00.000} \code{06/08/1984 00:00:00.000} \code{06/08/1985 00:00:00.000} \code{06/08/1986 00:00:00.000} \code{06/08/1987 00:00:00.000} \code{06/08/1988 00:00:00.000} \code{06/08/1989 00:00:00.000} \code{06/08/1990 00:00:00.000} \code{06/08/1991 00:00:00.000} \code{06/08/1992 00:00:00.000} \code{06/08/1993 00:00:00.000} \code{06/08/1994 00:00:00.000} \code{06/08/1995 00:00:00.000} \code{06/08/1996 00:00:00.000} \code{06/08/1997 00:00:00.000} \code{06/08/1998 00:00:00.000} \code{06/15/1980 00:00:00.000} \code{06/15/1981 00:00:00.000} \code{06/15/1982 00:00:00.000} \code{06/15/1983 00:00:00.000} \code{06/15/1984 00:00:00.000} \code{06/15/1985 00:00:00.000} \code{06/15/1986 00:00:00.000} \code{06/15/1987 00:00:00.000} \code{06/15/1988 00:00:00.000} \code{06/15/1989 00:00:00.000} \code{06/15/1990 00:00:00.000} \code{06/15/1991 00:00:00.000} \code{06/15/1992 00:00:00.000} \code{06/15/1993 00:00:00.000} \code{06/15/1994 00:00:00.000} \code{06/15/1995 00:00:00.000} \code{06/15/1996 00:00:00.000} \code{06/15/1997 00:00:00.000} \code{06/15/1998 00:00:00.000} \code{06/22/1980 00:00:00.000} \code{06/22/1981 00:00:00.000} \code{06/22/1982 00:00:00.000} \code{06/22/1983 00:00:00.000} \code{06/22/1984 00:00:00.000} \code{06/22/1985 00:00:00.000} \code{06/22/1986 00:00:00.000} \code{06/22/1987 00:00:00.000} \code{06/22/1988 00:00:00.000} \code{06/22/1989 00:00:00.000} \code{06/22/1990 00:00:00.000} \code{06/22/1991 00:00:00.000} \code{06/22/1992 00:00:00.000} \code{06/22/1993 00:00:00.000} \code{06/22/1994 00:00:00.000} \code{06/22/1995 00:00:00.000} \code{06/22/1996 00:00:00.000} \code{06/22/1997 00:00:00.000} \code{06/22/1998 00:00:00.000} \code{06/29/1980 00:00:00.000} \code{06/29/1981 00:00:00.000} \code{06/29/1982 00:00:00.000} \code{06/29/1983 00:00:00.000} \code{06/29/1984 00:00:00.000} \code{06/29/1986 00:00:00.000} \code{06/29/1987 00:00:00.000} \code{06/29/1988 00:00:00.000} \code{06/29/1989 00:00:00.000} \code{06/29/1990 00:00:00.000} \code{06/29/1992 00:00:00.000} \code{06/29/1993 00:00:00.000} \code{06/29/1994 00:00:00.000} \code{06/29/1995 00:00:00.000} \code{06/29/1997 00:00:00.000} \code{06/29/1998 00:00:00.000} \code{07/01/1980 00:00:00.000} \code{07/01/1981 00:00:00.000} \code{07/01/1982 00:00:00.000} \code{07/01/1983 00:00:00.000} \code{07/01/1984 00:00:00.000} \code{07/01/1985 00:00:00.000} \code{07/01/1986 00:00:00.000} \code{07/01/1987 00:00:00.000} \code{07/01/1988 00:00:00.000} \code{07/01/1989 00:00:00.000} \code{07/01/1990 00:00:00.000} \code{07/01/1991 00:00:00.000} \code{07/01/1992 00:00:00.000} \code{07/01/1993 00:00:00.000} \code{07/01/1994 00:00:00.000} \code{07/01/1995 00:00:00.000} \code{07/01/1996 00:00:00.000} \code{07/01/1997 00:00:00.000} \code{07/01/1998 00:00:00.000} \code{07/08/1980 00:00:00.000} \code{07/08/1981 00:00:00.000} \code{07/08/1982 00:00:00.000} \code{07/08/1983 00:00:00.000} \code{07/08/1984 00:00:00.000} \code{07/08/1985 00:00:00.000} \code{07/08/1986 00:00:00.000} \code{07/08/1987 00:00:00.000} \code{07/08/1988 00:00:00.000} \code{07/08/1989 00:00:00.000} \code{07/08/1990 00:00:00.000} \code{07/08/1991 00:00:00.000} \code{07/08/1992 00:00:00.000} \code{07/08/1993 00:00:00.000} \code{07/08/1994 00:00:00.000} \code{07/08/1995 00:00:00.000} \code{07/08/1996 00:00:00.000} \code{07/08/1997 00:00:00.000} \code{07/08/1998 00:00:00.000} \code{07/15/1980 00:00:00.000} \code{07/15/1981 00:00:00.000} \code{07/15/1982 00:00:00.000} \code{07/15/1983 00:00:00.000} \code{07/15/1984 00:00:00.000} \code{07/15/1985 00:00:00.000} \code{07/15/1986 00:00:00.000} \code{07/15/1987 00:00:00.000} \code{07/15/1988 00:00:00.000} \code{07/15/1989 00:00:00.000} \code{07/15/1990 00:00:00.000} \code{07/15/1991 00:00:00.000} \code{07/15/1992 00:00:00.000} \code{07/15/1993 00:00:00.000} \code{07/15/1994 00:00:00.000} \code{07/15/1995 00:00:00.000} \code{07/15/1996 00:00:00.000} \code{07/15/1997 00:00:00.000} \code{07/15/1998 00:00:00.000} \code{07/22/1980 00:00:00.000} \code{07/22/1981 00:00:00.000} \code{07/22/1982 00:00:00.000} \code{07/22/1983 00:00:00.000} \code{07/22/1984 00:00:00.000} \code{07/22/1985 00:00:00.000} \code{07/22/1986 00:00:00.000} \code{07/22/1987 00:00:00.000} \code{07/22/1988 00:00:00.000} \code{07/22/1989 00:00:00.000} \code{07/22/1990 00:00:00.000} \code{07/22/1991 00:00:00.000} \code{07/22/1992 00:00:00.000} \code{07/22/1993 00:00:00.000} \code{07/22/1994 00:00:00.000} \code{07/22/1995 00:00:00.000} \code{07/22/1996 00:00:00.000} \code{07/22/1997 00:00:00.000} \code{07/22/1998 00:00:00.000} \code{07/29/1980 00:00:00.000} \code{07/29/1981 00:00:00.000} \code{07/29/1982 00:00:00.000} \code{07/29/1983 00:00:00.000} \code{07/29/1984 00:00:00.000} \code{07/29/1985 00:00:00.000} \code{07/29/1986 00:00:00.000} \code{07/29/1987 00:00:00.000} \code{07/29/1988 00:00:00.000} \code{07/29/1989 00:00:00.000} \code{07/29/1990 00:00:00.000} \code{07/29/1991 00:00:00.000} \code{07/29/1992 00:00:00.000} \code{07/29/1993 00:00:00.000} \code{07/29/1994 00:00:00.000} \code{07/29/1995 00:00:00.000} \code{07/29/1996 00:00:00.000} \code{07/29/1997 00:00:00.000} \code{07/29/1998 00:00:00.000} \code{08/01/1980 00:00:00.000} \code{08/01/1981 00:00:00.000} \code{08/01/1982 00:00:00.000} \code{08/01/1983 00:00:00.000} \code{08/01/1984 00:00:00.000} \code{08/01/1985 00:00:00.000} \code{08/01/1986 00:00:00.000} \code{08/01/1987 00:00:00.000} \code{08/01/1988 00:00:00.000} \code{08/01/1989 00:00:00.000} \code{08/01/1990 00:00:00.000} \code{08/01/1991 00:00:00.000} \code{08/01/1992 00:00:00.000} \code{08/01/1993 00:00:00.000} \code{08/01/1994 00:00:00.000} \code{08/01/1995 00:00:00.000} \code{08/01/1996 00:00:00.000} \code{08/01/1997 00:00:00.000} \code{08/01/1998 00:00:00.000} \code{08/08/1980 00:00:00.000} \code{08/08/1981 00:00:00.000} \code{08/08/1982 00:00:00.000} \code{08/08/1983 00:00:00.000} \code{08/08/1984 00:00:00.000} \code{08/08/1985 00:00:00.000} \code{08/08/1986 00:00:00.000} \code{08/08/1987 00:00:00.000} \code{08/08/1988 00:00:00.000} \code{08/08/1989 00:00:00.000} \code{08/08/1990 00:00:00.000} \code{08/08/1991 00:00:00.000} \code{08/08/1992 00:00:00.000} \code{08/08/1993 00:00:00.000} \code{08/08/1994 00:00:00.000} \code{08/08/1995 00:00:00.000} \code{08/08/1996 00:00:00.000} \code{08/08/1997 00:00:00.000} \code{08/08/1998 00:00:00.000} \code{08/15/1980 00:00:00.000} \code{08/15/1981 00:00:00.000} \code{08/15/1982 00:00:00.000} \code{08/15/1983 00:00:00.000} \code{08/15/1984 00:00:00.000} \code{08/15/1985 00:00:00.000} \code{08/15/1986 00:00:00.000} \code{08/15/1987 00:00:00.000} \code{08/15/1988 00:00:00.000} \code{08/15/1989 00:00:00.000} \code{08/15/1990 00:00:00.000} \code{08/15/1991 00:00:00.000} \code{08/15/1992 00:00:00.000} \code{08/15/1993 00:00:00.000} \code{08/15/1994 00:00:00.000} \code{08/15/1995 00:00:00.000} \code{08/15/1996 00:00:00.000} \code{08/15/1997 00:00:00.000} \code{08/15/1998 00:00:00.000} \code{08/22/1980 00:00:00.000} \code{08/22/1981 00:00:00.000} \code{08/22/1982 00:00:00.000} \code{08/22/1983 00:00:00.000} \code{08/22/1984 00:00:00.000} \code{08/22/1985 00:00:00.000} \code{08/22/1986 00:00:00.000} \code{08/22/1987 00:00:00.000} \code{08/22/1988 00:00:00.000} \code{08/22/1989 00:00:00.000} \code{08/22/1990 00:00:00.000} \code{08/22/1991 00:00:00.000} \code{08/22/1992 00:00:00.000} \code{08/22/1993 00:00:00.000} \code{08/22/1994 00:00:00.000} \code{08/22/1995 00:00:00.000} \code{08/22/1996 00:00:00.000} \code{08/22/1997 00:00:00.000} \code{08/22/1998 00:00:00.000} \code{08/29/1980 00:00:00.000} \code{08/29/1981 00:00:00.000} \code{08/29/1982 00:00:00.000} \code{08/29/1983 00:00:00.000} \code{08/29/1984 00:00:00.000} \code{08/29/1985 00:00:00.000} \code{08/29/1986 00:00:00.000} \code{08/29/1987 00:00:00.000} \code{08/29/1988 00:00:00.000} \code{08/29/1989 00:00:00.000} \code{08/29/1990 00:00:00.000} \code{08/29/1991 00:00:00.000} \code{08/29/1992 00:00:00.000} \code{08/29/1993 00:00:00.000} \code{08/29/1994 00:00:00.000} \code{08/29/1995 00:00:00.000} \code{08/29/1996 00:00:00.000} \code{08/29/1997 00:00:00.000} \code{08/29/1998 00:00:00.000} \code{09/01/1980 00:00:00.000} \code{09/01/1981 00:00:00.000} \code{09/01/1982 00:00:00.000} \code{09/01/1983 00:00:00.000} \code{09/01/1984 00:00:00.000} \code{09/01/1985 00:00:00.000} \code{09/01/1986 00:00:00.000} \code{09/01/1987 00:00:00.000} \code{09/01/1988 00:00:00.000} \code{09/01/1989 00:00:00.000} \code{09/01/1990 00:00:00.000} \code{09/01/1991 00:00:00.000} \code{09/01/1992 00:00:00.000} \code{09/01/1993 00:00:00.000} \code{09/01/1994 00:00:00.000} \code{09/01/1995 00:00:00.000} \code{09/01/1996 00:00:00.000} \code{09/01/1997 00:00:00.000} \code{09/01/1998 00:00:00.000} \code{09/08/1980 00:00:00.000} \code{09/08/1981 00:00:00.000} \code{09/08/1982 00:00:00.000} \code{09/08/1983 00:00:00.000} \code{09/08/1984 00:00:00.000} \code{09/08/1985 00:00:00.000} \code{09/08/1986 00:00:00.000} \code{09/08/1987 00:00:00.000} \code{09/08/1988 00:00:00.000} \code{09/08/1989 00:00:00.000} \code{09/08/1990 00:00:00.000} \code{09/08/1991 00:00:00.000} \code{09/08/1992 00:00:00.000} \code{09/08/1993 00:00:00.000} \code{09/08/1994 00:00:00.000} \code{09/08/1995 00:00:00.000} \code{09/08/1996 00:00:00.000} \code{09/08/1997 00:00:00.000} \code{09/08/1998 00:00:00.000} \code{09/15/1980 00:00:00.000} \code{09/15/1981 00:00:00.000} \code{09/15/1982 00:00:00.000} \code{09/15/1983 00:00:00.000} \code{09/15/1984 00:00:00.000} \code{09/15/1985 00:00:00.000} \code{09/15/1986 00:00:00.000} \code{09/15/1987 00:00:00.000} \code{09/15/1988 00:00:00.000} \code{09/15/1989 00:00:00.000} \code{09/15/1990 00:00:00.000} \code{09/15/1991 00:00:00.000} \code{09/15/1992 00:00:00.000} \code{09/15/1993 00:00:00.000} \code{09/15/1994 00:00:00.000} \code{09/15/1995 00:00:00.000} \code{09/15/1996 00:00:00.000} \code{09/15/1997 00:00:00.000} \code{09/15/1998 00:00:00.000} \code{09/22/1980 00:00:00.000} \code{09/22/1981 00:00:00.000} \code{09/22/1982 00:00:00.000} \code{09/22/1983 00:00:00.000} \code{09/22/1984 00:00:00.000} \code{09/22/1985 00:00:00.000} \code{09/22/1986 00:00:00.000} \code{09/22/1987 00:00:00.000} \code{09/22/1988 00:00:00.000} \code{09/22/1989 00:00:00.000} \code{09/22/1990 00:00:00.000} \code{09/22/1991 00:00:00.000} \code{09/22/1992 00:00:00.000} \code{09/22/1993 00:00:00.000} \code{09/22/1994 00:00:00.000} \code{09/22/1995 00:00:00.000} \code{09/22/1996 00:00:00.000} \code{09/22/1997 00:00:00.000} \code{09/22/1998 00:00:00.000} \code{09/29/1980 00:00:00.000} \code{09/29/1981 00:00:00.000} \code{09/29/1982 00:00:00.000} \code{09/29/1983 00:00:00.000} \code{09/29/1985 00:00:00.000} \code{09/29/1986 00:00:00.000} \code{09/29/1987 00:00:00.000} \code{09/29/1988 00:00:00.000} \code{09/29/1989 00:00:00.000} \code{09/29/1991 00:00:00.000} \code{09/29/1992 00:00:00.000} \code{09/29/1993 00:00:00.000} \code{09/29/1994 00:00:00.000} \code{09/29/1995 00:00:00.000} \code{09/29/1996 00:00:00.000} \code{09/29/1997 00:00:00.000} \code{09/29/1998 00:00:00.000} \code{10/01/1980 00:00:00.000} \code{10/01/1981 00:00:00.000} \code{10/01/1982 00:00:00.000} \code{10/01/1983 00:00:00.000} \code{10/01/1984 00:00:00.000} \code{10/01/1985 00:00:00.000} \code{10/01/1986 00:00:00.000} \code{10/01/1987 00:00:00.000} \code{10/01/1988 00:00:00.000} \code{10/01/1989 00:00:00.000} \code{10/01/1990 00:00:00.000} \code{10/01/1991 00:00:00.000} \code{10/01/1992 00:00:00.000} \code{10/01/1993 00:00:00.000} \code{10/01/1994 00:00:00.000} \code{10/01/1995 00:00:00.000} \code{10/01/1996 00:00:00.000} \code{10/01/1997 00:00:00.000} \code{10/01/1998 00:00:00.000} \code{10/08/1980 00:00:00.000} \code{10/08/1981 00:00:00.000} \code{10/08/1982 00:00:00.000} \code{10/08/1983 00:00:00.000} \code{10/08/1984 00:00:00.000} \code{10/08/1985 00:00:00.000} \code{10/08/1986 00:00:00.000} \code{10/08/1987 00:00:00.000} \code{10/08/1988 00:00:00.000} \code{10/08/1989 00:00:00.000} \code{10/08/1990 00:00:00.000} \code{10/08/1991 00:00:00.000} \code{10/08/1992 00:00:00.000} \code{10/08/1993 00:00:00.000} \code{10/08/1994 00:00:00.000} \code{10/08/1995 00:00:00.000} \code{10/08/1996 00:00:00.000} \code{10/08/1997 00:00:00.000} \code{10/08/1998 00:00:00.000} \code{10/15/1980 00:00:00.000} \code{10/15/1981 00:00:00.000} \code{10/15/1982 00:00:00.000} \code{10/15/1983 00:00:00.000} \code{10/15/1984 00:00:00.000} \code{10/15/1985 00:00:00.000} \code{10/15/1986 00:00:00.000} \code{10/15/1987 00:00:00.000} \code{10/15/1988 00:00:00.000} \code{10/15/1989 00:00:00.000} \code{10/15/1990 00:00:00.000} \code{10/15/1991 00:00:00.000} \code{10/15/1992 00:00:00.000} \code{10/15/1993 00:00:00.000} \code{10/15/1994 00:00:00.000} \code{10/15/1995 00:00:00.000} \code{10/15/1996 00:00:00.000} \code{10/15/1997 00:00:00.000} \code{10/15/1998 00:00:00.000} \code{10/22/1980 00:00:00.000} \code{10/22/1981 00:00:00.000} \code{10/22/1982 00:00:00.000} \code{10/22/1983 00:00:00.000} \code{10/22/1984 00:00:00.000} \code{10/22/1985 00:00:00.000} \code{10/22/1986 00:00:00.000} \code{10/22/1987 00:00:00.000} \code{10/22/1988 00:00:00.000} \code{10/22/1989 00:00:00.000} \code{10/22/1990 00:00:00.000} \code{10/22/1991 00:00:00.000} \code{10/22/1992 00:00:00.000} \code{10/22/1993 00:00:00.000} \code{10/22/1994 00:00:00.000} \code{10/22/1995 00:00:00.000} \code{10/22/1996 00:00:00.000} \code{10/22/1997 00:00:00.000} \code{10/22/1998 00:00:00.000} \code{10/29/1980 00:00:00.000} \code{10/29/1981 00:00:00.000} \code{10/29/1982 00:00:00.000} \code{10/29/1983 00:00:00.000} \code{10/29/1984 00:00:00.000} \code{10/29/1985 00:00:00.000} \code{10/29/1986 00:00:00.000} \code{10/29/1987 00:00:00.000} \code{10/29/1988 00:00:00.000} \code{10/29/1989 00:00:00.000} \code{10/29/1990 00:00:00.000} \code{10/29/1991 00:00:00.000} \code{10/29/1992 00:00:00.000} \code{10/29/1993 00:00:00.000} \code{10/29/1994 00:00:00.000} \code{10/29/1995 00:00:00.000} \code{10/29/1996 00:00:00.000} \code{10/29/1997 00:00:00.000} \code{10/29/1998 00:00:00.000} \code{11/01/1980 00:00:00.000} \code{11/01/1981 00:00:00.000} \code{11/01/1982 00:00:00.000} \code{11/01/1983 00:00:00.000} \code{11/01/1984 00:00:00.000} \code{11/01/1985 00:00:00.000} \code{11/01/1986 00:00:00.000} \code{11/01/1987 00:00:00.000} \code{11/01/1988 00:00:00.000} \code{11/01/1989 00:00:00.000} \code{11/01/1990 00:00:00.000} \code{11/01/1991 00:00:00.000} \code{11/01/1992 00:00:00.000} \code{11/01/1993 00:00:00.000} \code{11/01/1994 00:00:00.000} \code{11/01/1995 00:00:00.000} \code{11/01/1996 00:00:00.000} \code{11/01/1997 00:00:00.000} \code{11/01/1998 00:00:00.000} \code{11/08/1980 00:00:00.000} \code{11/08/1981 00:00:00.000} \code{11/08/1982 00:00:00.000} \code{11/08/1983 00:00:00.000} \code{11/08/1984 00:00:00.000} \code{11/08/1985 00:00:00.000} \code{11/08/1986 00:00:00.000} \code{11/08/1987 00:00:00.000} \code{11/08/1988 00:00:00.000} \code{11/08/1989 00:00:00.000} \code{11/08/1990 00:00:00.000} \code{11/08/1991 00:00:00.000} \code{11/08/1992 00:00:00.000} \code{11/08/1993 00:00:00.000} \code{11/08/1994 00:00:00.000} \code{11/08/1995 00:00:00.000} \code{11/08/1996 00:00:00.000} \code{11/08/1997 00:00:00.000} \code{11/08/1998 00:00:00.000} \code{11/15/1980 00:00:00.000} \code{11/15/1981 00:00:00.000} \code{11/15/1982 00:00:00.000} \code{11/15/1983 00:00:00.000} \code{11/15/1984 00:00:00.000} \code{11/15/1985 00:00:00.000} \code{11/15/1986 00:00:00.000} \code{11/15/1987 00:00:00.000} \code{11/15/1988 00:00:00.000} \code{11/15/1989 00:00:00.000} \code{11/15/1990 00:00:00.000} \code{11/15/1991 00:00:00.000} \code{11/15/1992 00:00:00.000} \code{11/15/1993 00:00:00.000} \code{11/15/1994 00:00:00.000} \code{11/15/1995 00:00:00.000} \code{11/15/1996 00:00:00.000} \code{11/15/1997 00:00:00.000} \code{11/15/1998 00:00:00.000} \code{11/22/1980 00:00:00.000} \code{11/22/1981 00:00:00.000} \code{11/22/1982 00:00:00.000} \code{11/22/1983 00:00:00.000} \code{11/22/1984 00:00:00.000} \code{11/22/1985 00:00:00.000} \code{11/22/1986 00:00:00.000} \code{11/22/1987 00:00:00.000} \code{11/22/1988 00:00:00.000} \code{11/22/1989 00:00:00.000} \code{11/22/1990 00:00:00.000} \code{11/22/1991 00:00:00.000} \code{11/22/1992 00:00:00.000} \code{11/22/1993 00:00:00.000} \code{11/22/1994 00:00:00.000} \code{11/22/1995 00:00:00.000} \code{11/22/1996 00:00:00.000} \code{11/22/1997 00:00:00.000} \code{11/22/1998 00:00:00.000} \code{11/29/1981 00:00:00.000} \code{11/29/1982 00:00:00.000} \code{11/29/1983 00:00:00.000} \code{11/29/1984 00:00:00.000} \code{11/29/1985 00:00:00.000} \code{11/29/1987 00:00:00.000} \code{11/29/1988 00:00:00.000} \code{11/29/1989 00:00:00.000} \code{11/29/1990 00:00:00.000} \code{11/29/1991 00:00:00.000} \code{11/29/1992 00:00:00.000} \code{11/29/1993 00:00:00.000} \code{11/29/1994 00:00:00.000} \code{11/29/1995 00:00:00.000} \code{11/29/1996 00:00:00.000} \code{11/29/1998 00:00:00.000} \code{12/01/1980 00:00:00.000} \code{12/01/1981 00:00:00.000} \code{12/01/1982 00:00:00.000} \code{12/01/1983 00:00:00.000} \code{12/01/1984 00:00:00.000} \code{12/01/1985 00:00:00.000} \code{12/01/1986 00:00:00.000} \code{12/01/1987 00:00:00.000} \code{12/01/1988 00:00:00.000} \code{12/01/1989 00:00:00.000} \code{12/01/1990 00:00:00.000} \code{12/01/1991 00:00:00.000} \code{12/01/1992 00:00:00.000} \code{12/01/1993 00:00:00.000} \code{12/01/1994 00:00:00.000} \code{12/01/1995 00:00:00.000} \code{12/01/1996 00:00:00.000} \code{12/01/1997 00:00:00.000} \code{12/01/1998 00:00:00.000} \code{12/08/1980 00:00:00.000} \code{12/08/1981 00:00:00.000} \code{12/08/1982 00:00:00.000} \code{12/08/1983 00:00:00.000} \code{12/08/1984 00:00:00.000} \code{12/08/1985 00:00:00.000} \code{12/08/1986 00:00:00.000} \code{12/08/1987 00:00:00.000} \code{12/08/1988 00:00:00.000} \code{12/08/1989 00:00:00.000} \code{12/08/1990 00:00:00.000} \code{12/08/1991 00:00:00.000} \code{12/08/1992 00:00:00.000} \code{12/08/1993 00:00:00.000} \code{12/08/1994 00:00:00.000} \code{12/08/1995 00:00:00.000} \code{12/08/1996 00:00:00.000} \code{12/08/1997 00:00:00.000} \code{12/08/1998 00:00:00.000} \code{12/15/1980 00:00:00.000} \code{12/15/1981 00:00:00.000} \code{12/15/1982 00:00:00.000} \code{12/15/1983 00:00:00.000} \code{12/15/1984 00:00:00.000} \code{12/15/1985 00:00:00.000} \code{12/15/1986 00:00:00.000} \code{12/15/1987 00:00:00.000} \code{12/15/1988 00:00:00.000} \code{12/15/1989 00:00:00.000} \code{12/15/1990 00:00:00.000} \code{12/15/1991 00:00:00.000} \code{12/15/1992 00:00:00.000} \code{12/15/1993 00:00:00.000} \code{12/15/1994 00:00:00.000} \code{12/15/1995 00:00:00.000} \code{12/15/1996 00:00:00.000} \code{12/15/1997 00:00:00.000} \code{12/15/1998 00:00:00.000} \code{12/22/1980 00:00:00.000} \code{12/22/1981 00:00:00.000} \code{12/22/1982 00:00:00.000} \code{12/22/1983 00:00:00.000} \code{12/22/1984 00:00:00.000} \code{12/22/1985 00:00:00.000} \code{12/22/1986 00:00:00.000} \code{12/22/1987 00:00:00.000} \code{12/22/1988 00:00:00.000} \code{12/22/1989 00:00:00.000} \code{12/22/1990 00:00:00.000} \code{12/22/1991 00:00:00.000} \code{12/22/1992 00:00:00.000} \code{12/22/1993 00:00:00.000} \code{12/22/1994 00:00:00.000} \code{12/22/1995 00:00:00.000} \code{12/22/1996 00:00:00.000} \code{12/22/1997 00:00:00.000} \code{12/22/1998 00:00:00.000} \code{12/29/1980 00:00:00.000} \code{12/29/1981 00:00:00.000} \code{12/29/1982 00:00:00.000} \code{12/29/1983 00:00:00.000} \code{12/29/1984 00:00:00.000} \code{12/29/1985 00:00:00.000} \code{12/29/1986 00:00:00.000} \code{12/29/1987 00:00:00.000} \code{12/29/1988 00:00:00.000} \code{12/29/1989 00:00:00.000} \code{12/29/1990 00:00:00.000} \code{12/29/1991 00:00:00.000} \code{12/29/1992 00:00:00.000} \code{12/29/1993 00:00:00.000} \code{12/29/1994 00:00:00.000} \code{12/29/1995 00:00:00.000} \code{12/29/1996 00:00:00.000} \code{12/29/1997 00:00:00.000} \code{12/29/1998 00:00:00.000}}
    \item{\code{Dutch.FX}}{a numeric vector}
  }
}

\source{
S+ FinMetrics DFX.ts
}
\references{
S+ FinMetrics DFX.ts
}
\examples{
data(DFX.ts)
## maybe str(DFX.ts) ; plot(DFX.ts) ...
}
\keyword{datasets}
