% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DJ30}
\alias{DJ30}
\title{Daily Dow Jones 30 stock price from January 1992 to September 1999}
\format{A \samp{xts} object on 1991-01-02/2001-01-02 containing:
\describe{
  \item{AA}{a numeric vector}
  \item{AXP}{a numeric vector}
  \item{BA}{a numeric vector}
  \item{C}{a numeric vector}
  \item{CAT}{a numeric vector}
  \item{DD}{a numeric vector}
  \item{DIS}{a numeric vector}
  \item{EK}{a numeric vector}
  \item{GE}{a numeric vector}
  \item{GM}{a numeric vector}
  \item{HD}{a numeric vector}
  \item{HON}{a numeric vector}
  \item{HWP}{a numeric vector}
  \item{IBM}{a numeric vector}
  \item{INTC}{a numeric vector}
  \item{IP}{a numeric vector}
  \item{JNJ}{a numeric vector}
  \item{JPM}{a numeric vector}
  \item{KO}{a numeric vector}
  \item{MCD}{a numeric vector}
  \item{MMM}{a numeric vector}
  \item{MO}{a numeric vector}
  \item{MRK}{a numeric vector}
  \item{MSFT}{a numeric vector}
  \item{PG}{a numeric vector}
  \item{SBC}{a numeric vector}
  \item{T}{a numeric vector}
  \item{UTX}{a numeric vector}
  \item{WMT}{a numeric vector}
  \item{XOM}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
DJ30
}
\description{
A \samp{xts} object of daily Dow Jones 30 stock price from January 1992 to September 1999.
}
\keyword{datasets}

