\name{djia}
\alias{djia}
\docType{data}
\title{
Dow Jone Industrial
Average over the period January 1, 1990 to February 20, 1990
}
\description{
Dow Jone Industrial
Average over the period January 1, 1990 to February 20, 1990
}
\usage{data(djia)}
\format{
  A data frame with 18840 observations on the following 6 variables.
}
\source{
S+ FinMetrics djia.
}
\references{
S+ FinMetrics djia.
}
\examples{
data(djia)
## maybe str(djia) ; plot(djia) ...
}
\keyword{datasets}
