% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DOC}
\alias{DOC}
\title{Monthly U.S. Department of Commerce data from January 1959 to January 1995}
\format{A \samp{xts} object on Jan 1959/Jan 1995 containing:
\describe{
  \item{IP}{a numeric vector}
  \item{PI}{a numeric vector}
  \item{Sales}{a numeric vector}
  \item{Payroll}{a numeric vector}
  \item{DOC}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
DOC
}
\description{
A \samp{xts} object of monthly industrial production (IP), total personal income less transfer
payments (PI), total manufacturing and trade sales (Sales), employees on nonagricultural payrolls
(Payroll), and Index of Coincident Economic Indicators (DOC) compiled by the U.S. Department of
Commerce from the period January 1959 to January 1995.
}
\keyword{datasets}

