\name{DowJones30}
\alias{DowJones30}
\docType{data}
\title{
DowJones30 stock price from 1/10/1992 to 9/9/1999
}
\description{
DowJones30 stock price from 1/10/1992 to 9/9/1999
}
\usage{data(DowJones30)}
\format{
  A data frame with 2528 observations on the following 31 variables.
}
\source{
S+ FinMetrics DowJones30.
}
\references{
S+ FinMetrics DowJones30.
}
\examples{
data(DowJones30)
## maybe str(DowJones30) ; plot(DowJones30) ...
}
\keyword{datasets}
