% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excessReturns}
\alias{excessReturns}
\title{Monthly excess returns on Microsoft and S&P 500 from February 1990 to December 2000}
\format{A \samp{xts} object on Feb 1990/Dec 2000 containing:
\describe{
  \item{MSFT}{a numeric vector}
  \item{SP500}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
excessReturns
}
\description{
A \samp{xts} object of monthly excess returns on Microsoft and S&P 500 from February 1990 to
December 2000.
}
\keyword{datasets}

