\name{excessReturns.ts}
\alias{excessReturns.ts}
\docType{data}
\title{
monthly excess return data on Microsoft stock and S&P 500 index 
over the ten year period February 1990 through December 2000
}
\description{
monthly excess return data on Microsoft stock and S&P 500 index
over the ten year period February 1990 through December 2000
}
\usage{data(excessReturns.ts)}
\format{
  A data frame with 131 observations on the following 2 variables.
  \describe{
    \item{\code{MSFT}}{a numeric vector}
    \item{\code{SP500}}{a numeric vector}
  }
}

\source{
S+ FinMetric "excessReturns.ts"
}
\references{
S+ FinMetric "excessReturns.ts"
}
\examples{
data(excessReturns.ts)
## maybe str(excessReturns.ts) ; plot(excessReturns.ts) ...
}
\keyword{datasets}
