% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fittedewmaVol.R
\name{fitted.ewmaVol}
\alias{fitted.ewmaVol}
\title{fitted.ewmavol()}
\usage{
\method{fitted}{ewmaVol}(object)
}
\arguments{
\item{object}{object}
}
\value{
object$sigma
}
\description{
fitted.ewmavol()
}
\examples{
\dontrun{
fitted.ewmavol(r)
}
}
\author{
Eric Zivot
}

