\name{folio.dat}
\alias{folio.dat}
\docType{data}
\title{
Weekly data on 1618 stocks over the period January 8, 1997
to June 28, 2000.
}
\description{
Weekly data on 1618 stocks over the period January 8, 1997
to June 28, 2000.
}
\usage{data(folio.dat)}
\format{
  A data frame with 182 observations on the following 1619 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{01/05/2000} \code{01/06/1999} \code{01/07/1998} \code{01/08/1997} \code{01/12/2000} \code{01/13/1999} \code{01/14/1998} \code{01/15/1997} \code{01/19/2000} \code{01/20/1999} \code{01/21/1998} \code{01/22/1997} \code{01/26/2000} \code{01/27/1999} \code{01/28/1998} \code{01/29/1997} \code{02/02/2000} \code{02/03/1999} \code{02/04/1998} \code{02/05/1997} \code{02/09/2000} \code{02/10/1999} \code{02/11/1998} \code{02/12/1997} \code{02/16/2000} \code{02/17/1999} \code{02/18/1998} \code{02/19/1997} \code{02/23/2000} \code{02/24/1999} \code{02/25/1998} \code{02/26/1997} \code{03/01/2000} \code{03/03/1999} \code{03/04/1998} \code{03/05/1997} \code{03/08/2000} \code{03/10/1999} \code{03/11/1998} \code{03/12/1997} \code{03/15/2000} \code{03/17/1999} \code{03/18/1998} \code{03/19/1997} \code{03/22/2000} \code{03/24/1999} \code{03/25/1998} \code{03/26/1997} \code{03/29/2000} \code{03/31/1999} \code{04/01/1998} \code{04/02/1997} \code{04/05/2000} \code{04/07/1999} \code{04/08/1998} \code{04/09/1997} \code{04/12/2000} \code{04/14/1999} \code{04/15/1998} \code{04/16/1997} \code{04/19/2000} \code{04/21/1999} \code{04/22/1998} \code{04/23/1997} \code{04/26/2000} \code{04/28/1999} \code{04/29/1998} \code{04/30/1997} \code{05/03/2000} \code{05/05/1999} \code{05/06/1998} \code{05/07/1997} \code{05/10/2000} \code{05/12/1999} \code{05/13/1998} \code{05/14/1997} \code{05/17/2000} \code{05/19/1999} \code{05/20/1998} \code{05/21/1997} \code{05/24/2000} \code{05/26/1999} \code{05/27/1998} \code{05/28/1997} \code{05/31/2000} \code{06/02/1999} \code{06/03/1998} \code{06/04/1997} \code{06/07/2000} \code{06/09/1999} \code{06/10/1998} \code{06/11/1997} \code{06/14/2000} \code{06/16/1999} \code{06/17/1998} \code{06/18/1997} \code{06/21/2000} \code{06/23/1999} \code{06/24/1998} \code{06/25/1997} \code{06/28/2000} \code{06/30/1999} \code{07/01/1998} \code{07/02/1997} \code{07/07/1999} \code{07/08/1998} \code{07/09/1997} \code{07/14/1999} \code{07/15/1998} \code{07/16/1997} \code{07/21/1999} \code{07/22/1998} \code{07/23/1997} \code{07/28/1999} \code{07/29/1998} \code{07/30/1997} \code{08/04/1999} \code{08/05/1998} \code{08/06/1997} \code{08/11/1999} \code{08/12/1998} \code{08/13/1997} \code{08/18/1999} \code{08/19/1998} \code{08/20/1997} \code{08/25/1999} \code{08/26/1998} \code{08/27/1997} \code{09/01/1999} \code{09/02/1998} \code{09/03/1997} \code{09/08/1999} \code{09/09/1998} \code{09/10/1997} \code{09/15/1999} \code{09/16/1998} \code{09/17/1997} \code{09/22/1999} \code{09/23/1998} \code{09/24/1997} \code{09/29/1999} \code{09/30/1998} \code{10/01/1997} \code{10/06/1999} \code{10/07/1998} \code{10/08/1997} \code{10/13/1999} \code{10/14/1998} \code{10/15/1997} \code{10/20/1999} \code{10/21/1998} \code{10/22/1997} \code{10/27/1999} \code{10/28/1998} \code{10/29/1997} \code{11/03/1999} \code{11/04/1998} \code{11/05/1997} \code{11/10/1999} \code{11/11/1998} \code{11/12/1997} \code{11/17/1999} \code{11/18/1998} \code{11/19/1997} \code{11/24/1999} \code{11/25/1998} \code{11/26/1997} \code{12/01/1999} \code{12/02/1998} \code{12/03/1997} \code{12/08/1999} \code{12/09/1998} \code{12/10/1997} \code{12/15/1999} \code{12/16/1998} \code{12/17/1997} \code{12/22/1999} \code{12/23/1998} \code{12/24/1997} \code{12/29/1999} \code{12/30/1998} \code{12/31/1997}}
    \item{\code{IATV}}{a numeric vector}
    \item{\code{ADCT}}{a numeric vector}
    \item{\code{ADEX}}{a numeric vector}
    \item{\code{ABM}}{a numeric vector}
    \item{\code{ACTM}}{a numeric vector}
    \item{\code{AFL}}{a numeric vector}
    \item{\code{AG}}{a numeric vector}
    \item{\code{ATG}}{a numeric vector}
    \item{\code{AES}}{a numeric vector}
    \item{\code{AKS}}{a numeric vector}
    \item{\code{AML}}{a numeric vector}
    \item{\code{AMR}}{a numeric vector}
    \item{\code{APAC}}{a numeric vector}
    \item{\code{T}}{a numeric vector}
    \item{\code{ATSI}}{a numeric vector}
    \item{\code{AVTC}}{a numeric vector}
    \item{\code{AVX}}{a numeric vector}
    \item{\code{ABT}}{a numeric vector}
    \item{\code{ANF}}{a numeric vector}
    \item{\code{AK}}{a numeric vector}
    \item{\code{ART}}{a numeric vector}
    \item{\code{ACTL}}{a numeric vector}
    \item{\code{ACXM}}{a numeric vector}
    \item{\code{ADLAC}}{a numeric vector}
    \item{\code{ADTK}}{a numeric vector}
    \item{\code{ADBE}}{a numeric vector}
    \item{\code{ADTN}}{a numeric vector}
    \item{\code{ADVP}}{a numeric vector}
    \item{\code{ADIC}}{a numeric vector}
    \item{\code{ADLT}}{a numeric vector}
    \item{\code{AFCI}}{a numeric vector}
    \item{\code{ATIS}}{a numeric vector}
    \item{\code{AD}}{a numeric vector}
    \item{\code{AMD}}{a numeric vector}
    \item{\code{ADVNA}}{a numeric vector}
    \item{\code{AEIS}}{a numeric vector}
    \item{\code{ADVS}}{a numeric vector}
    \item{\code{AET}}{a numeric vector}
    \item{\code{AFFX}}{a numeric vector}
    \item{\code{APD}}{a numeric vector}
    \item{\code{ABF}}{a numeric vector}
    \item{\code{ARG}}{a numeric vector}
    \item{\code{ALK}}{a numeric vector}
    \item{\code{AIN}}{a numeric vector}
    \item{\code{ALB}}{a numeric vector}
    \item{\code{ACV}}{a numeric vector}
    \item{\code{ABS}}{a numeric vector}
    \item{\code{AA}}{a numeric vector}
    \item{\code{ALEX}}{a numeric vector}
    \item{\code{ALXN}}{a numeric vector}
    \item{\code{ALFA}}{a numeric vector}
    \item{\code{ALKS}}{a numeric vector}
    \item{\code{ALN}}{a numeric vector}
    \item{\code{AGN}}{a numeric vector}
    \item{\code{ALSC}}{a numeric vector}
    \item{\code{ALLP}}{a numeric vector}
    \item{\code{AFC}}{a numeric vector}
    \item{\code{ALL}}{a numeric vector}
    \item{\code{AT}}{a numeric vector}
    \item{\code{ALO}}{a numeric vector}
    \item{\code{ALTR}}{a numeric vector}
    \item{\code{AZA}}{a numeric vector}
    \item{\code{ABK}}{a numeric vector}
    \item{\code{AHC}}{a numeric vector}
    \item{\code{UHAL}}{a numeric vector}
    \item{\code{AOL}}{a numeric vector}
    \item{\code{AWA}}{a numeric vector}
    \item{\code{AEOS}}{a numeric vector}
    \item{\code{AEP}}{a numeric vector}
    \item{\code{AXP}}{a numeric vector}
    \item{\code{AFG}}{a numeric vector}
    \item{\code{AFWY}}{a numeric vector}
    \item{\code{AGC}}{a numeric vector}
    \item{\code{AHP}}{a numeric vector}
    \item{\code{AIG}}{a numeric vector}
    \item{\code{AMSY}}{a numeric vector}
    \item{\code{APCC}}{a numeric vector}
    \item{\code{ASD}}{a numeric vector}
    \item{\code{AWK}}{a numeric vector}
    \item{\code{ACF}}{a numeric vector}
    \item{\code{AAS}}{a numeric vector}
    \item{\code{AMES}}{a numeric vector}
    \item{\code{AMGN}}{a numeric vector}
    \item{\code{APH}}{a numeric vector}
    \item{\code{ASO}}{a numeric vector}
    \item{\code{AMLN}}{a numeric vector}
    \item{\code{APC}}{a numeric vector}
    \item{\code{ANAD}}{a numeric vector}
    \item{\code{ADI}}{a numeric vector}
    \item{\code{ALOG}}{a numeric vector}
    \item{\code{ANLY}}{a numeric vector}
    \item{\code{SLOT}}{a numeric vector}
    \item{\code{ANDB}}{a numeric vector}
    \item{\code{ANDW}}{a numeric vector}
    \item{\code{ADRX}}{a numeric vector}
    \item{\code{BUD}}{a numeric vector}
    \item{\code{AXE}}{a numeric vector}
    \item{\code{ANN}}{a numeric vector}
    \item{\code{ANTC}}{a numeric vector}
    \item{\code{AOC}}{a numeric vector}
    \item{\code{APA}}{a numeric vector}
    \item{\code{AIV}}{a numeric vector}
    \item{\code{APHT}}{a numeric vector}
    \item{\code{APOL}}{a numeric vector}
    \item{\code{AAPL}}{a numeric vector}
    \item{\code{APPB}}{a numeric vector}
    \item{\code{AHG}}{a numeric vector}
    \item{\code{AMAT}}{a numeric vector}
    \item{\code{ASTX}}{a numeric vector}
    \item{\code{ATR}}{a numeric vector}
    \item{\code{ADM}}{a numeric vector}
    \item{\code{ACAT}}{a numeric vector}
    \item{\code{ARI}}{a numeric vector}
    \item{\code{AGII}}{a numeric vector}
    \item{\code{AGY}}{a numeric vector}
    \item{\code{ARIA}}{a numeric vector}
    \item{\code{ABFS}}{a numeric vector}
    \item{\code{ACK}}{a numeric vector}
    \item{\code{AIND}}{a numeric vector}
    \item{\code{ARW}}{a numeric vector}
    \item{\code{ARRO}}{a numeric vector}
    \item{\code{ARTC}}{a numeric vector}
    \item{\code{ASH}}{a numeric vector}
    \item{\code{ASPT}}{a numeric vector}
    \item{\code{AZPN}}{a numeric vector}
    \item{\code{ASBC}}{a numeric vector}
    \item{\code{ASFC}}{a numeric vector}
    \item{\code{ASYT}}{a numeric vector}
    \item{\code{ACAI}}{a numeric vector}
    \item{\code{ATML}}{a numeric vector}
    \item{\code{ATO}}{a numeric vector}
    \item{\code{ADSK}}{a numeric vector}
    \item{\code{ADP}}{a numeric vector}
    \item{\code{AZO}}{a numeric vector}
    \item{\code{AVNT}}{a numeric vector}
    \item{\code{AVY}}{a numeric vector}
    \item{\code{AVID}}{a numeric vector}
    \item{\code{AVIR}}{a numeric vector}
    \item{\code{AVT}}{a numeric vector}
    \item{\code{AVP}}{a numeric vector}
    \item{\code{AWRE}}{a numeric vector}
    \item{\code{AZR}}{a numeric vector}
    \item{\code{BBT}}{a numeric vector}
    \item{\code{BSYS}}{a numeric vector}
    \item{\code{BJS}}{a numeric vector}
    \item{\code{BRE}}{a numeric vector}
    \item{\code{BMCS}}{a numeric vector}
    \item{\code{BHI}}{a numeric vector}
    \item{\code{BEZ}}{a numeric vector}
    \item{\code{BLL}}{a numeric vector}
    \item{\code{BDG}}{a numeric vector}
    \item{\code{BAC}}{a numeric vector}
    \item{\code{BK}}{a numeric vector}
    \item{\code{ONE}}{a numeric vector}
    \item{\code{BNKU}}{a numeric vector}
    \item{\code{BCR}}{a numeric vector}
    \item{\code{BKS}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{BRL}}{a numeric vector}
    \item{\code{BARZ}}{a numeric vector}
    \item{\code{BRR}}{a numeric vector}
    \item{\code{BSNX}}{a numeric vector}
    \item{\code{BMG}}{a numeric vector}
    \item{\code{BOL}}{a numeric vector}
    \item{\code{BAX}}{a numeric vector}
    \item{\code{BSC}}{a numeric vector}
    \item{\code{BEC}}{a numeric vector}
    \item{\code{BDX}}{a numeric vector}
    \item{\code{BBBY}}{a numeric vector}
    \item{\code{BED}}{a numeric vector}
    \item{\code{BOG}}{a numeric vector}
    \item{\code{BWC}}{a numeric vector}
    \item{\code{BHW}}{a numeric vector}
    \item{\code{BLS}}{a numeric vector}
    \item{\code{BLC}}{a numeric vector}
    \item{\code{BMS}}{a numeric vector}
    \item{\code{BHE}}{a numeric vector}
    \item{\code{BBC}}{a numeric vector}
    \item{\code{BKLY}}{a numeric vector}
    \item{\code{BRY}}{a numeric vector}
    \item{\code{BBY}}{a numeric vector}
    \item{\code{BS}}{a numeric vector}
    \item{\code{BILL}}{a numeric vector}
    \item{\code{BDY}}{a numeric vector}
    \item{\code{BIOA}}{a numeric vector}
    \item{\code{BTGC}}{a numeric vector}
    \item{\code{BCRX}}{a numeric vector}
    \item{\code{BGEN}}{a numeric vector}
    \item{\code{BMET}}{a numeric vector}
    \item{\code{BDK}}{a numeric vector}
    \item{\code{BBOX}}{a numeric vector}
    \item{\code{BKH}}{a numeric vector}
    \item{\code{EWB}}{a numeric vector}
    \item{\code{HRB}}{a numeric vector}
    \item{\code{BTH}}{a numeric vector}
    \item{\code{BOBE}}{a numeric vector}
    \item{\code{BA}}{a numeric vector}
    \item{\code{BCC}}{a numeric vector}
    \item{\code{BGP}}{a numeric vector}
    \item{\code{BWA}}{a numeric vector}
    \item{\code{BSX}}{a numeric vector}
    \item{\code{BOW}}{a numeric vector}
    \item{\code{BNE}}{a numeric vector}
    \item{\code{BYD}}{a numeric vector}
    \item{\code{BOY}}{a numeric vector}
    \item{\code{BDN}}{a numeric vector}
    \item{\code{BGG}}{a numeric vector}
    \item{\code{CELL}}{a numeric vector}
    \item{\code{EAT}}{a numeric vector}
    \item{\code{BMY}}{a numeric vector}
    \item{\code{BVSN}}{a numeric vector}
    \item{\code{BRKS}}{a numeric vector}
    \item{\code{BRKT}}{a numeric vector}
    \item{\code{TMBR}}{a numeric vector}
    \item{\code{BC}}{a numeric vector}
    \item{\code{BW}}{a numeric vector}
    \item{\code{BKI}}{a numeric vector}
    \item{\code{BCF}}{a numeric vector}
    \item{\code{BNI}}{a numeric vector}
    \item{\code{BR}}{a numeric vector}
    \item{\code{BPP}}{a numeric vector}
    \item{\code{BSH}}{a numeric vector}
    \item{\code{CHP}}{a numeric vector}
    \item{\code{CBL}}{a numeric vector}
    \item{\code{CCCG}}{a numeric vector}
    \item{\code{CBRL}}{a numeric vector}
    \item{\code{CCBL}}{a numeric vector}
    \item{\code{CDI}}{a numeric vector}
    \item{\code{CDWC}}{a numeric vector}
    \item{\code{CI}}{a numeric vector}
    \item{\code{CNL}}{a numeric vector}
    \item{\code{CMGI}}{a numeric vector}
    \item{\code{CMS}}{a numeric vector}
    \item{\code{CPY}}{a numeric vector}
    \item{\code{CNA}}{a numeric vector}
    \item{\code{CNF}}{a numeric vector}
    \item{\code{CSGS}}{a numeric vector}
    \item{\code{CSX}}{a numeric vector}
    \item{\code{CPTL}}{a numeric vector}
    \item{\code{CUNO}}{a numeric vector}
    \item{\code{CVS}}{a numeric vector}
    \item{\code{CVC}}{a numeric vector}
    \item{\code{CS}}{a numeric vector}
    \item{\code{CBT}}{a numeric vector}
    \item{\code{COG}}{a numeric vector}
    \item{\code{CACI}}{a numeric vector}
    \item{\code{CDN}}{a numeric vector}
    \item{\code{CLCI}}{a numeric vector}
    \item{\code{CCC}}{a numeric vector}
    \item{\code{CAMP}}{a numeric vector}
    \item{\code{CWT}}{a numeric vector}
    \item{\code{ELY}}{a numeric vector}
    \item{\code{CPN}}{a numeric vector}
    \item{\code{CBM}}{a numeric vector}
    \item{\code{CATP}}{a numeric vector}
    \item{\code{CPT}}{a numeric vector}
    \item{\code{CPB}}{a numeric vector}
    \item{\code{COF}}{a numeric vector}
    \item{\code{CMO}}{a numeric vector}
    \item{\code{CSAR}}{a numeric vector}
    \item{\code{CAH}}{a numeric vector}
    \item{\code{CSL}}{a numeric vector}
    \item{\code{CRS}}{a numeric vector}
    \item{\code{CRE}}{a numeric vector}
    \item{\code{CAR}}{a numeric vector}
    \item{\code{CGC}}{a numeric vector}
    \item{\code{CASY}}{a numeric vector}
    \item{\code{PWN}}{a numeric vector}
    \item{\code{POS}}{a numeric vector}
    \item{\code{CDX}}{a numeric vector}
    \item{\code{CAT}}{a numeric vector}
    \item{\code{CACOA}}{a numeric vector}
    \item{\code{CEGE}}{a numeric vector}
    \item{\code{CLTK}}{a numeric vector}
    \item{\code{CELG}}{a numeric vector}
    \item{\code{CNT}}{a numeric vector}
    \item{\code{CTX}}{a numeric vector}
    \item{\code{CENT}}{a numeric vector}
    \item{\code{CBC}}{a numeric vector}
    \item{\code{CTL}}{a numeric vector}
    \item{\code{CEPH}}{a numeric vector}
    \item{\code{CEN}}{a numeric vector}
    \item{\code{CERN}}{a numeric vector}
    \item{\code{CHB}}{a numeric vector}
    \item{\code{CHRS}}{a numeric vector}
    \item{\code{CMB}}{a numeric vector}
    \item{\code{CPJ}}{a numeric vector}
    \item{\code{CKFR}}{a numeric vector}
    \item{\code{CKP}}{a numeric vector}
    \item{\code{CAKE}}{a numeric vector}
    \item{\code{CCG}}{a numeric vector}
    \item{\code{CHE}}{a numeric vector}
    \item{\code{CSK}}{a numeric vector}
    \item{\code{CHK}}{a numeric vector}
    \item{\code{CHV}}{a numeric vector}
    \item{\code{CHCS}}{a numeric vector}
    \item{\code{CHH}}{a numeric vector}
    \item{\code{CQB}}{a numeric vector}
    \item{\code{CHIR}}{a numeric vector}
    \item{\code{CCN}}{a numeric vector}
    \item{\code{CB}}{a numeric vector}
    \item{\code{CHD}}{a numeric vector}
    \item{\code{CINF}}{a numeric vector}
    \item{\code{CIN}}{a numeric vector}
    \item{\code{CC}}{a numeric vector}
    \item{\code{CSCO}}{a numeric vector}
    \item{\code{CRUS}}{a numeric vector}
    \item{\code{CTAS}}{a numeric vector}
    \item{\code{CBCF}}{a numeric vector}
    \item{\code{CTXS}}{a numeric vector}
    \item{\code{CYN}}{a numeric vector}
    \item{\code{CLE}}{a numeric vector}
    \item{\code{CLC}}{a numeric vector}
    \item{\code{CMH}}{a numeric vector}
    \item{\code{CCU}}{a numeric vector}
    \item{\code{CLF}}{a numeric vector}
    \item{\code{CLX}}{a numeric vector}
    \item{\code{COA}}{a numeric vector}
    \item{\code{CGP}}{a numeric vector}
    \item{\code{COKE}}{a numeric vector}
    \item{\code{KO}}{a numeric vector}
    \item{\code{CCE}}{a numeric vector}
    \item{\code{CGNX}}{a numeric vector}
    \item{\code{COHR}}{a numeric vector}
    \item{\code{COHU}}{a numeric vector}
    \item{\code{KCP}}{a numeric vector}
    \item{\code{CL}}{a numeric vector}
    \item{\code{CKC}}{a numeric vector}
    \item{\code{CNB}}{a numeric vector}
    \item{\code{CLP}}{a numeric vector}
    \item{\code{COB}}{a numeric vector}
    \item{\code{CMCSK}}{a numeric vector}
    \item{\code{CDO}}{a numeric vector}
    \item{\code{CMA}}{a numeric vector}
    \item{\code{CBH}}{a numeric vector}
    \item{\code{CBSH}}{a numeric vector}
    \item{\code{CGI}}{a numeric vector}
    \item{\code{CFB}}{a numeric vector}
    \item{\code{CMC}}{a numeric vector}
    \item{\code{NNN}}{a numeric vector}
    \item{\code{CFBX}}{a numeric vector}
    \item{\code{CBSS}}{a numeric vector}
    \item{\code{CPQ}}{a numeric vector}
    \item{\code{CA}}{a numeric vector}
    \item{\code{CMNT}}{a numeric vector}
    \item{\code{CSC}}{a numeric vector}
    \item{\code{CPWR}}{a numeric vector}
    \item{\code{CRK}}{a numeric vector}
    \item{\code{CMVT}}{a numeric vector}
    \item{\code{CAG}}{a numeric vector}
    \item{\code{CHRZ}}{a numeric vector}
    \item{\code{LENS}}{a numeric vector}
    \item{\code{CEFT}}{a numeric vector}
    \item{\code{CCUR}}{a numeric vector}
    \item{\code{CNMD}}{a numeric vector}
    \item{\code{CNCT}}{a numeric vector}
    \item{\code{CNC}}{a numeric vector}
    \item{\code{ED}}{a numeric vector}
    \item{\code{COP}}{a numeric vector}
    \item{\code{CNS}}{a numeric vector}
    \item{\code{COO}}{a numeric vector}
    \item{\code{CBE}}{a numeric vector}
    \item{\code{CTB}}{a numeric vector}
    \item{\code{CPRT}}{a numeric vector}
    \item{\code{GLW}}{a numeric vector}
    \item{\code{CPWM}}{a numeric vector}
    \item{\code{CCR}}{a numeric vector}
    \item{\code{CUZ}}{a numeric vector}
    \item{\code{COX}}{a numeric vector}
    \item{\code{CR}}{a numeric vector}
    \item{\code{CMOS}}{a numeric vector}
    \item{\code{CACC}}{a numeric vector}
    \item{\code{CREE}}{a numeric vector}
    \item{\code{CEI}}{a numeric vector}
    \item{\code{XTO}}{a numeric vector}
    \item{\code{CCK}}{a numeric vector}
    \item{\code{CUB}}{a numeric vector}
    \item{\code{CBST}}{a numeric vector}
    \item{\code{CUM}}{a numeric vector}
    \item{\code{CYBE}}{a numeric vector}
    \item{\code{CYGN}}{a numeric vector}
    \item{\code{CYMI}}{a numeric vector}
    \item{\code{CY}}{a numeric vector}
    \item{\code{CYT}}{a numeric vector}
    \item{\code{CYTO}}{a numeric vector}
    \item{\code{CYTC}}{a numeric vector}
    \item{\code{DQE}}{a numeric vector}
    \item{\code{DPL}}{a numeric vector}
    \item{\code{DHI}}{a numeric vector}
    \item{\code{DSPG}}{a numeric vector}
    \item{\code{DST}}{a numeric vector}
    \item{\code{DTE}}{a numeric vector}
    \item{\code{DVI}}{a numeric vector}
    \item{\code{DTL}}{a numeric vector}
    \item{\code{DS}}{a numeric vector}
    \item{\code{DCN}}{a numeric vector}
    \item{\code{DHR}}{a numeric vector}
    \item{\code{DRI}}{a numeric vector}
    \item{\code{DBCC}}{a numeric vector}
    \item{\code{DSCP}}{a numeric vector}
    \item{\code{DSTM}}{a numeric vector}
    \item{\code{TRL}}{a numeric vector}
    \item{\code{DAVX}}{a numeric vector}
    \item{\code{DF}}{a numeric vector}
    \item{\code{DE}}{a numeric vector}
    \item{\code{DELL}}{a numeric vector}
    \item{\code{DFG}}{a numeric vector}
    \item{\code{DLP}}{a numeric vector}
    \item{\code{DAL}}{a numeric vector}
    \item{\code{DEL}}{a numeric vector}
    \item{\code{DLX}}{a numeric vector}
    \item{\code{DRTE}}{a numeric vector}
    \item{\code{XRAY}}{a numeric vector}
    \item{\code{DDR}}{a numeric vector}
    \item{\code{DVN}}{a numeric vector}
    \item{\code{DV}}{a numeric vector}
    \item{\code{DP}}{a numeric vector}
    \item{\code{DL}}{a numeric vector}
    \item{\code{DMED}}{a numeric vector}
    \item{\code{DO}}{a numeric vector}
    \item{\code{DBD}}{a numeric vector}
    \item{\code{DIGE}}{a numeric vector}
    \item{\code{DDS}}{a numeric vector}
    \item{\code{DME}}{a numeric vector}
    \item{\code{DNEX}}{a numeric vector}
    \item{\code{DIS}}{a numeric vector}
    \item{\code{DCTM}}{a numeric vector}
    \item{\code{DOL}}{a numeric vector}
    \item{\code{DG}}{a numeric vector}
    \item{\code{DLTR}}{a numeric vector}
    \item{\code{D}}{a numeric vector}
    \item{\code{DCI}}{a numeric vector}
    \item{\code{DNY}}{a numeric vector}
    \item{\code{DOV}}{a numeric vector}
    \item{\code{DVD}}{a numeric vector}
    \item{\code{DOW}}{a numeric vector}
    \item{\code{DJ}}{a numeric vector}
    \item{\code{DSL}}{a numeric vector}
    \item{\code{DBRN}}{a numeric vector}
    \item{\code{DRYR}}{a numeric vector}
    \item{\code{DD}}{a numeric vector}
    \item{\code{DUK}}{a numeric vector}
    \item{\code{DRE}}{a numeric vector}
    \item{\code{DNB}}{a numeric vector}
    \item{\code{DPMI}}{a numeric vector}
    \item{\code{DUSA}}{a numeric vector}
    \item{\code{DY}}{a numeric vector}
    \item{\code{EEX}}{a numeric vector}
    \item{\code{EMC}}{a numeric vector}
    \item{\code{ESV}}{a numeric vector}
    \item{\code{EOG}}{a numeric vector}
    \item{\code{ESST}}{a numeric vector}
    \item{\code{EGRP}}{a numeric vector}
    \item{\code{EGR}}{a numeric vector}
    \item{\code{EGP}}{a numeric vector}
    \item{\code{EMN}}{a numeric vector}
    \item{\code{EK}}{a numeric vector}
    \item{\code{ETN}}{a numeric vector}
    \item{\code{EV}}{a numeric vector}
    \item{\code{DISH}}{a numeric vector}
    \item{\code{ECL}}{a numeric vector}
    \item{\code{EIX}}{a numeric vector}
    \item{\code{EDMC}}{a numeric vector}
    \item{\code{AGE}}{a numeric vector}
    \item{\code{EE}}{a numeric vector}
    \item{\code{EPG}}{a numeric vector}
    \item{\code{ELNT}}{a numeric vector}
    \item{\code{ELK}}{a numeric vector}
    \item{\code{ESIO}}{a numeric vector}
    \item{\code{EGLS}}{a numeric vector}
    \item{\code{ERTS}}{a numeric vector}
    \item{\code{EDS}}{a numeric vector}
    \item{\code{EFII}}{a numeric vector}
    \item{\code{EMR}}{a numeric vector}
    \item{\code{EMIS}}{a numeric vector}
    \item{\code{EMMS}}{a numeric vector}
    \item{\code{EDE}}{a numeric vector}
    \item{\code{EMLX}}{a numeric vector}
    \item{\code{ENDO}}{a numeric vector}
    \item{\code{EGN}}{a numeric vector}
    \item{\code{ENER}}{a numeric vector}
    \item{\code{EC}}{a numeric vector}
    \item{\code{EFS}}{a numeric vector}
    \item{\code{ENE}}{a numeric vector}
    \item{\code{ETR}}{a numeric vector}
    \item{\code{ENMD}}{a numeric vector}
    \item{\code{ENZN}}{a numeric vector}
    \item{\code{ENZ}}{a numeric vector}
    \item{\code{EFX}}{a numeric vector}
    \item{\code{EQT}}{a numeric vector}
    \item{\code{ENN}}{a numeric vector}
    \item{\code{EQR}}{a numeric vector}
    \item{\code{ERIE}}{a numeric vector}
    \item{\code{ESE}}{a numeric vector}
    \item{\code{ESS}}{a numeric vector}
    \item{\code{ESL}}{a numeric vector}
    \item{\code{ETH}}{a numeric vector}
    \item{\code{EY}}{a numeric vector}
    \item{\code{EXAR}}{a numeric vector}
    \item{\code{XLTC}}{a numeric vector}
    \item{\code{EXLN}}{a numeric vector}
    \item{\code{EX}}{a numeric vector}
    \item{\code{EXPD}}{a numeric vector}
    \item{\code{ESRX}}{a numeric vector}
    \item{\code{FMN}}{a numeric vector}
    \item{\code{FEIC}}{a numeric vector}
    \item{\code{FMC}}{a numeric vector}
    \item{\code{FBAN}}{a numeric vector}
    \item{\code{FPL}}{a numeric vector}
    \item{\code{FSII}}{a numeric vector}
    \item{\code{FDS}}{a numeric vector}
    \item{\code{FIC}}{a numeric vector}
    \item{\code{FFD}}{a numeric vector}
    \item{\code{FDO}}{a numeric vector}
    \item{\code{FAST}}{a numeric vector}
    \item{\code{FJC}}{a numeric vector}
    \item{\code{FRE}}{a numeric vector}
    \item{\code{FMO}}{a numeric vector}
    \item{\code{FNM}}{a numeric vector}
    \item{\code{FRT}}{a numeric vector}
    \item{\code{FSS}}{a numeric vector}
    \item{\code{FD}}{a numeric vector}
    \item{\code{FDX}}{a numeric vector}
    \item{\code{FCH}}{a numeric vector}
    \item{\code{FOE}}{a numeric vector}
    \item{\code{FNF}}{a numeric vector}
    \item{\code{FITB}}{a numeric vector}
    \item{\code{FILE}}{a numeric vector}
    \item{\code{FIF}}{a numeric vector}
    \item{\code{FNV}}{a numeric vector}
    \item{\code{FAF}}{a numeric vector}
    \item{\code{FBP}}{a numeric vector}
    \item{\code{FCF}}{a numeric vector}
    \item{\code{FTFC}}{a numeric vector}
    \item{\code{FDC}}{a numeric vector}
    \item{\code{FR}}{a numeric vector}
    \item{\code{FMBI}}{a numeric vector}
    \item{\code{FTU}}{a numeric vector}
    \item{\code{FVB}}{a numeric vector}
    \item{\code{FRW}}{a numeric vector}
    \item{\code{FSR}}{a numeric vector}
    \item{\code{FISV}}{a numeric vector}
    \item{\code{FED}}{a numeric vector}
    \item{\code{FMER}}{a numeric vector}
    \item{\code{FSH}}{a numeric vector}
    \item{\code{FLE}}{a numeric vector}
    \item{\code{FLM}}{a numeric vector}
    \item{\code{FLA}}{a numeric vector}
    \item{\code{FRK}}{a numeric vector}
    \item{\code{FLR}}{a numeric vector}
    \item{\code{FLO}}{a numeric vector}
    \item{\code{FTS}}{a numeric vector}
    \item{\code{F}}{a numeric vector}
    \item{\code{FRX}}{a numeric vector}
    \item{\code{FOSL}}{a numeric vector}
    \item{\code{FWC}}{a numeric vector}
    \item{\code{FFA}}{a numeric vector}
    \item{\code{BEN}}{a numeric vector}
    \item{\code{FCX}}{a numeric vector}
    \item{\code{FMT}}{a numeric vector}
    \item{\code{FRTZ}}{a numeric vector}
    \item{\code{FRNT}}{a numeric vector}
    \item{\code{FULL}}{a numeric vector}
    \item{\code{FULT}}{a numeric vector}
    \item{\code{FBN}}{a numeric vector}
    \item{\code{GKSRA}}{a numeric vector}
    \item{\code{GMT}}{a numeric vector}
    \item{\code{GCX}}{a numeric vector}
    \item{\code{GPU}}{a numeric vector}
    \item{\code{GBL}}{a numeric vector}
    \item{\code{GBP}}{a numeric vector}
    \item{\code{AJG}}{a numeric vector}
    \item{\code{GCI}}{a numeric vector}
    \item{\code{GPS}}{a numeric vector}
    \item{\code{GSNX}}{a numeric vector}
    \item{\code{GCR}}{a numeric vector}
    \item{\code{GY}}{a numeric vector}
    \item{\code{GNCMA}}{a numeric vector}
    \item{\code{GD}}{a numeric vector}
    \item{\code{GE}}{a numeric vector}
    \item{\code{GGP}}{a numeric vector}
    \item{\code{GIS}}{a numeric vector}
    \item{\code{GM}}{a numeric vector}
    \item{\code{GMH}}{a numeric vector}
    \item{\code{GCO}}{a numeric vector}
    \item{\code{GNTX}}{a numeric vector}
    \item{\code{GLYT}}{a numeric vector}
    \item{\code{GENE}}{a numeric vector}
    \item{\code{GEN}}{a numeric vector}
    \item{\code{GZTC}}{a numeric vector}
    \item{\code{GPC}}{a numeric vector}
    \item{\code{GENZ}}{a numeric vector}
    \item{\code{GGC}}{a numeric vector}
    \item{\code{GP}}{a numeric vector}
    \item{\code{GWRX}}{a numeric vector}
    \item{\code{GRB}}{a numeric vector}
    \item{\code{GERN}}{a numeric vector}
    \item{\code{ROCK}}{a numeric vector}
    \item{\code{GILD}}{a numeric vector}
    \item{\code{G}}{a numeric vector}
    \item{\code{GLT}}{a numeric vector}
    \item{\code{GEMS}}{a numeric vector}
    \item{\code{GLB}}{a numeric vector}
    \item{\code{GLIA}}{a numeric vector}
    \item{\code{GRT}}{a numeric vector}
    \item{\code{GLBL}}{a numeric vector}
    \item{\code{GLM}}{a numeric vector}
    \item{\code{GLDB}}{a numeric vector}
    \item{\code{GDW}}{a numeric vector}
    \item{\code{GR}}{a numeric vector}
    \item{\code{GT}}{a numeric vector}
    \item{\code{GGG}}{a numeric vector}
    \item{\code{GWW}}{a numeric vector}
    \item{\code{GAP}}{a numeric vector}
    \item{\code{GLK}}{a numeric vector}
    \item{\code{GPT}}{a numeric vector}
    \item{\code{GBCOA}}{a numeric vector}
    \item{\code{GFF}}{a numeric vector}
    \item{\code{GTK}}{a numeric vector}
    \item{\code{GES}}{a numeric vector}
    \item{\code{GDT}}{a numeric vector}
    \item{\code{GLFD}}{a numeric vector}
    \item{\code{HCC}}{a numeric vector}
    \item{\code{HNCS}}{a numeric vector}
    \item{\code{HRP}}{a numeric vector}
    \item{\code{HSE}}{a numeric vector}
    \item{\code{HAE}}{a numeric vector}
    \item{\code{HAL}}{a numeric vector}
    \item{\code{HDL}}{a numeric vector}
    \item{\code{HNV}}{a numeric vector}
    \item{\code{JH}}{a numeric vector}
    \item{\code{HDI}}{a numeric vector}
    \item{\code{HGIC}}{a numeric vector}
    \item{\code{HAR}}{a numeric vector}
    \item{\code{HLIT}}{a numeric vector}
    \item{\code{HET}}{a numeric vector}
    \item{\code{HRS}}{a numeric vector}
    \item{\code{HSC}}{a numeric vector}
    \item{\code{HHS}}{a numeric vector}
    \item{\code{HIG}}{a numeric vector}
    \item{\code{HAS}}{a numeric vector}
    \item{\code{HE}}{a numeric vector}
    \item{\code{HCP}}{a numeric vector}
    \item{\code{HRC}}{a numeric vector}
    \item{\code{HMA}}{a numeric vector}
    \item{\code{HR}}{a numeric vector}
    \item{\code{HCN}}{a numeric vector}
    \item{\code{HTLD}}{a numeric vector}
    \item{\code{HEI}}{a numeric vector}
    \item{\code{HNZ}}{a numeric vector}
    \item{\code{HELX}}{a numeric vector}
    \item{\code{HP}}{a numeric vector}
    \item{\code{JKHY}}{a numeric vector}
    \item{\code{HPC}}{a numeric vector}
    \item{\code{HSY}}{a numeric vector}
    \item{\code{HWP}}{a numeric vector}
    \item{\code{HXL}}{a numeric vector}
    \item{\code{HIB}}{a numeric vector}
    \item{\code{HIW}}{a numeric vector}
    \item{\code{HRH}}{a numeric vector}
    \item{\code{HB}}{a numeric vector}
    \item{\code{HLT}}{a numeric vector}
    \item{\code{HLR}}{a numeric vector}
    \item{\code{HLYW}}{a numeric vector}
    \item{\code{HD}}{a numeric vector}
    \item{\code{HME}}{a numeric vector}
    \item{\code{HM}}{a numeric vector}
    \item{\code{HON}}{a numeric vector}
    \item{\code{HH}}{a numeric vector}
    \item{\code{HMN}}{a numeric vector}
    \item{\code{HRL}}{a numeric vector}
    \item{\code{HPT}}{a numeric vector}
    \item{\code{HMT}}{a numeric vector}
    \item{\code{HOTT}}{a numeric vector}
    \item{\code{HTN}}{a numeric vector}
    \item{\code{HI}}{a numeric vector}
    \item{\code{THX}}{a numeric vector}
    \item{\code{HUBB}}{a numeric vector}
    \item{\code{HUG}}{a numeric vector}
    \item{\code{HUM}}{a numeric vector}
    \item{\code{HGSI}}{a numeric vector}
    \item{\code{JBHT}}{a numeric vector}
    \item{\code{HBAN}}{a numeric vector}
    \item{\code{HTCH}}{a numeric vector}
    \item{\code{ICN}}{a numeric vector}
    \item{\code{IDTC}}{a numeric vector}
    \item{\code{IBP}}{a numeric vector}
    \item{\code{ICOS}}{a numeric vector}
    \item{\code{IDPH}}{a numeric vector}
    \item{\code{IDXC}}{a numeric vector}
    \item{\code{IGEN}}{a numeric vector}
    \item{\code{IGL}}{a numeric vector}
    \item{\code{IRT}}{a numeric vector}
    \item{\code{STAT}}{a numeric vector}
    \item{\code{ESI}}{a numeric vector}
    \item{\code{IBIS}}{a numeric vector}
    \item{\code{IDA}}{a numeric vector}
    \item{\code{IEX}}{a numeric vector}
    \item{\code{IDXX}}{a numeric vector}
    \item{\code{IGTE}}{a numeric vector}
    \item{\code{IDX}}{a numeric vector}
    \item{\code{ITW}}{a numeric vector}
    \item{\code{IMN}}{a numeric vector}
    \item{\code{IMCL}}{a numeric vector}
    \item{\code{IMNR}}{a numeric vector}
    \item{\code{IMNX}}{a numeric vector}
    \item{\code{IMGN}}{a numeric vector}
    \item{\code{IMP}}{a numeric vector}
    \item{\code{IMAT}}{a numeric vector}
    \item{\code{IMMU}}{a numeric vector}
    \item{\code{IMRS}}{a numeric vector}
    \item{\code{INCY}}{a numeric vector}
    \item{\code{INDB}}{a numeric vector}
    \item{\code{INFS}}{a numeric vector}
    \item{\code{IR}}{a numeric vector}
    \item{\code{IMKTA}}{a numeric vector}
    \item{\code{IM}}{a numeric vector}
    \item{\code{INHL}}{a numeric vector}
    \item{\code{KPA}}{a numeric vector}
    \item{\code{NSIT}}{a numeric vector}
    \item{\code{IO}}{a numeric vector}
    \item{\code{INPR}}{a numeric vector}
    \item{\code{INSUA}}{a numeric vector}
    \item{\code{IFS}}{a numeric vector}
    \item{\code{IAAI}}{a numeric vector}
    \item{\code{IDTI}}{a numeric vector}
    \item{\code{ISSI}}{a numeric vector}
    \item{\code{INTD}}{a numeric vector}
    \item{\code{INTC}}{a numeric vector}
    \item{\code{INTL}}{a numeric vector}
    \item{\code{IFSIA}}{a numeric vector}
    \item{\code{INGR}}{a numeric vector}
    \item{\code{LINK}}{a numeric vector}
    \item{\code{ILXI}}{a numeric vector}
    \item{\code{ICIX}}{a numeric vector}
    \item{\code{IBM}}{a numeric vector}
    \item{\code{IFCI}}{a numeric vector}
    \item{\code{IFF}}{a numeric vector}
    \item{\code{IGT}}{a numeric vector}
    \item{\code{IMC}}{a numeric vector}
    \item{\code{IP}}{a numeric vector}
    \item{\code{IRF}}{a numeric vector}
    \item{\code{ISCA}}{a numeric vector}
    \item{\code{ISP}}{a numeric vector}
    \item{\code{IPG}}{a numeric vector}
    \item{\code{IBC}}{a numeric vector}
    \item{\code{INRS}}{a numeric vector}
    \item{\code{ITN}}{a numeric vector}
    \item{\code{INTV}}{a numeric vector}
    \item{\code{IBI}}{a numeric vector}
    \item{\code{INTU}}{a numeric vector}
    \item{\code{INTZ}}{a numeric vector}
    \item{\code{IOM}}{a numeric vector}
    \item{\code{ION}}{a numeric vector}
    \item{\code{IPL}}{a numeric vector}
    \item{\code{ISIP}}{a numeric vector}
    \item{\code{ITWO}}{a numeric vector}
    \item{\code{IVX}}{a numeric vector}
    \item{\code{JDN}}{a numeric vector}
    \item{\code{JDAS}}{a numeric vector}
    \item{\code{JLG}}{a numeric vector}
    \item{\code{JPR}}{a numeric vector}
    \item{\code{JEC}}{a numeric vector}
    \item{\code{JAKK}}{a numeric vector}
    \item{\code{JEF}}{a numeric vector}
    \item{\code{JP}}{a numeric vector}
    \item{\code{JNC}}{a numeric vector}
    \item{\code{JNJ}}{a numeric vector}
    \item{\code{JCI}}{a numeric vector}
    \item{\code{JNY}}{a numeric vector}
    \item{\code{KLAC}}{a numeric vector}
    \item{\code{KM}}{a numeric vector}
    \item{\code{KLU}}{a numeric vector}
    \item{\code{KAMNA}}{a numeric vector}
    \item{\code{KLT}}{a numeric vector}
    \item{\code{KSU}}{a numeric vector}
    \item{\code{KBH}}{a numeric vector}
    \item{\code{KDN}}{a numeric vector}
    \item{\code{KEA}}{a numeric vector}
    \item{\code{KEI}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{KWD}}{a numeric vector}
    \item{\code{KELYA}}{a numeric vector}
    \item{\code{KMT}}{a numeric vector}
    \item{\code{KNT}}{a numeric vector}
    \item{\code{KMG}}{a numeric vector}
    \item{\code{KEG}}{a numeric vector}
    \item{\code{KP}}{a numeric vector}
    \item{\code{KEY}}{a numeric vector}
    \item{\code{KFRC}}{a numeric vector}
    \item{\code{KBALB}}{a numeric vector}
    \item{\code{KMB}}{a numeric vector}
    \item{\code{KIM}}{a numeric vector}
    \item{\code{KEX}}{a numeric vector}
    \item{\code{KRI}}{a numeric vector}
    \item{\code{KNGT}}{a numeric vector}
    \item{\code{KE}}{a numeric vector}
    \item{\code{KSS}}{a numeric vector}
    \item{\code{KOPN}}{a numeric vector}
    \item{\code{KR}}{a numeric vector}
    \item{\code{KRON}}{a numeric vector}
    \item{\code{KLIC}}{a numeric vector}
    \item{\code{LCCI}}{a numeric vector}
    \item{\code{LTV}}{a numeric vector}
    \item{\code{LSI}}{a numeric vector}
    \item{\code{LTXX}}{a numeric vector}
    \item{\code{LZB}}{a numeric vector}
    \item{\code{LH}}{a numeric vector}
    \item{\code{LG}}{a numeric vector}
    \item{\code{LAF}}{a numeric vector}
    \item{\code{LRCX}}{a numeric vector}
    \item{\code{LAMR}}{a numeric vector}
    \item{\code{LANC}}{a numeric vector}
    \item{\code{LNCE}}{a numeric vector}
    \item{\code{LE}}{a numeric vector}
    \item{\code{LSTR}}{a numeric vector}
    \item{\code{LSCC}}{a numeric vector}
    \item{\code{EL}}{a numeric vector}
    \item{\code{LAWS}}{a numeric vector}
    \item{\code{LEA}}{a numeric vector}
    \item{\code{LTRE}}{a numeric vector}
    \item{\code{LEE}}{a numeric vector}
    \item{\code{LGTO}}{a numeric vector}
    \item{\code{LEG}}{a numeric vector}
    \item{\code{LM}}{a numeric vector}
    \item{\code{LEH}}{a numeric vector}
    \item{\code{LEN}}{a numeric vector}
    \item{\code{LUK}}{a numeric vector}
    \item{\code{LVEL}}{a numeric vector}
    \item{\code{LXP}}{a numeric vector}
    \item{\code{LXK}}{a numeric vector}
    \item{\code{LBY}}{a numeric vector}
    \item{\code{LC}}{a numeric vector}
    \item{\code{L}}{a numeric vector}
    \item{\code{LRY}}{a numeric vector}
    \item{\code{LGND}}{a numeric vector}
    \item{\code{LTBG}}{a numeric vector}
    \item{\code{LLY}}{a numeric vector}
    \item{\code{LTD}}{a numeric vector}
    \item{\code{LNCR}}{a numeric vector}
    \item{\code{LECO}}{a numeric vector}
    \item{\code{LNC}}{a numeric vector}
    \item{\code{LLTC}}{a numeric vector}
    \item{\code{LIN}}{a numeric vector}
    \item{\code{LFUS}}{a numeric vector}
    \item{\code{LIT}}{a numeric vector}
    \item{\code{LIZ}}{a numeric vector}
    \item{\code{LMT}}{a numeric vector}
    \item{\code{LNET}}{a numeric vector}
    \item{\code{LTR}}{a numeric vector}
    \item{\code{STAR}}{a numeric vector}
    \item{\code{LDG}}{a numeric vector}
    \item{\code{LFB}}{a numeric vector}
    \item{\code{LD}}{a numeric vector}
    \item{\code{LPX}}{a numeric vector}
    \item{\code{LOW}}{a numeric vector}
    \item{\code{LZ}}{a numeric vector}
    \item{\code{LUB}}{a numeric vector}
    \item{\code{LU}}{a numeric vector}
    \item{\code{LYO}}{a numeric vector}
    \item{\code{MAFB}}{a numeric vector}
    \item{\code{MBI}}{a numeric vector}
    \item{\code{KRB}}{a numeric vector}
    \item{\code{MCN}}{a numeric vector}
    \item{\code{MDC}}{a numeric vector}
    \item{\code{MDU}}{a numeric vector}
    \item{\code{WFR}}{a numeric vector}
    \item{\code{MTG}}{a numeric vector}
    \item{\code{MOGN}}{a numeric vector}
    \item{\code{MGG}}{a numeric vector}
    \item{\code{MRVC}}{a numeric vector}
    \item{\code{MSM}}{a numeric vector}
    \item{\code{MSCA}}{a numeric vector}
    \item{\code{MTIC}}{a numeric vector}
    \item{\code{MAC}}{a numeric vector}
    \item{\code{CLI}}{a numeric vector}
    \item{\code{MACR}}{a numeric vector}
    \item{\code{SHOO}}{a numeric vector}
    \item{\code{MDSN}}{a numeric vector}
    \item{\code{MAG}}{a numeric vector}
    \item{\code{MWL}}{a numeric vector}
    \item{\code{MTW}}{a numeric vector}
    \item{\code{HCR}}{a numeric vector}
    \item{\code{MAN}}{a numeric vector}
    \item{\code{MHC}}{a numeric vector}
    \item{\code{MANU}}{a numeric vector}
    \item{\code{MAPS}}{a numeric vector}
    \item{\code{MRCH}}{a numeric vector}
    \item{\code{MCS}}{a numeric vector}
    \item{\code{MMC}}{a numeric vector}
    \item{\code{MAR}}{a numeric vector}
    \item{\code{MATK}}{a numeric vector}
    \item{\code{MLM}}{a numeric vector}
    \item{\code{MAS}}{a numeric vector}
    \item{\code{MATX}}{a numeric vector}
    \item{\code{MAT}}{a numeric vector}
    \item{\code{MTSN}}{a numeric vector}
    \item{\code{MXIM}}{a numeric vector}
    \item{\code{MAY}}{a numeric vector}
    \item{\code{MYG}}{a numeric vector}
    \item{\code{MNI}}{a numeric vector}
    \item{\code{MCD}}{a numeric vector}
    \item{\code{MHP}}{a numeric vector}
    \item{\code{MCK}}{a numeric vector}
    \item{\code{MCLD}}{a numeric vector}
    \item{\code{MEA}}{a numeric vector}
    \item{\code{MEDX}}{a numeric vector}
    \item{\code{TAXI}}{a numeric vector}
    \item{\code{MDEA}}{a numeric vector}
    \item{\code{MEGA}}{a numeric vector}
    \item{\code{MAI}}{a numeric vector}
    \item{\code{MEDI}}{a numeric vector}
    \item{\code{MT}}{a numeric vector}
    \item{\code{MDT}}{a numeric vector}
    \item{\code{MEL}}{a numeric vector}
    \item{\code{MBRS}}{a numeric vector}
    \item{\code{MNTR}}{a numeric vector}
    \item{\code{MENT}}{a numeric vector}
    \item{\code{MRBK}}{a numeric vector}
    \item{\code{MRK}}{a numeric vector}
    \item{\code{MCY}}{a numeric vector}
    \item{\code{MERQ}}{a numeric vector}
    \item{\code{MDP}}{a numeric vector}
    \item{\code{TMR}}{a numeric vector}
    \item{\code{MER}}{a numeric vector}
    \item{\code{MESA}}{a numeric vector}
    \item{\code{MAIR}}{a numeric vector}
    \item{\code{METHA}}{a numeric vector}
    \item{\code{MCOM}}{a numeric vector}
    \item{\code{MCLL}}{a numeric vector}
    \item{\code{MMG}}{a numeric vector}
    \item{\code{MIKE}}{a numeric vector}
    \item{\code{MCRL}}{a numeric vector}
    \item{\code{MCRS}}{a numeric vector}
    \item{\code{MSFT}}{a numeric vector}
    \item{\code{MVIS}}{a numeric vector}
    \item{\code{MCHP}}{a numeric vector}
    \item{\code{MUEI}}{a numeric vector}
    \item{\code{MU}}{a numeric vector}
    \item{\code{MSCC}}{a numeric vector}
    \item{\code{MAA}}{a numeric vector}
    \item{\code{MME}}{a numeric vector}
    \item{\code{MRS}}{a numeric vector}
    \item{\code{MEH}}{a numeric vector}
    \item{\code{MWY}}{a numeric vector}
    \item{\code{MLNM}}{a numeric vector}
    \item{\code{MCH}}{a numeric vector}
    \item{\code{MLHR}}{a numeric vector}
    \item{\code{MIL}}{a numeric vector}
    \item{\code{MLS}}{a numeric vector}
    \item{\code{MTX}}{a numeric vector}
    \item{\code{MNMD}}{a numeric vector}
    \item{\code{MMM}}{a numeric vector}
    \item{\code{MINI}}{a numeric vector}
    \item{\code{MODI}}{a numeric vector}
    \item{\code{MDCC}}{a numeric vector}
    \item{\code{MOLX}}{a numeric vector}
    \item{\code{MOND}}{a numeric vector}
    \item{\code{MTP}}{a numeric vector}
    \item{\code{MOGA}}{a numeric vector}
    \item{\code{JPM}}{a numeric vector}
    \item{\code{MOR}}{a numeric vector}
    \item{\code{MHI}}{a numeric vector}
    \item{\code{MTNT}}{a numeric vector}
    \item{\code{MOT}}{a numeric vector}
    \item{\code{MLI}}{a numeric vector}
    \item{\code{MUR}}{a numeric vector}
    \item{\code{MLG}}{a numeric vector}
    \item{\code{MYE}}{a numeric vector}
    \item{\code{MYL}}{a numeric vector}
    \item{\code{MYGN}}{a numeric vector}
    \item{\code{NABI}}{a numeric vector}
    \item{\code{NBTY}}{a numeric vector}
    \item{\code{NCH}}{a numeric vector}
    \item{\code{NCOG}}{a numeric vector}
    \item{\code{NCR}}{a numeric vector}
    \item{\code{NSS}}{a numeric vector}
    \item{\code{NL}}{a numeric vector}
    \item{\code{NPCI}}{a numeric vector}
    \item{\code{NUI}}{a numeric vector}
    \item{\code{NVR}}{a numeric vector}
    \item{\code{NBR}}{a numeric vector}
    \item{\code{NC}}{a numeric vector}
    \item{\code{NANO}}{a numeric vector}
    \item{\code{NCC}}{a numeric vector}
    \item{\code{NCBC}}{a numeric vector}
    \item{\code{NDC}}{a numeric vector}
    \item{\code{NFG}}{a numeric vector}
    \item{\code{TEE}}{a numeric vector}
    \item{\code{NHI}}{a numeric vector}
    \item{\code{NATI}}{a numeric vector}
    \item{\code{NOI}}{a numeric vector}
    \item{\code{NPK}}{a numeric vector}
    \item{\code{NAP}}{a numeric vector}
    \item{\code{NSM}}{a numeric vector}
    \item{\code{NSI}}{a numeric vector}
    \item{\code{NS}}{a numeric vector}
    \item{\code{NHP}}{a numeric vector}
    \item{\code{NMSS}}{a numeric vector}
    \item{\code{NAUT}}{a numeric vector}
    \item{\code{NAV}}{a numeric vector}
    \item{\code{NERX}}{a numeric vector}
    \item{\code{NTEC}}{a numeric vector}
    \item{\code{NETM}}{a numeric vector}
    \item{\code{NTAP}}{a numeric vector}
    \item{\code{NWK}}{a numeric vector}
    \item{\code{NPIX}}{a numeric vector}
    \item{\code{NRGN}}{a numeric vector}
    \item{\code{NBIX}}{a numeric vector}
    \item{\code{NEB}}{a numeric vector}
    \item{\code{NJR}}{a numeric vector}
    \item{\code{NWL}}{a numeric vector}
    \item{\code{NFX}}{a numeric vector}
    \item{\code{NEM}}{a numeric vector}
    \item{\code{NR}}{a numeric vector}
    \item{\code{NEWP}}{a numeric vector}
    \item{\code{NNS}}{a numeric vector}
    \item{\code{NMK}}{a numeric vector}
    \item{\code{GAS}}{a numeric vector}
    \item{\code{NKE}}{a numeric vector}
    \item{\code{NDN}}{a numeric vector}
    \item{\code{NI}}{a numeric vector}
    \item{\code{NBL}}{a numeric vector}
    \item{\code{NE}}{a numeric vector}
    \item{\code{NDSN}}{a numeric vector}
    \item{\code{NSC}}{a numeric vector}
    \item{\code{NTK}}{a numeric vector}
    \item{\code{NFB}}{a numeric vector}
    \item{\code{NU}}{a numeric vector}
    \item{\code{NTRS}}{a numeric vector}
    \item{\code{NOC}}{a numeric vector}
    \item{\code{NWAC}}{a numeric vector}
    \item{\code{NWSB}}{a numeric vector}
    \item{\code{NIS}}{a numeric vector}
    \item{\code{NVDM}}{a numeric vector}
    \item{\code{NOVL}}{a numeric vector}
    \item{\code{NVLS}}{a numeric vector}
    \item{\code{NOVN}}{a numeric vector}
    \item{\code{NUS}}{a numeric vector}
    \item{\code{NUE}}{a numeric vector}
    \item{\code{NEV}}{a numeric vector}
    \item{\code{CHUX}}{a numeric vector}
    \item{\code{OGE}}{a numeric vector}
    \item{\code{OAKT}}{a numeric vector}
    \item{\code{OO}}{a numeric vector}
    \item{\code{OSII}}{a numeric vector}
    \item{\code{OXY}}{a numeric vector}
    \item{\code{OII}}{a numeric vector}
    \item{\code{OCFC}}{a numeric vector}
    \item{\code{OMX}}{a numeric vector}
    \item{\code{ODP}}{a numeric vector}
    \item{\code{OLOG}}{a numeric vector}
    \item{\code{OG}}{a numeric vector}
    \item{\code{OCAS}}{a numeric vector}
    \item{\code{OLDB}}{a numeric vector}
    \item{\code{ORI}}{a numeric vector}
    \item{\code{OLN}}{a numeric vector}
    \item{\code{OCR}}{a numeric vector}
    \item{\code{OMC}}{a numeric vector}
    \item{\code{ASGN}}{a numeric vector}
    \item{\code{OCQ}}{a numeric vector}
    \item{\code{OKE}}{a numeric vector}
    \item{\code{OMKT}}{a numeric vector}
    \item{\code{OCCF}}{a numeric vector}
    \item{\code{ORCL}}{a numeric vector}
    \item{\code{ORG}}{a numeric vector}
    \item{\code{ORLY}}{a numeric vector}
    \item{\code{GOSHA}}{a numeric vector}
    \item{\code{FIBR}}{a numeric vector}
    \item{\code{OTTR}}{a numeric vector}
    \item{\code{OSG}}{a numeric vector}
    \item{\code{OWC}}{a numeric vector}
    \item{\code{OMI}}{a numeric vector}
    \item{\code{OI}}{a numeric vector}
    \item{\code{OXHP}}{a numeric vector}
    \item{\code{PCOM}}{a numeric vector}
    \item{\code{PCG}}{a numeric vector}
    \item{\code{PFFB}}{a numeric vector}
    \item{\code{PNC}}{a numeric vector}
    \item{\code{PPG}}{a numeric vector}
    \item{\code{PPL}}{a numeric vector}
    \item{\code{PRIA}}{a numeric vector}
    \item{\code{PSSI}}{a numeric vector}
    \item{\code{PCAR}}{a numeric vector}
    \item{\code{BOH}}{a numeric vector}
    \item{\code{PAG}}{a numeric vector}
    \item{\code{PSUN}}{a numeric vector}
    \item{\code{PLL}}{a numeric vector}
    \item{\code{PZZA}}{a numeric vector}
    \item{\code{PMTC}}{a numeric vector}
    \item{\code{PRXL}}{a numeric vector}
    \item{\code{PKE}}{a numeric vector}
    \item{\code{PKD}}{a numeric vector}
    \item{\code{PH}}{a numeric vector}
    \item{\code{PRKR}}{a numeric vector}
    \item{\code{PKY}}{a numeric vector}
    \item{\code{PDCO}}{a numeric vector}
    \item{\code{PTEN}}{a numeric vector}
    \item{\code{PXR}}{a numeric vector}
    \item{\code{PAYX}}{a numeric vector}
    \item{\code{PSS}}{a numeric vector}
    \item{\code{PNN}}{a numeric vector}
    \item{\code{PENN}}{a numeric vector}
    \item{\code{JCP}}{a numeric vector}
    \item{\code{PEI}}{a numeric vector}
    \item{\code{PZL}}{a numeric vector}
    \item{\code{PNR}}{a numeric vector}
    \item{\code{PBCT}}{a numeric vector}
    \item{\code{PGL}}{a numeric vector}
    \item{\code{PSFT}}{a numeric vector}
    \item{\code{PBY}}{a numeric vector}
    \item{\code{PEP}}{a numeric vector}
    \item{\code{PFGC}}{a numeric vector}
    \item{\code{PRGO}}{a numeric vector}
    \item{\code{PETM}}{a numeric vector}
    \item{\code{PFE}}{a numeric vector}
    \item{\code{PPDI}}{a numeric vector}
    \item{\code{PCOP}}{a numeric vector}
    \item{\code{PD}}{a numeric vector}
    \item{\code{PSC}}{a numeric vector}
    \item{\code{MO}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
    \item{\code{PVH}}{a numeric vector}
    \item{\code{PTEC}}{a numeric vector}
    \item{\code{PHTN}}{a numeric vector}
    \item{\code{PLAB}}{a numeric vector}
    \item{\code{PNY}}{a numeric vector}
    \item{\code{PIR}}{a numeric vector}
    \item{\code{CHX}}{a numeric vector}
    \item{\code{PCLE}}{a numeric vector}
    \item{\code{PNW}}{a numeric vector}
    \item{\code{PIOS}}{a numeric vector}
    \item{\code{PBI}}{a numeric vector}
    \item{\code{PZB}}{a numeric vector}
    \item{\code{PIXR}}{a numeric vector}
    \item{\code{PLX}}{a numeric vector}
    \item{\code{PLA}}{a numeric vector}
    \item{\code{PYX}}{a numeric vector}
    \item{\code{PLXS}}{a numeric vector}
    \item{\code{PCL}}{a numeric vector}
    \item{\code{PPP}}{a numeric vector}
    \item{\code{PII}}{a numeric vector}
    \item{\code{PRD}}{a numeric vector}
    \item{\code{PLCM}}{a numeric vector}
    \item{\code{POP}}{a numeric vector}
    \item{\code{BPOP}}{a numeric vector}
    \item{\code{PPS}}{a numeric vector}
    \item{\code{PCH}}{a numeric vector}
    \item{\code{POM}}{a numeric vector}
    \item{\code{PWAV}}{a numeric vector}
    \item{\code{PX}}{a numeric vector}
    \item{\code{PPD}}{a numeric vector}
    \item{\code{PCP}}{a numeric vector}
    \item{\code{PP}}{a numeric vector}
    \item{\code{PLFE}}{a numeric vector}
    \item{\code{PRST}}{a numeric vector}
    \item{\code{TROW}}{a numeric vector}
    \item{\code{PDQ}}{a numeric vector}
    \item{\code{PRTL}}{a numeric vector}
    \item{\code{PRCM}}{a numeric vector}
    \item{\code{PG}}{a numeric vector}
    \item{\code{PRGX}}{a numeric vector}
    \item{\code{PRGS}}{a numeric vector}
    \item{\code{PGR}}{a numeric vector}
    \item{\code{PSDI}}{a numeric vector}
    \item{\code{POSO}}{a numeric vector}
    \item{\code{PL}}{a numeric vector}
    \item{\code{PDLI}}{a numeric vector}
    \item{\code{PBKS}}{a numeric vector}
    \item{\code{PROX}}{a numeric vector}
    \item{\code{PSIX}}{a numeric vector}
    \item{\code{PNM}}{a numeric vector}
    \item{\code{PEG}}{a numeric vector}
    \item{\code{PSA}}{a numeric vector}
    \item{\code{PSD}}{a numeric vector}
    \item{\code{PTZ}}{a numeric vector}
    \item{\code{PHM}}{a numeric vector}
    \item{\code{PUMA}}{a numeric vector}
    \item{\code{QLGC}}{a numeric vector}
    \item{\code{OAT}}{a numeric vector}
    \item{\code{QCOM}}{a numeric vector}
    \item{\code{NX}}{a numeric vector}
    \item{\code{STR}}{a numeric vector}
    \item{\code{QTRN}}{a numeric vector}
    \item{\code{QHGI}}{a numeric vector}
    \item{\code{FLC}}{a numeric vector}
    \item{\code{RFS}}{a numeric vector}
    \item{\code{RGS}}{a numeric vector}
    \item{\code{RTI}}{a numeric vector}
    \item{\code{RSYS}}{a numeric vector}
    \item{\code{RNBO}}{a numeric vector}
    \item{\code{RAL}}{a numeric vector}
    \item{\code{RATL}}{a numeric vector}
    \item{\code{RJF}}{a numeric vector}
    \item{\code{RYN}}{a numeric vector}
    \item{\code{RDA}}{a numeric vector}
    \item{\code{O}}{a numeric vector}
    \item{\code{RA}}{a numeric vector}
    \item{\code{RBK}}{a numeric vector}
    \item{\code{RBC}}{a numeric vector}
    \item{\code{REGN}}{a numeric vector}
    \item{\code{RGIS}}{a numeric vector}
    \item{\code{REG}}{a numeric vector}
    \item{\code{RGBK}}{a numeric vector}
    \item{\code{RGA}}{a numeric vector}
    \item{\code{RS}}{a numeric vector}
    \item{\code{REI}}{a numeric vector}
    \item{\code{REMC}}{a numeric vector}
    \item{\code{RMDY}}{a numeric vector}
    \item{\code{RCGI}}{a numeric vector}
    \item{\code{RCII}}{a numeric vector}
    \item{\code{RBNC}}{a numeric vector}
    \item{\code{RSFC}}{a numeric vector}
    \item{\code{REFR}}{a numeric vector}
    \item{\code{RSCR}}{a numeric vector}
    \item{\code{RESP}}{a numeric vector}
    \item{\code{REV}}{a numeric vector}
    \item{\code{REY}}{a numeric vector}
    \item{\code{RIGS}}{a numeric vector}
    \item{\code{RAD}}{a numeric vector}
    \item{\code{ROAD}}{a numeric vector}
    \item{\code{RHI}}{a numeric vector}
    \item{\code{ROBV}}{a numeric vector}
    \item{\code{RKT}}{a numeric vector}
    \item{\code{ROK}}{a numeric vector}
    \item{\code{ROG}}{a numeric vector}
    \item{\code{ROH}}{a numeric vector}
    \item{\code{ROL}}{a numeric vector}
    \item{\code{RLC}}{a numeric vector}
    \item{\code{ROP}}{a numeric vector}
    \item{\code{ROST}}{a numeric vector}
    \item{\code{RSE}}{a numeric vector}
    \item{\code{RDC}}{a numeric vector}
    \item{\code{RI}}{a numeric vector}
    \item{\code{RDK}}{a numeric vector}
    \item{\code{RUS}}{a numeric vector}
    \item{\code{RML}}{a numeric vector}
    \item{\code{RYAN}}{a numeric vector}
    \item{\code{R}}{a numeric vector}
    \item{\code{RYL}}{a numeric vector}
    \item{\code{SBC}}{a numeric vector}
    \item{\code{SBSE}}{a numeric vector}
    \item{\code{SDLI}}{a numeric vector}
    \item{\code{SEIC}}{a numeric vector}
    \item{\code{SPSS}}{a numeric vector}
    \item{\code{SPW}}{a numeric vector}
    \item{\code{TSG}}{a numeric vector}
    \item{\code{SAFC}}{a numeric vector}
    \item{\code{SFE}}{a numeric vector}
    \item{\code{SWY}}{a numeric vector}
    \item{\code{JOE}}{a numeric vector}
    \item{\code{STJ}}{a numeric vector}
    \item{\code{MARY}}{a numeric vector}
    \item{\code{SPC}}{a numeric vector}
    \item{\code{SKS}}{a numeric vector}
    \item{\code{SCAI}}{a numeric vector}
    \item{\code{SNDK}}{a numeric vector}
    \item{\code{SANM}}{a numeric vector}
    \item{\code{SANG}}{a numeric vector}
    \item{\code{SAPE}}{a numeric vector}
    \item{\code{SLE}}{a numeric vector}
    \item{\code{SATC}}{a numeric vector}
    \item{\code{BFS}}{a numeric vector}
    \item{\code{SAWS}}{a numeric vector}
    \item{\code{SCG}}{a numeric vector}
    \item{\code{HSIC}}{a numeric vector}
    \item{\code{SGP}}{a numeric vector}
    \item{\code{SCHL}}{a numeric vector}
    \item{\code{SHLM}}{a numeric vector}
    \item{\code{SCH}}{a numeric vector}
    \item{\code{SWM}}{a numeric vector}
    \item{\code{SCLN}}{a numeric vector}
    \item{\code{SFA}}{a numeric vector}
    \item{\code{SMG}}{a numeric vector}
    \item{\code{SSP}}{a numeric vector}
    \item{\code{SEAC}}{a numeric vector}
    \item{\code{CKH}}{a numeric vector}
    \item{\code{S}}{a numeric vector}
    \item{\code{SCUR}}{a numeric vector}
    \item{\code{SEI}}{a numeric vector}
    \item{\code{SIGI}}{a numeric vector}
    \item{\code{SMTC}}{a numeric vector}
    \item{\code{SMTL}}{a numeric vector}
    \item{\code{SRM}}{a numeric vector}
    \item{\code{SEPR}}{a numeric vector}
    \item{\code{SQAA}}{a numeric vector}
    \item{\code{SRV}}{a numeric vector}
    \item{\code{SVM}}{a numeric vector}
    \item{\code{SGR}}{a numeric vector}
    \item{\code{SHX}}{a numeric vector}
    \item{\code{SHW}}{a numeric vector}
    \item{\code{SKO}}{a numeric vector}
    \item{\code{SHU}}{a numeric vector}
    \item{\code{SEBL}}{a numeric vector}
    \item{\code{SRP}}{a numeric vector}
    \item{\code{SIAL}}{a numeric vector}
    \item{\code{SGI}}{a numeric vector}
    \item{\code{SSTI}}{a numeric vector}
    \item{\code{SIVB}}{a numeric vector}
    \item{\code{SVGI}}{a numeric vector}
    \item{\code{SPG}}{a numeric vector}
    \item{\code{SBGI}}{a numeric vector}
    \item{\code{SWW}}{a numeric vector}
    \item{\code{SIPX}}{a numeric vector}
    \item{\code{SKY}}{a numeric vector}
    \item{\code{SKYW}}{a numeric vector}
    \item{\code{SMF}}{a numeric vector}
    \item{\code{AOS}}{a numeric vector}
    \item{\code{SII}}{a numeric vector}
    \item{\code{SRW}}{a numeric vector}
    \item{\code{SNA}}{a numeric vector}
    \item{\code{SLR}}{a numeric vector}
    \item{\code{SONC}}{a numeric vector}
    \item{\code{SON}}{a numeric vector}
    \item{\code{BID}}{a numeric vector}
    \item{\code{TSFG}}{a numeric vector}
    \item{\code{SJI}}{a numeric vector}
    \item{\code{SO}}{a numeric vector}
    \item{\code{PCU}}{a numeric vector}
    \item{\code{SUG}}{a numeric vector}
    \item{\code{SOTR}}{a numeric vector}
    \item{\code{LUV}}{a numeric vector}
    \item{\code{SWX}}{a numeric vector}
    \item{\code{SWN}}{a numeric vector}
    \item{\code{SVRN}}{a numeric vector}
    \item{\code{SSS}}{a numeric vector}
    \item{\code{SEH}}{a numeric vector}
    \item{\code{SLNK}}{a numeric vector}
    \item{\code{SFAM}}{a numeric vector}
    \item{\code{TRK}}{a numeric vector}
    \item{\code{SPGLA}}{a numeric vector}
    \item{\code{SPK}}{a numeric vector}
    \item{\code{SMI}}{a numeric vector}
    \item{\code{FON}}{a numeric vector}
    \item{\code{STAA}}{a numeric vector}
    \item{\code{SMSC}}{a numeric vector}
    \item{\code{SPF}}{a numeric vector}
    \item{\code{SR}}{a numeric vector}
    \item{\code{SXI}}{a numeric vector}
    \item{\code{SWK}}{a numeric vector}
    \item{\code{SBAS}}{a numeric vector}
    \item{\code{SPLS}}{a numeric vector}
    \item{\code{SBUX}}{a numeric vector}
    \item{\code{STFC}}{a numeric vector}
    \item{\code{HOT}}{a numeric vector}
    \item{\code{STT}}{a numeric vector}
    \item{\code{STN}}{a numeric vector}
    \item{\code{STLD}}{a numeric vector}
    \item{\code{SMRT}}{a numeric vector}
    \item{\code{SBIB}}{a numeric vector}
    \item{\code{SSSS}}{a numeric vector}
    \item{\code{STEI}}{a numeric vector}
    \item{\code{SGY}}{a numeric vector}
    \item{\code{SUS}}{a numeric vector}
    \item{\code{STK}}{a numeric vector}
    \item{\code{STRA}}{a numeric vector}
    \item{\code{SRR}}{a numeric vector}
    \item{\code{SDRC}}{a numeric vector}
    \item{\code{RGR}}{a numeric vector}
    \item{\code{SUB}}{a numeric vector}
    \item{\code{SMT}}{a numeric vector}
    \item{\code{SUI}}{a numeric vector}
    \item{\code{SUNW}}{a numeric vector}
    \item{\code{SOC}}{a numeric vector}
    \item{\code{RAYS}}{a numeric vector}
    \item{\code{SUN}}{a numeric vector}
    \item{\code{SNRZ}}{a numeric vector}
    \item{\code{STI}}{a numeric vector}
    \item{\code{SCON}}{a numeric vector}
    \item{\code{SUPG}}{a numeric vector}
    \item{\code{SESI}}{a numeric vector}
    \item{\code{SUP}}{a numeric vector}
    \item{\code{SUPX}}{a numeric vector}
    \item{\code{SVU}}{a numeric vector}
    \item{\code{SUSQ}}{a numeric vector}
    \item{\code{SFY}}{a numeric vector}
    \item{\code{SWFT}}{a numeric vector}
    \item{\code{SYBS}}{a numeric vector}
    \item{\code{SYKE}}{a numeric vector}
    \item{\code{SLVN}}{a numeric vector}
    \item{\code{SYMC}}{a numeric vector}
    \item{\code{SBL}}{a numeric vector}
    \item{\code{SYMM}}{a numeric vector}
    \item{\code{SCOR}}{a numeric vector}
    \item{\code{SNPS}}{a numeric vector}
    \item{\code{SNV}}{a numeric vector}
    \item{\code{SYY}}{a numeric vector}
    \item{\code{SCTC}}{a numeric vector}
    \item{\code{TCB}}{a numeric vector}
    \item{\code{TE}}{a numeric vector}
    \item{\code{TJX}}{a numeric vector}
    \item{\code{TRW}}{a numeric vector}
    \item{\code{TXU}}{a numeric vector}
    \item{\code{TLB}}{a numeric vector}
    \item{\code{TALK}}{a numeric vector}
    \item{\code{SKT}}{a numeric vector}
    \item{\code{TGEN}}{a numeric vector}
    \item{\code{TCO}}{a numeric vector}
    \item{\code{TECD}}{a numeric vector}
    \item{\code{TECH}}{a numeric vector}
    \item{\code{TNL}}{a numeric vector}
    \item{\code{TSCC}}{a numeric vector}
    \item{\code{TECUA}}{a numeric vector}
    \item{\code{TRC}}{a numeric vector}
    \item{\code{TKLC}}{a numeric vector}
    \item{\code{TEK}}{a numeric vector}
    \item{\code{TLCM}}{a numeric vector}
    \item{\code{TFX}}{a numeric vector}
    \item{\code{TDS}}{a numeric vector}
    \item{\code{TLAB}}{a numeric vector}
    \item{\code{TIN}}{a numeric vector}
    \item{\code{TTEC}}{a numeric vector}
    \item{\code{THC}}{a numeric vector}
    \item{\code{TEN}}{a numeric vector}
    \item{\code{TER}}{a numeric vector}
    \item{\code{TSO}}{a numeric vector}
    \item{\code{TX}}{a numeric vector}
    \item{\code{TXB}}{a numeric vector}
    \item{\code{TXI}}{a numeric vector}
    \item{\code{TXN}}{a numeric vector}
    \item{\code{TXT}}{a numeric vector}
    \item{\code{TGX}}{a numeric vector}
    \item{\code{TCA}}{a numeric vector}
    \item{\code{TFT}}{a numeric vector}
    \item{\code{TMO}}{a numeric vector}
    \item{\code{TNB}}{a numeric vector}
    \item{\code{TII}}{a numeric vector}
    \item{\code{THO}}{a numeric vector}
    \item{\code{THDO}}{a numeric vector}
    \item{\code{COMS}}{a numeric vector}
    \item{\code{TFS}}{a numeric vector}
    \item{\code{TDW}}{a numeric vector}
    \item{\code{TIF}}{a numeric vector}
    \item{\code{TBL}}{a numeric vector}
    \item{\code{TWX}}{a numeric vector}
    \item{\code{TKR}}{a numeric vector}
    \item{\code{TTN}}{a numeric vector}
    \item{\code{TOL}}{a numeric vector}
    \item{\code{TLGD}}{a numeric vector}
    \item{\code{TR}}{a numeric vector}
    \item{\code{TOPP}}{a numeric vector}
    \item{\code{TMK}}{a numeric vector}
    \item{\code{TTC}}{a numeric vector}
    \item{\code{TOS}}{a numeric vector}
    \item{\code{TSS}}{a numeric vector}
    \item{\code{TCT}}{a numeric vector}
    \item{\code{TOY}}{a numeric vector}
    \item{\code{TWMC}}{a numeric vector}
    \item{\code{TSAI}}{a numeric vector}
    \item{\code{TRH}}{a numeric vector}
    \item{\code{TKTX}}{a numeric vector}
    \item{\code{TMG}}{a numeric vector}
    \item{\code{TXCC}}{a numeric vector}
    \item{\code{TG}}{a numeric vector}
    \item{\code{VIRS}}{a numeric vector}
    \item{\code{TGIC}}{a numeric vector}
    \item{\code{TRY}}{a numeric vector}
    \item{\code{TRB}}{a numeric vector}
    \item{\code{TMAR}}{a numeric vector}
    \item{\code{TRCD}}{a numeric vector}
    \item{\code{TRMB}}{a numeric vector}
    \item{\code{TRN}}{a numeric vector}
    \item{\code{TQNT}}{a numeric vector}
    \item{\code{TNO}}{a numeric vector}
    \item{\code{TCNJ}}{a numeric vector}
    \item{\code{TRST}}{a numeric vector}
    \item{\code{TRMK}}{a numeric vector}
    \item{\code{TUP}}{a numeric vector}
    \item{\code{TW}}{a numeric vector}
    \item{\code{TWLB}}{a numeric vector}
    \item{\code{IIVI}}{a numeric vector}
    \item{\code{UAL}}{a numeric vector}
    \item{\code{UCR}}{a numeric vector}
    \item{\code{UGI}}{a numeric vector}
    \item{\code{UIL}}{a numeric vector}
    \item{\code{UMBF}}{a numeric vector}
    \item{\code{MRO}}{a numeric vector}
    \item{\code{UST}}{a numeric vector}
    \item{\code{URS}}{a numeric vector}
    \item{\code{USG}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{UTI}}{a numeric vector}
    \item{\code{SLM}}{a numeric vector}
    \item{\code{UDS}}{a numeric vector}
    \item{\code{UTEK}}{a numeric vector}
    \item{\code{UFI}}{a numeric vector}
    \item{\code{UK}}{a numeric vector}
    \item{\code{UNP}}{a numeric vector}
    \item{\code{UPC}}{a numeric vector}
    \item{\code{UIS}}{a numeric vector}
    \item{\code{UNT}}{a numeric vector}
    \item{\code{UAG}}{a numeric vector}
    \item{\code{UBSI}}{a numeric vector}
    \item{\code{UDR}}{a numeric vector}
    \item{\code{UNFI}}{a numeric vector}
    \item{\code{USM}}{a numeric vector}
    \item{\code{U}}{a numeric vector}
    \item{\code{USI}}{a numeric vector}
    \item{\code{UTX}}{a numeric vector}
    \item{\code{UNH}}{a numeric vector}
    \item{\code{UNIT}}{a numeric vector}
    \item{\code{UVV}}{a numeric vector}
    \item{\code{UEIC}}{a numeric vector}
    \item{\code{UFPI}}{a numeric vector}
    \item{\code{UHS}}{a numeric vector}
    \item{\code{UVN}}{a numeric vector}
    \item{\code{UCL}}{a numeric vector}
    \item{\code{UNM}}{a numeric vector}
    \item{\code{USFC}}{a numeric vector}
    \item{\code{UCU}}{a numeric vector}
    \item{\code{VFC}}{a numeric vector}
    \item{\code{VCI}}{a numeric vector}
    \item{\code{VLNC}}{a numeric vector}
    \item{\code{VLY}}{a numeric vector}
    \item{\code{VALM}}{a numeric vector}
    \item{\code{VAL}}{a numeric vector}
    \item{\code{VCD}}{a numeric vector}
    \item{\code{VVTV}}{a numeric vector}
    \item{\code{VANS}}{a numeric vector}
    \item{\code{VRC}}{a numeric vector}
    \item{\code{VAR}}{a numeric vector}
    \item{\code{VASO}}{a numeric vector}
    \item{\code{VECO}}{a numeric vector}
    \item{\code{VMSI}}{a numeric vector}
    \item{\code{VRTY}}{a numeric vector}
    \item{\code{VTS}}{a numeric vector}
    \item{\code{VRTS}}{a numeric vector}
    \item{\code{VRTX}}{a numeric vector}
    \item{\code{VVI}}{a numeric vector}
    \item{\code{VSAT}}{a numeric vector}
    \item{\code{VIAB}}{a numeric vector}
    \item{\code{VICL}}{a numeric vector}
    \item{\code{VICR}}{a numeric vector}
    \item{\code{VPI}}{a numeric vector}
    \item{\code{VSH}}{a numeric vector}
    \item{\code{VITL}}{a numeric vector}
    \item{\code{VTSS}}{a numeric vector}
    \item{\code{VVUS}}{a numeric vector}
    \item{\code{VNO}}{a numeric vector}
    \item{\code{VMC}}{a numeric vector}
    \item{\code{WFSI}}{a numeric vector}
    \item{\code{WDFC}}{a numeric vector}
    \item{\code{WGL}}{a numeric vector}
    \item{\code{WBPR}}{a numeric vector}
    \item{\code{WMS}}{a numeric vector}
    \item{\code{WPS}}{a numeric vector}
    \item{\code{WNC}}{a numeric vector}
    \item{\code{WAB}}{a numeric vector}
    \item{\code{WB}}{a numeric vector}
    \item{\code{WMT}}{a numeric vector}
    \item{\code{WAG}}{a numeric vector}
    \item{\code{WCS}}{a numeric vector}
    \item{\code{WAC}}{a numeric vector}
    \item{\code{WFSL}}{a numeric vector}
    \item{\code{WPO}}{a numeric vector}
    \item{\code{WRE}}{a numeric vector}
    \item{\code{WMI}}{a numeric vector}
    \item{\code{WAT}}{a numeric vector}
    \item{\code{WSO}}{a numeric vector}
    \item{\code{WTS}}{a numeric vector}
    \item{\code{WBB}}{a numeric vector}
    \item{\code{WLNK}}{a numeric vector}
    \item{\code{WBST}}{a numeric vector}
    \item{\code{WRI}}{a numeric vector}
    \item{\code{WS}}{a numeric vector}
    \item{\code{WMK}}{a numeric vector}
    \item{\code{WLM}}{a numeric vector}
    \item{\code{WLP}}{a numeric vector}
    \item{\code{WFC}}{a numeric vector}
    \item{\code{WEN}}{a numeric vector}
    \item{\code{WERN}}{a numeric vector}
    \item{\code{WST}}{a numeric vector}
    \item{\code{WTSC}}{a numeric vector}
    \item{\code{WABC}}{a numeric vector}
    \item{\code{WSTL}}{a numeric vector}
    \item{\code{WDC}}{a numeric vector}
    \item{\code{WGR}}{a numeric vector}
    \item{\code{WR}}{a numeric vector}
    \item{\code{WWCA}}{a numeric vector}
    \item{\code{W}}{a numeric vector}
    \item{\code{WY}}{a numeric vector}
    \item{\code{WHR}}{a numeric vector}
    \item{\code{WH}}{a numeric vector}
    \item{\code{WTNY}}{a numeric vector}
    \item{\code{WFMI}}{a numeric vector}
    \item{\code{OATS}}{a numeric vector}
    \item{\code{WLL}}{a numeric vector}
    \item{\code{WMB}}{a numeric vector}
    \item{\code{CWEI}}{a numeric vector}
    \item{\code{WIND}}{a numeric vector}
    \item{\code{WIN}}{a numeric vector}
    \item{\code{WGO}}{a numeric vector}
    \item{\code{WCII}}{a numeric vector}
    \item{\code{WCLX}}{a numeric vector}
    \item{\code{WEC}}{a numeric vector}
    \item{\code{WLV}}{a numeric vector}
    \item{\code{WWW}}{a numeric vector}
    \item{\code{WGOV}}{a numeric vector}
    \item{\code{WAXS}}{a numeric vector}
    \item{\code{WCOM}}{a numeric vector}
    \item{\code{XCED}}{a numeric vector}
    \item{\code{WWY}}{a numeric vector}
    \item{\code{XRIT}}{a numeric vector}
    \item{\code{XLNX}}{a numeric vector}
    \item{\code{XIRC}}{a numeric vector}
    \item{\code{XRX}}{a numeric vector}
    \item{\code{XTR}}{a numeric vector}
    \item{\code{XYBR}}{a numeric vector}
    \item{\code{YHOO}}{a numeric vector}
    \item{\code{YELL}}{a numeric vector}
    \item{\code{YRK}}{a numeric vector}
    \item{\code{YBTVA}}{a numeric vector}
    \item{\code{ZBRA}}{a numeric vector}
    \item{\code{ZIGO}}{a numeric vector}
    \item{\code{GZNASD}}{a numeric vector}
    \item{\code{GZMSFT}}{a numeric vector}
    \item{\code{GZOIL}}{a numeric vector}
    \item{\code{GZSOXX}}{a numeric vector}
  }
}
\source{
S+ FinMetric "folio.dat".
}
\references{
S+ FinMetric "folio.dat".
}
\examples{
data(folio.dat)
## maybe str(folio.dat) ; plot(folio.dat) ...
}
\keyword{datasets}
