% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{folio}
\alias{folio}
\title{Weekly returns on 1618 stocks from January 1997 to June 2000}
\format{A \samp{xts} object on Feb 1990/Dec 2000. Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
folio
}
\description{
A \samp{xts} object of weekly returns on 1618 stocks from January 1997 to June 2000.
}
\keyword{datasets}

