% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ford}
\alias{ford}
\title{Daily returns on Ford stock from February 1984 to December 1991}
\format{A \samp{xts} object on 1984-02-02/1991-12-31 containing:
\describe{
  \item{ford}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
ford
}
\description{
A \samp{xts} object of daily returns on Ford stock from February 1984 to December 1991.
}
\keyword{datasets}

