% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frip}
\alias{frip}
\title{Monthly industrial production of France from January 1960 to December 1989}
\format{A \samp{xts} object on Jan 1960/Dec 1989 containing:
\describe{
  \item{frip}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
frip
}
\description{
A \samp{xts} object of monthly industrial production of France from January 1960 to December
1989.
}
\keyword{datasets}

