% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwdRate}
\alias{fwdRate}
\title{Monthly forward rates of U.S. coupons from August 1985 to February 1991}
\format{A \samp{xts} object on Aug 1985/Feb 1991. Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
fwdRate
}
\description{
A \samp{xts} object of monthly forward rates of U.S. coupons from August 1985 to February 1991
computed by McCulloch and Kwon (1993).
}
\keyword{datasets}

