% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german}
\alias{german}
\title{Daily log returns on BMW and Siemens stock from January 1973 through July 1996}
\format{A \samp{xts} object on 1973-01-02/1996-07-23 containing:
\describe{
  \item{bmw}{a numeric vector}
  \item{siemens}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
german
}
\description{
A \samp{xts} object of daily log returns on BMW and Siemens stock from January 1973 through July
1996.
}
\keyword{datasets}

