\name{highFreq3M.df}
\alias{highFreq3M.df}
\docType{data}
\title{
the transactions level data on 3M corporation stock for December 1999
}
\description{
the transactions level data on 3M corporation stock for December 1999
}
\usage{data(highFreq3M.df)}
\format{
  A data frame with 21368 observations on the following 3 variables.
  \describe{
    \item{\code{trade.day}}{a numeric vector}
    \item{\code{trade.time}}{a numeric vector}
    \item{\code{trade.price}}{a numeric vector}
  }
}

\source{
S+ FinMetric "highFreq3M.df"
}
\references{
S+ FinMetric "highFreq3M.df"
}
\examples{
data(highFreq3M.df)
## maybe str(highFreq3M.df) ; plot(highFreq3M.df) ...
}
\keyword{datasets}
