% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highFreq3M}
\alias{highFreq3M}
\title{Transactions level data on 3M Corporation stock for December 1999}
\format{A \samp{xts} object on 1999-12-01 09:33:32/1999-12-31 13:02:53 containing:
\describe{
  \item{3M}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
highFreq3M
}
\description{
A \samp{xts} object of transactions level data on 3M Corporation stock for December 1999.
}
\keyword{datasets}

