\name{hp.s}
\alias{hp.s}
\docType{data}
\title{
daily time series of Ford stock returns
}
\description{
daily time series of Ford stock returns from 1/10/1986 to 9/9/1999
}
\usage{data(hp.s)}
\format{
  A data frame with 2000 observations on the following 2 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{1/10/1985} \code{1/10/1986} \code{1/10/1989} \code{1/10/1990} \code{1/10/1991} \code{1/11/1985} \code{1/11/1988} \code{1/11/1989} \code{1/11/1990} \code{1/11/1991} \code{1/12/1987} \code{1/12/1988} \code{1/12/1989} \code{1/12/1990} \code{1/13/1986} \code{1/13/1987} \code{1/13/1988} \code{1/13/1989} \code{1/14/1985} \code{1/14/1986} \code{1/14/1987} \code{1/14/1988} \code{1/14/1991} \code{1/15/1985} \code{1/15/1986} \code{1/15/1987} \code{1/15/1988} \code{1/15/1990} \code{1/15/1991} \code{1/16/1985} \code{1/16/1986} \code{1/16/1987} \code{1/16/1989} \code{1/16/1990} \code{1/16/1991} \code{1/17/1985} \code{1/17/1986} \code{1/17/1989} \code{1/17/1990} \code{1/17/1991} \code{1/18/1985} \code{1/18/1988} \code{1/18/1989} \code{1/18/1990} \code{1/18/1991} \code{1/19/1987} \code{1/19/1988} \code{1/19/1989} \code{1/19/1990} \code{1/2/1985} \code{1/2/1986} \code{1/2/1987} \code{1/2/1990} \code{1/2/1991} \code{1/20/1986} \code{1/20/1987} \code{1/20/1988} \code{1/20/1989} \code{1/21/1985} \code{1/21/1986} \code{1/21/1987} \code{1/21/1988} \code{1/21/1991} \code{1/22/1985} \code{1/22/1986} \code{1/22/1987} \code{1/22/1988} \code{1/22/1990} \code{1/22/1991} \code{1/23/1985} \code{1/23/1986} \code{1/23/1987} \code{1/23/1989} \code{1/23/1990} \code{1/23/1991} \code{1/24/1985} \code{1/24/1986} \code{1/24/1989} \code{1/24/1990} \code{1/24/1991} \code{1/25/1985} \code{1/25/1988} \code{1/25/1989} \code{1/25/1990} \code{1/25/1991} \code{1/26/1987} \code{1/26/1988} \code{1/26/1989} \code{1/26/1990} \code{1/27/1986} \code{1/27/1987} \code{1/27/1988} \code{1/27/1989} \code{1/28/1985} \code{1/28/1986} \code{1/28/1987} \code{1/28/1988} \code{1/28/1991} \code{1/29/1985} \code{1/29/1986} \code{1/29/1987} \code{1/29/1988} \code{1/29/1990} \code{1/29/1991} \code{1/3/1985} \code{1/3/1986} \code{1/3/1989} \code{1/3/1990} \code{1/3/1991} \code{1/30/1985} \code{1/30/1986} \code{1/30/1987} \code{1/30/1989} \code{1/30/1990} \code{1/30/1991} \code{1/31/1985} \code{1/31/1986} \code{1/31/1989} \code{1/31/1990} \code{1/31/1991} \code{1/4/1985} \code{1/4/1988} \code{1/4/1989} \code{1/4/1990} \code{1/4/1991} \code{1/5/1987} \code{1/5/1988} \code{1/5/1989} \code{1/5/1990} \code{1/6/1986} \code{1/6/1987} \code{1/6/1988} \code{1/6/1989} \code{1/7/1985} \code{1/7/1986} \code{1/7/1987} \code{1/7/1988} \code{1/7/1991} \code{1/8/1985} \code{1/8/1986} \code{1/8/1987} \code{1/8/1988} \code{1/8/1990} \code{1/8/1991} \code{1/9/1985} \code{1/9/1986} \code{1/9/1987} \code{1/9/1989} \code{1/9/1990} \code{1/9/1991} \code{10/1/1984} \code{10/1/1985} \code{10/1/1986} \code{10/1/1987} \code{10/1/1990} \code{10/1/1991} \code{10/10/1984} \code{10/10/1985} \code{10/10/1986} \code{10/10/1988} \code{10/10/1989} \code{10/10/1990} \code{10/10/1991} \code{10/11/1984} \code{10/11/1985} \code{10/11/1988} \code{10/11/1989} \code{10/11/1990} \code{10/11/1991} \code{10/12/1984} \code{10/12/1987} \code{10/12/1988} \code{10/12/1989} \code{10/12/1990} \code{10/13/1986} \code{10/13/1987} \code{10/13/1988} \code{10/13/1989} \code{10/14/1985} \code{10/14/1986} \code{10/14/1987} \code{10/14/1988} \code{10/14/1991} \code{10/15/1984} \code{10/15/1985} \code{10/15/1986} \code{10/15/1987} \code{10/15/1990} \code{10/15/1991} \code{10/16/1984} \code{10/16/1985} \code{10/16/1986} \code{10/16/1987} \code{10/16/1989} \code{10/16/1990} \code{10/16/1991} \code{10/17/1984} \code{10/17/1985} \code{10/17/1986} \code{10/17/1988} \code{10/17/1989} \code{10/17/1990} \code{10/17/1991} \code{10/18/1984} \code{10/18/1985} \code{10/18/1988} \code{10/18/1989} \code{10/18/1990} \code{10/18/1991} \code{10/19/1984} \code{10/19/1987} \code{10/19/1988} \code{10/19/1989} \code{10/19/1990} \code{10/2/1984} \code{10/2/1985} \code{10/2/1986} \code{10/2/1987} \code{10/2/1989} \code{10/2/1990} \code{10/2/1991} \code{10/20/1986} \code{10/20/1987} \code{10/20/1988} \code{10/20/1989} \code{10/21/1985} \code{10/21/1986} \code{10/21/1987} \code{10/21/1988} \code{10/21/1991} \code{10/22/1984} \code{10/22/1985} \code{10/22/1986} \code{10/22/1987} \code{10/22/1990} \code{10/22/1991} \code{10/23/1984} \code{10/23/1985} \code{10/23/1986} \code{10/23/1987} \code{10/23/1989} \code{10/23/1990} \code{10/23/1991} \code{10/24/1984} \code{10/24/1985} \code{10/24/1986} \code{10/24/1988} \code{10/24/1989} \code{10/24/1990} \code{10/24/1991} \code{10/25/1984} \code{10/25/1985} \code{10/25/1988} \code{10/25/1989} \code{10/25/1990} \code{10/25/1991} \code{10/26/1984} \code{10/26/1987} \code{10/26/1988} \code{10/26/1989} \code{10/26/1990} \code{10/27/1986} \code{10/27/1987} \code{10/27/1988} \code{10/27/1989} \code{10/28/1985} \code{10/28/1986} \code{10/28/1987} \code{10/28/1988} \code{10/28/1991} \code{10/29/1984} \code{10/29/1985} \code{10/29/1986} \code{10/29/1987} \code{10/29/1990} \code{10/29/1991} \code{10/3/1984} \code{10/3/1985} \code{10/3/1986} \code{10/3/1988} \code{10/3/1989} \code{10/3/1990} \code{10/3/1991} \code{10/30/1984} \code{10/30/1985} \code{10/30/1986} \code{10/30/1987} \code{10/30/1989} \code{10/30/1990} \code{10/30/1991} \code{10/31/1984} \code{10/31/1985} \code{10/31/1986} \code{10/31/1988} \code{10/31/1989} \code{10/31/1990} \code{10/31/1991} \code{10/4/1984} \code{10/4/1985} \code{10/4/1988} \code{10/4/1989} \code{10/4/1990} \code{10/4/1991} \code{10/5/1984} \code{10/5/1987} \code{10/5/1988} \code{10/5/1989} \code{10/5/1990} \code{10/6/1986} \code{10/6/1987} \code{10/6/1988} \code{10/6/1989} \code{10/7/1985} \code{10/7/1986} \code{10/7/1987} \code{10/7/1988} \code{10/7/1991} \code{10/8/1984} \code{10/8/1985} \code{10/8/1986} \code{10/8/1987} \code{10/8/1990} \code{10/8/1991} \code{10/9/1984} \code{10/9/1985} \code{10/9/1986} \code{10/9/1987} \code{10/9/1989} \code{10/9/1990} \code{10/9/1991} \code{11/1/1984} \code{11/1/1985} \code{11/1/1988} \code{11/1/1989} \code{11/1/1990} \code{11/1/1991} \code{11/10/1986} \code{11/10/1987} \code{11/10/1988} \code{11/10/1989} \code{11/11/1985} \code{11/11/1986} \code{11/11/1987} \code{11/11/1988} \code{11/11/1991} \code{11/12/1984} \code{11/12/1985} \code{11/12/1986} \code{11/12/1987} \code{11/12/1990} \code{11/12/1991} \code{11/13/1984} \code{11/13/1985} \code{11/13/1986} \code{11/13/1987} \code{11/13/1989} \code{11/13/1990} \code{11/13/1991} \code{11/14/1984} \code{11/14/1985} \code{11/14/1986} \code{11/14/1988} \code{11/14/1989} \code{11/14/1990} \code{11/14/1991} \code{11/15/1984} \code{11/15/1985} \code{11/15/1988} \code{11/15/1989} \code{11/15/1990} \code{11/15/1991} \code{11/16/1984} \code{11/16/1987} \code{11/16/1988} \code{11/16/1989} \code{11/16/1990} \code{11/17/1986} \code{11/17/1987} \code{11/17/1988} \code{11/17/1989} \code{11/18/1985} \code{11/18/1986} \code{11/18/1987} \code{11/18/1988} \code{11/18/1991} \code{11/19/1984} \code{11/19/1985} \code{11/19/1986} \code{11/19/1987} \code{11/19/1990} \code{11/19/1991} \code{11/2/1984} \code{11/2/1987} \code{11/2/1988} \code{11/2/1989} \code{11/2/1990} \code{11/20/1984} \code{11/20/1985} \code{11/20/1986} \code{11/20/1987} \code{11/20/1989} \code{11/20/1990} \code{11/20/1991} \code{11/21/1984} \code{11/21/1985} \code{11/21/1986} \code{11/21/1988} \code{11/21/1989} \code{11/21/1990} \code{11/21/1991} \code{11/22/1985} \code{11/22/1988} \code{11/22/1989} \code{11/22/1991} \code{11/23/1984} \code{11/23/1987} \code{11/23/1988} \code{11/23/1990} \code{11/24/1986} \code{11/24/1987} \code{11/24/1989} \code{11/25/1985} \code{11/25/1986} \code{11/25/1987} \code{11/25/1988} \code{11/25/1991} \code{11/26/1984} \code{11/26/1985} \code{11/26/1986} \code{11/26/1990} \code{11/26/1991} \code{11/27/1984} \code{11/27/1985} \code{11/27/1987} \code{11/27/1989} \code{11/27/1990} \code{11/27/1991} \code{11/28/1984} \code{11/28/1986} \code{11/28/1988} \code{11/28/1989} \code{11/28/1990} \code{11/29/1984} \code{11/29/1985} \code{11/29/1988} \code{11/29/1989} \code{11/29/1990} \code{11/29/1991} \code{11/3/1986} \code{11/3/1987} \code{11/3/1988} \code{11/3/1989} \code{11/30/1984} \code{11/30/1987} \code{11/30/1988} \code{11/30/1989} \code{11/30/1990} \code{11/4/1985} \code{11/4/1986} \code{11/4/1987} \code{11/4/1988} \code{11/4/1991} \code{11/5/1984} \code{11/5/1985} \code{11/5/1986} \code{11/5/1987} \code{11/5/1990} \code{11/5/1991} \code{11/6/1984} \code{11/6/1985} \code{11/6/1986} \code{11/6/1987} \code{11/6/1989} \code{11/6/1990} \code{11/6/1991} \code{11/7/1984} \code{11/7/1985} \code{11/7/1986} \code{11/7/1988} \code{11/7/1989} \code{11/7/1990} \code{11/7/1991} \code{11/8/1984} \code{11/8/1985} \code{11/8/1988} \code{11/8/1989} \code{11/8/1990} \code{11/8/1991} \code{11/9/1984} \code{11/9/1987} \code{11/9/1988} \code{11/9/1989} \code{11/9/1990} \code{12/1/1986} \code{12/1/1987} \code{12/1/1988} \code{12/1/1989} \code{12/10/1984} \code{12/10/1985} \code{12/10/1986} \code{12/10/1987} \code{12/10/1990} \code{12/10/1991} \code{12/11/1984} \code{12/11/1985} \code{12/11/1986} \code{12/11/1987} \code{12/11/1989} \code{12/11/1990} \code{12/11/1991} \code{12/12/1984} \code{12/12/1985} \code{12/12/1986} \code{12/12/1988} \code{12/12/1989} \code{12/12/1990} \code{12/12/1991} \code{12/13/1984} \code{12/13/1985} \code{12/13/1988} \code{12/13/1989} \code{12/13/1990} \code{12/13/1991} \code{12/14/1984} \code{12/14/1987} \code{12/14/1988} \code{12/14/1989} \code{12/14/1990} \code{12/15/1986} \code{12/15/1987} \code{12/15/1988} \code{12/15/1989} \code{12/16/1985} \code{12/16/1986} \code{12/16/1987} \code{12/16/1988} \code{12/16/1991} \code{12/17/1984} \code{12/17/1985} \code{12/17/1986} \code{12/17/1987} \code{12/17/1990} \code{12/17/1991} \code{12/18/1984} \code{12/18/1985} \code{12/18/1986} \code{12/18/1987} \code{12/18/1989} \code{12/18/1990} \code{12/18/1991} \code{12/19/1984} \code{12/19/1985} \code{12/19/1986} \code{12/19/1988} \code{12/19/1989} \code{12/19/1990} \code{12/19/1991} \code{12/2/1985} \code{12/2/1986} \code{12/2/1987} \code{12/2/1988} \code{12/2/1991} \code{12/20/1984} \code{12/20/1985} \code{12/20/1988} \code{12/20/1989} \code{12/20/1990} \code{12/20/1991} \code{12/21/1984} \code{12/21/1987} \code{12/21/1988} \code{12/21/1989} \code{12/21/1990} \code{12/22/1986} \code{12/22/1987} \code{12/22/1988} \code{12/22/1989} \code{12/23/1985} \code{12/23/1986} \code{12/23/1987} \code{12/23/1988} \code{12/23/1991} \code{12/24/1984} \code{12/24/1985} \code{12/24/1986} \code{12/24/1987} \code{12/24/1990} \code{12/24/1991} \code{12/26/1984} \code{12/26/1985} \code{12/26/1986} \code{12/26/1989} \code{12/26/1990} \code{12/26/1991} \code{12/27/1984} \code{12/27/1985} \code{12/27/1988} \code{12/27/1989} \code{12/27/1990} \code{12/27/1991} \code{12/28/1984} \code{12/28/1987} \code{12/28/1988} \code{12/28/1989} \code{12/28/1990} \code{12/29/1986} \code{12/29/1987} \code{12/29/1988} \code{12/29/1989} \code{12/3/1984} \code{12/3/1985} \code{12/3/1986} \code{12/3/1987} \code{12/3/1990} \code{12/3/1991} \code{12/30/1985} \code{12/30/1986} \code{12/30/1987} \code{12/30/1988} \code{12/30/1991} \code{12/31/1984} \code{12/31/1985} \code{12/31/1986} \code{12/31/1987} \code{12/31/1990} \code{12/31/1991} \code{12/4/1984} \code{12/4/1985} \code{12/4/1986} \code{12/4/1987} \code{12/4/1989} \code{12/4/1990} \code{12/4/1991} \code{12/5/1984} \code{12/5/1985} \code{12/5/1986} \code{12/5/1988} \code{12/5/1989} \code{12/5/1990} \code{12/5/1991} \code{12/6/1984} \code{12/6/1985} \code{12/6/1988} \code{12/6/1989} \code{12/6/1990} \code{12/6/1991} \code{12/7/1984} \code{12/7/1987} \code{12/7/1988} \code{12/7/1989} \code{12/7/1990} \code{12/8/1986} \code{12/8/1987} \code{12/8/1988} \code{12/8/1989} \code{12/9/1985} \code{12/9/1986} \code{12/9/1987} \code{12/9/1988} \code{12/9/1991} \code{2/1/1985} \code{2/1/1988} \code{2/1/1989} \code{2/1/1990} \code{2/1/1991} \code{2/10/1984} \code{2/10/1986} \code{2/10/1987} \code{2/10/1988} \code{2/10/1989} \code{2/11/1985} \code{2/11/1986} \code{2/11/1987} \code{2/11/1988} \code{2/11/1991} \code{2/12/1985} \code{2/12/1986} \code{2/12/1987} \code{2/12/1988} \code{2/12/1990} \code{2/12/1991} \code{2/13/1984} \code{2/13/1985} \code{2/13/1986} \code{2/13/1987} \code{2/13/1989} \code{2/13/1990} \code{2/13/1991} \code{2/14/1984} \code{2/14/1985} \code{2/14/1986} \code{2/14/1989} \code{2/14/1990} \code{2/14/1991} \code{2/15/1984} \code{2/15/1985} \code{2/15/1989} \code{2/15/1990} \code{2/15/1991} \code{2/16/1984} \code{2/16/1988} \code{2/16/1989} \code{2/16/1990} \code{2/17/1984} \code{2/17/1987} \code{2/17/1988} \code{2/17/1989} \code{2/18/1986} \code{2/18/1987} \code{2/18/1988} \code{2/19/1985} \code{2/19/1986} \code{2/19/1987} \code{2/19/1988} \code{2/19/1991} \code{2/2/1984} \code{2/2/1987} \code{2/2/1988} \code{2/2/1989} \code{2/2/1990} \code{2/20/1985} \code{2/20/1986} \code{2/20/1987} \code{2/20/1990} \code{2/20/1991} \code{2/21/1984} \code{2/21/1985} \code{2/21/1986} \code{2/21/1989} \code{2/21/1990} \code{2/21/1991} \code{2/22/1984} \code{2/22/1985} \code{2/22/1988} \code{2/22/1989} \code{2/22/1990} \code{2/22/1991} \code{2/23/1984} \code{2/23/1987} \code{2/23/1988} \code{2/23/1989} \code{2/23/1990} \code{2/24/1984} \code{2/24/1986} \code{2/24/1987} \code{2/24/1988} \code{2/24/1989} \code{2/25/1985} \code{2/25/1986} \code{2/25/1987} \code{2/25/1988} \code{2/25/1991} \code{2/26/1985} \code{2/26/1986} \code{2/26/1987} \code{2/26/1988} \code{2/26/1990} \code{2/26/1991} \code{2/27/1984} \code{2/27/1985} \code{2/27/1986} \code{2/27/1987} \code{2/27/1989} \code{2/27/1990} \code{2/27/1991} \code{2/28/1984} \code{2/28/1985} \code{2/28/1986} \code{2/28/1989} \code{2/28/1990} \code{2/28/1991} \code{2/29/1984} \code{2/29/1988} \code{2/3/1984} \code{2/3/1986} \code{2/3/1987} \code{2/3/1988} \code{2/3/1989} \code{2/4/1985} \code{2/4/1986} \code{2/4/1987} \code{2/4/1988} \code{2/4/1991} \code{2/5/1985} \code{2/5/1986} \code{2/5/1987} \code{2/5/1988} \code{2/5/1990} \code{2/5/1991} \code{2/6/1984} \code{2/6/1985} \code{2/6/1986} \code{2/6/1987} \code{2/6/1989} \code{2/6/1990} \code{2/6/1991} \code{2/7/1984} \code{2/7/1985} \code{2/7/1986} \code{2/7/1989} \code{2/7/1990} \code{2/7/1991} \code{2/8/1984} \code{2/8/1985} \code{2/8/1988} \code{2/8/1989} \code{2/8/1990} \code{2/8/1991} \code{2/9/1984} \code{2/9/1987} \code{2/9/1988} \code{2/9/1989} \code{2/9/1990} \code{3/1/1984} \code{3/1/1985} \code{3/1/1988} \code{3/1/1989} \code{3/1/1990} \code{3/1/1991} \code{3/10/1986} \code{3/10/1987} \code{3/10/1988} \code{3/10/1989} \code{3/11/1985} \code{3/11/1986} \code{3/11/1987} \code{3/11/1988} \code{3/11/1991} \code{3/12/1984} \code{3/12/1985} \code{3/12/1986} \code{3/12/1987} \code{3/12/1990} \code{3/12/1991} \code{3/13/1984} \code{3/13/1985} \code{3/13/1986} \code{3/13/1987} \code{3/13/1989} \code{3/13/1990} \code{3/13/1991} \code{3/14/1984} \code{3/14/1985} \code{3/14/1986} \code{3/14/1988} \code{3/14/1989} \code{3/14/1990} \code{3/14/1991} \code{3/15/1984} \code{3/15/1985} \code{3/15/1988} \code{3/15/1989} \code{3/15/1990} \code{3/15/1991} \code{3/16/1984} \code{3/16/1987} \code{3/16/1988} \code{3/16/1989} \code{3/16/1990} \code{3/17/1986} \code{3/17/1987} \code{3/17/1988} \code{3/17/1989} \code{3/18/1985} \code{3/18/1986} \code{3/18/1987} \code{3/18/1988} \code{3/18/1991} \code{3/19/1984} \code{3/19/1985} \code{3/19/1986} \code{3/19/1987} \code{3/19/1990} \code{3/19/1991} \code{3/2/1984} \code{3/2/1987} \code{3/2/1988} \code{3/2/1989} \code{3/2/1990} \code{3/20/1984} \code{3/20/1985} \code{3/20/1986} \code{3/20/1987} \code{3/20/1989} \code{3/20/1990} \code{3/20/1991} \code{3/21/1984} \code{3/21/1985} \code{3/21/1986} \code{3/21/1988} \code{3/21/1989} \code{3/21/1990} \code{3/21/1991} \code{3/22/1984} \code{3/22/1985} \code{3/22/1988} \code{3/22/1989} \code{3/22/1990} \code{3/22/1991} \code{3/23/1984} \code{3/23/1987} \code{3/23/1988} \code{3/23/1989} \code{3/23/1990} \code{3/24/1986} \code{3/24/1987} \code{3/24/1988} \code{3/25/1985} \code{3/25/1986} \code{3/25/1987} \code{3/25/1988} \code{3/25/1991} \code{3/26/1984} \code{3/26/1985} \code{3/26/1986} \code{3/26/1987} \code{3/26/1990} \code{3/26/1991} \code{3/27/1984} \code{3/27/1985} \code{3/27/1986} \code{3/27/1987} \code{3/27/1989} \code{3/27/1990} \code{3/27/1991} \code{3/28/1984} \code{3/28/1985} \code{3/28/1988} \code{3/28/1989} \code{3/28/1990} \code{3/28/1991} \code{3/29/1984} \code{3/29/1985} \code{3/29/1988} \code{3/29/1989} \code{3/29/1990} \code{3/3/1986} \code{3/3/1987} \code{3/3/1988} \code{3/3/1989} \code{3/30/1984} \code{3/30/1987} \code{3/30/1988} \code{3/30/1989} \code{3/30/1990} \code{3/31/1986} \code{3/31/1987} \code{3/31/1988} \code{3/31/1989} \code{3/4/1985} \code{3/4/1986} \code{3/4/1987} \code{3/4/1988} \code{3/4/1991} \code{3/5/1984} \code{3/5/1985} \code{3/5/1986} \code{3/5/1987} \code{3/5/1990} \code{3/5/1991} \code{3/6/1984} \code{3/6/1985} \code{3/6/1986} \code{3/6/1987} \code{3/6/1989} \code{3/6/1990} \code{3/6/1991} \code{3/7/1984} \code{3/7/1985} \code{3/7/1986} \code{3/7/1988} \code{3/7/1989} \code{3/7/1990} \code{3/7/1991} \code{3/8/1984} \code{3/8/1985} \code{3/8/1988} \code{3/8/1989} \code{3/8/1990} \code{3/8/1991} \code{3/9/1984} \code{3/9/1987} \code{3/9/1988} \code{3/9/1989} \code{3/9/1990} \code{4/1/1985} \code{4/1/1986} \code{4/1/1987} \code{4/1/1991} \code{4/10/1984} \code{4/10/1985} \code{4/10/1986} \code{4/10/1987} \code{4/10/1989} \code{4/10/1990} \code{4/10/1991} \code{4/11/1984} \code{4/11/1985} \code{4/11/1986} \code{4/11/1988} \code{4/11/1989} \code{4/11/1990} \code{4/11/1991} \code{4/12/1984} \code{4/12/1985} \code{4/12/1988} \code{4/12/1989} \code{4/12/1990} \code{4/12/1991} \code{4/13/1984} \code{4/13/1987} \code{4/13/1988} \code{4/13/1989} \code{4/14/1986} \code{4/14/1987} \code{4/14/1988} \code{4/14/1989} \code{4/15/1985} \code{4/15/1986} \code{4/15/1987} \code{4/15/1988} \code{4/15/1991} \code{4/16/1984} \code{4/16/1985} \code{4/16/1986} \code{4/16/1987} \code{4/16/1990} \code{4/16/1991} \code{4/17/1984} \code{4/17/1985} \code{4/17/1986} \code{4/17/1989} \code{4/17/1990} \code{4/17/1991} \code{4/18/1984} \code{4/18/1985} \code{4/18/1986} \code{4/18/1988} \code{4/18/1989} \code{4/18/1990} \code{4/18/1991} \code{4/19/1984} \code{4/19/1985} \code{4/19/1988} \code{4/19/1989} \code{4/19/1990} \code{4/19/1991} \code{4/2/1984} \code{4/2/1985} \code{4/2/1986} \code{4/2/1987} \code{4/2/1990} \code{4/2/1991} \code{4/20/1987} \code{4/20/1988} \code{4/20/1989} \code{4/20/1990} \code{4/21/1986} \code{4/21/1987} \code{4/21/1988} \code{4/21/1989} \code{4/22/1985} \code{4/22/1986} \code{4/22/1987} \code{4/22/1988} \code{4/22/1991} \code{4/23/1984} \code{4/23/1985} \code{4/23/1986} \code{4/23/1987} \code{4/23/1990} \code{4/23/1991} \code{4/24/1984} \code{4/24/1985} \code{4/24/1986} \code{4/24/1987} \code{4/24/1989} \code{4/24/1990} \code{4/24/1991} \code{4/25/1984} \code{4/25/1985} \code{4/25/1986} \code{4/25/1988} \code{4/25/1989} \code{4/25/1990} \code{4/25/1991} \code{4/26/1984} \code{4/26/1985} \code{4/26/1988} \code{4/26/1989} \code{4/26/1990} \code{4/26/1991} \code{4/27/1984} \code{4/27/1987} \code{4/27/1988} \code{4/27/1989} \code{4/27/1990} \code{4/28/1986} \code{4/28/1987} \code{4/28/1988} \code{4/28/1989} \code{4/29/1985} \code{4/29/1986} \code{4/29/1987} \code{4/29/1988} \code{4/29/1991} \code{4/3/1984} \code{4/3/1985} \code{4/3/1986} \code{4/3/1987} \code{4/3/1989} \code{4/3/1990} \code{4/3/1991} \code{4/30/1984} \code{4/30/1985} \code{4/30/1986} \code{4/30/1987} \code{4/30/1990} \code{4/30/1991} \code{4/4/1984} \code{4/4/1985} \code{4/4/1986} \code{4/4/1988} \code{4/4/1989} \code{4/4/1990} \code{4/4/1991} \code{4/5/1984} \code{4/5/1988} \code{4/5/1989} \code{4/5/1990} \code{4/5/1991} \code{4/6/1984} \code{4/6/1987} \code{4/6/1988} \code{4/6/1989} \code{4/6/1990} \code{4/7/1986} \code{4/7/1987} \code{4/7/1988} \code{4/7/1989} \code{4/8/1985} \code{4/8/1986} \code{4/8/1987} \code{4/8/1988} \code{4/8/1991} \code{4/9/1984} \code{4/9/1985} \code{4/9/1986} \code{4/9/1987} \code{4/9/1990} \code{4/9/1991} \code{5/1/1984} \code{5/1/1985} \code{5/1/1986} \code{5/1/1987} \code{5/1/1989} \code{5/1/1990} \code{5/1/1991} \code{5/10/1984} \code{5/10/1985} \code{5/10/1988} \code{5/10/1989} \code{5/10/1990} \code{5/10/1991} \code{5/11/1984} \code{5/11/1987} \code{5/11/1988} \code{5/11/1989} \code{5/11/1990} \code{5/12/1986} \code{5/12/1987} \code{5/12/1988} \code{5/12/1989} \code{5/13/1985} \code{5/13/1986} \code{5/13/1987} \code{5/13/1988} \code{5/13/1991} \code{5/14/1984} \code{5/14/1985} \code{5/14/1986} \code{5/14/1987} \code{5/14/1990} \code{5/14/1991} \code{5/15/1984} \code{5/15/1985} \code{5/15/1986} \code{5/15/1987} \code{5/15/1989} \code{5/15/1990} \code{5/15/1991} \code{5/16/1984} \code{5/16/1985} \code{5/16/1986} \code{5/16/1988} \code{5/16/1989} \code{5/16/1990} \code{5/16/1991} \code{5/17/1984} \code{5/17/1985} \code{5/17/1988} \code{5/17/1989} \code{5/17/1990} \code{5/17/1991} \code{5/18/1984} \code{5/18/1987} \code{5/18/1988} \code{5/18/1989} \code{5/18/1990} \code{5/19/1986} \code{5/19/1987} \code{5/19/1988} \code{5/19/1989} \code{5/2/1984} \code{5/2/1985} \code{5/2/1986} \code{5/2/1988} \code{5/2/1989} \code{5/2/1990} \code{5/2/1991} \code{5/20/1985} \code{5/20/1986} \code{5/20/1987} \code{5/20/1988} \code{5/20/1991} \code{5/21/1984} \code{5/21/1985} \code{5/21/1986} \code{5/21/1987} \code{5/21/1990} \code{5/21/1991} \code{5/22/1984} \code{5/22/1985} \code{5/22/1986} \code{5/22/1987} \code{5/22/1989} \code{5/22/1990} \code{5/22/1991} \code{5/23/1984} \code{5/23/1985} \code{5/23/1986} \code{5/23/1988} \code{5/23/1989} \code{5/23/1990} \code{5/23/1991} \code{5/24/1984} \code{5/24/1985} \code{5/24/1988} \code{5/24/1989} \code{5/24/1990} \code{5/24/1991} \code{5/25/1984} \code{5/25/1988} \code{5/25/1989} \code{5/25/1990} \code{5/26/1987} \code{5/26/1988} \code{5/26/1989} \code{5/27/1986} \code{5/27/1987} \code{5/27/1988} \code{5/28/1985} \code{5/28/1986} \code{5/28/1987} \code{5/28/1991} \code{5/29/1984} \code{5/29/1985} \code{5/29/1986} \code{5/29/1987} \code{5/29/1990} \code{5/29/1991} \code{5/3/1984} \code{5/3/1985} \code{5/3/1988} \code{5/3/1989} \code{5/3/1990} \code{5/3/1991} \code{5/30/1984} \code{5/30/1985} \code{5/30/1986} \code{5/30/1989} \code{5/30/1990} \code{5/30/1991} \code{5/31/1984} \code{5/31/1985} \code{5/31/1988} \code{5/31/1989} \code{5/31/1990} \code{5/31/1991} \code{5/4/1984} \code{5/4/1987} \code{5/4/1988} \code{5/4/1989} \code{5/4/1990} \code{5/5/1986} \code{5/5/1987} \code{5/5/1988} \code{5/5/1989} \code{5/6/1985} \code{5/6/1986} \code{5/6/1987} \code{5/6/1988} \code{5/6/1991} \code{5/7/1984} \code{5/7/1985} \code{5/7/1986} \code{5/7/1987} \code{5/7/1990} \code{5/7/1991} \code{5/8/1984} \code{5/8/1985} \code{5/8/1986} \code{5/8/1987} \code{5/8/1989} \code{5/8/1990} \code{5/8/1991} \code{5/9/1984} \code{5/9/1985} \code{5/9/1986} \code{5/9/1988} \code{5/9/1989} \code{5/9/1990} \code{5/9/1991} \code{6/1/1984} \code{6/1/1987} \code{6/1/1988} \code{6/1/1989} \code{6/1/1990} \code{6/10/1985} \code{6/10/1986} \code{6/10/1987} \code{6/10/1988} \code{6/10/1991} \code{6/11/1984} \code{6/11/1985} \code{6/11/1986} \code{6/11/1987} \code{6/11/1990} \code{6/11/1991} \code{6/12/1984} \code{6/12/1985} \code{6/12/1986} \code{6/12/1987} \code{6/12/1989} \code{6/12/1990} \code{6/12/1991} \code{6/13/1984} \code{6/13/1985} \code{6/13/1986} \code{6/13/1988} \code{6/13/1989} \code{6/13/1990} \code{6/13/1991} \code{6/14/1984} \code{6/14/1985} \code{6/14/1988} \code{6/14/1989} \code{6/14/1990} \code{6/14/1991} \code{6/15/1984} \code{6/15/1987} \code{6/15/1988} \code{6/15/1989} \code{6/15/1990} \code{6/16/1986} \code{6/16/1987} \code{6/16/1988} \code{6/16/1989} \code{6/17/1985} \code{6/17/1986} \code{6/17/1987} \code{6/17/1988} \code{6/17/1991} \code{6/18/1984} \code{6/18/1985} \code{6/18/1986} \code{6/18/1987} \code{6/18/1990} \code{6/18/1991} \code{6/19/1984} \code{6/19/1985} \code{6/19/1986} \code{6/19/1987} \code{6/19/1989} \code{6/19/1990} \code{6/19/1991} \code{6/2/1986} \code{6/2/1987} \code{6/2/1988} \code{6/2/1989} \code{6/20/1984} \code{6/20/1985} \code{6/20/1986} \code{6/20/1988} \code{6/20/1989} \code{6/20/1990} \code{6/20/1991} \code{6/21/1984} \code{6/21/1985} \code{6/21/1988} \code{6/21/1989} \code{6/21/1990} \code{6/21/1991} \code{6/22/1984} \code{6/22/1987} \code{6/22/1988} \code{6/22/1989} \code{6/22/1990} \code{6/23/1986} \code{6/23/1987} \code{6/23/1988} \code{6/23/1989} \code{6/24/1985} \code{6/24/1986} \code{6/24/1987} \code{6/24/1988} \code{6/24/1991} \code{6/25/1984} \code{6/25/1985} \code{6/25/1986} \code{6/25/1987} \code{6/25/1990} \code{6/25/1991} \code{6/26/1984} \code{6/26/1985} \code{6/26/1986} \code{6/26/1987} \code{6/26/1989} \code{6/26/1990} \code{6/26/1991} \code{6/27/1984} \code{6/27/1985} \code{6/27/1986} \code{6/27/1988} \code{6/27/1989} \code{6/27/1990} \code{6/27/1991} \code{6/28/1984} \code{6/28/1985} \code{6/28/1988} \code{6/28/1989} \code{6/28/1990} \code{6/28/1991} \code{6/29/1984} \code{6/29/1987} \code{6/29/1988} \code{6/29/1989} \code{6/29/1990} \code{6/3/1985} \code{6/3/1986} \code{6/3/1987} \code{6/3/1988} \code{6/3/1991} \code{6/30/1986} \code{6/30/1987} \code{6/30/1988} \code{6/30/1989} \code{6/4/1984} \code{6/4/1985} \code{6/4/1986} \code{6/4/1987} \code{6/4/1990} \code{6/4/1991} \code{6/5/1984} \code{6/5/1985} \code{6/5/1986} \code{6/5/1987} \code{6/5/1989} \code{6/5/1990} \code{6/5/1991} \code{6/6/1984} \code{6/6/1985} \code{6/6/1986} \code{6/6/1988} \code{6/6/1989} \code{6/6/1990} \code{6/6/1991} \code{6/7/1984} \code{6/7/1985} \code{6/7/1988} \code{6/7/1989} \code{6/7/1990} \code{6/7/1991} \code{6/8/1984} \code{6/8/1987} \code{6/8/1988} \code{6/8/1989} \code{6/8/1990} \code{6/9/1986} \code{6/9/1987} \code{6/9/1988} \code{6/9/1989} \code{7/1/1985} \code{7/1/1986} \code{7/1/1987} \code{7/1/1988} \code{7/1/1991} \code{7/10/1984} \code{7/10/1985} \code{7/10/1986} \code{7/10/1987} \code{7/10/1989} \code{7/10/1990} \code{7/10/1991} \code{7/11/1984} \code{7/11/1985} \code{7/11/1986} \code{7/11/1988} \code{7/11/1989} \code{7/11/1990} \code{7/11/1991} \code{7/12/1984} \code{7/12/1985} \code{7/12/1988} \code{7/12/1989} \code{7/12/1990} \code{7/12/1991} \code{7/13/1984} \code{7/13/1987} \code{7/13/1988} \code{7/13/1989} \code{7/13/1990} \code{7/14/1986} \code{7/14/1987} \code{7/14/1988} \code{7/14/1989} \code{7/15/1985} \code{7/15/1986} \code{7/15/1987} \code{7/15/1988} \code{7/15/1991} \code{7/16/1984} \code{7/16/1985} \code{7/16/1986} \code{7/16/1987} \code{7/16/1990} \code{7/16/1991} \code{7/17/1984} \code{7/17/1985} \code{7/17/1986} \code{7/17/1987} \code{7/17/1989} \code{7/17/1990} \code{7/17/1991} \code{7/18/1984} \code{7/18/1985} \code{7/18/1986} \code{7/18/1988} \code{7/18/1989} \code{7/18/1990} \code{7/18/1991} \code{7/19/1984} \code{7/19/1985} \code{7/19/1988} \code{7/19/1989} \code{7/19/1990} \code{7/19/1991} \code{7/2/1984} \code{7/2/1985} \code{7/2/1986} \code{7/2/1987} \code{7/2/1990} \code{7/2/1991} \code{7/20/1984} \code{7/20/1987} \code{7/20/1988} \code{7/20/1989} \code{7/20/1990} \code{7/21/1986} \code{7/21/1987} \code{7/21/1988} \code{7/21/1989} \code{7/22/1985} \code{7/22/1986} \code{7/22/1987} \code{7/22/1988} \code{7/22/1991} \code{7/23/1984} \code{7/23/1985} \code{7/23/1986} \code{7/23/1987} \code{7/23/1990} \code{7/23/1991} \code{7/24/1984} \code{7/24/1985} \code{7/24/1986} \code{7/24/1987} \code{7/24/1989} \code{7/24/1990} \code{7/24/1991} \code{7/25/1984} \code{7/25/1985} \code{7/25/1986} \code{7/25/1988} \code{7/25/1989} \code{7/25/1990} \code{7/25/1991} \code{7/26/1984} \code{7/26/1985} \code{7/26/1988} \code{7/26/1989} \code{7/26/1990} \code{7/26/1991} \code{7/27/1984} \code{7/27/1987} \code{7/27/1988} \code{7/27/1989} \code{7/27/1990} \code{7/28/1986} \code{7/28/1987} \code{7/28/1988} \code{7/28/1989} \code{7/29/1985} \code{7/29/1986} \code{7/29/1987} \code{7/29/1988} \code{7/29/1991} \code{7/3/1984} \code{7/3/1985} \code{7/3/1986} \code{7/3/1989} \code{7/3/1990} \code{7/3/1991} \code{7/30/1984} \code{7/30/1985} \code{7/30/1986} \code{7/30/1987} \code{7/30/1990} \code{7/30/1991} \code{7/31/1984} \code{7/31/1985} \code{7/31/1986} \code{7/31/1987} \code{7/31/1989} \code{7/31/1990} \code{7/31/1991} \code{7/5/1984} \code{7/5/1985} \code{7/5/1988} \code{7/5/1989} \code{7/5/1990} \code{7/5/1991} \code{7/6/1984} \code{7/6/1987} \code{7/6/1988} \code{7/6/1989} \code{7/6/1990} \code{7/7/1986} \code{7/7/1987} \code{7/7/1988} \code{7/7/1989} \code{7/8/1985} \code{7/8/1986} \code{7/8/1987} \code{7/8/1988} \code{7/8/1991} \code{7/9/1984} \code{7/9/1985} \code{7/9/1986} \code{7/9/1987} \code{7/9/1990} \code{7/9/1991} \code{8/1/1984} \code{8/1/1985} \code{8/1/1986} \code{8/1/1988} \code{8/1/1989} \code{8/1/1990} \code{8/1/1991} \code{8/10/1984} \code{8/10/1987} \code{8/10/1988} \code{8/10/1989} \code{8/10/1990} \code{8/11/1986} \code{8/11/1987} \code{8/11/1988} \code{8/11/1989} \code{8/12/1985} \code{8/12/1986} \code{8/12/1987} \code{8/12/1988} \code{8/12/1991} \code{8/13/1984} \code{8/13/1985} \code{8/13/1986} \code{8/13/1987} \code{8/13/1990} \code{8/13/1991} \code{8/14/1984} \code{8/14/1985} \code{8/14/1986} \code{8/14/1987} \code{8/14/1989} \code{8/14/1990} \code{8/14/1991} \code{8/15/1984} \code{8/15/1985} \code{8/15/1986} \code{8/15/1988} \code{8/15/1989} \code{8/15/1990} \code{8/15/1991} \code{8/16/1984} \code{8/16/1985} \code{8/16/1988} \code{8/16/1989} \code{8/16/1990} \code{8/16/1991} \code{8/17/1984} \code{8/17/1987} \code{8/17/1988} \code{8/17/1989} \code{8/17/1990} \code{8/18/1986} \code{8/18/1987} \code{8/18/1988} \code{8/18/1989} \code{8/19/1985} \code{8/19/1986} \code{8/19/1987} \code{8/19/1988} \code{8/19/1991} \code{8/2/1984} \code{8/2/1985} \code{8/2/1988} \code{8/2/1989} \code{8/2/1990} \code{8/2/1991} \code{8/20/1984} \code{8/20/1985} \code{8/20/1986} \code{8/20/1987} \code{8/20/1990} \code{8/20/1991} \code{8/21/1984} \code{8/21/1985} \code{8/21/1986} \code{8/21/1987} \code{8/21/1989} \code{8/21/1990} \code{8/21/1991} \code{8/22/1984} \code{8/22/1985} \code{8/22/1986} \code{8/22/1988} \code{8/22/1989} \code{8/22/1990} \code{8/22/1991} \code{8/23/1984} \code{8/23/1985} \code{8/23/1988} \code{8/23/1989} \code{8/23/1990} \code{8/23/1991} \code{8/24/1984} \code{8/24/1987} \code{8/24/1988} \code{8/24/1989} \code{8/24/1990} \code{8/25/1986} \code{8/25/1987} \code{8/25/1988} \code{8/25/1989} \code{8/26/1985} \code{8/26/1986} \code{8/26/1987} \code{8/26/1988} \code{8/26/1991} \code{8/27/1984} \code{8/27/1985} \code{8/27/1986} \code{8/27/1987} \code{8/27/1990} \code{8/27/1991} \code{8/28/1984} \code{8/28/1985} \code{8/28/1986} \code{8/28/1987} \code{8/28/1989} \code{8/28/1990} \code{8/28/1991} \code{8/29/1984} \code{8/29/1985} \code{8/29/1986} \code{8/29/1988} \code{8/29/1989} \code{8/29/1990} \code{8/29/1991} \code{8/3/1984} \code{8/3/1987} \code{8/3/1988} \code{8/3/1989} \code{8/3/1990} \code{8/30/1984} \code{8/30/1985} \code{8/30/1988} \code{8/30/1989} \code{8/30/1990} \code{8/30/1991} \code{8/31/1984} \code{8/31/1987} \code{8/31/1988} \code{8/31/1989} \code{8/31/1990} \code{8/4/1986} \code{8/4/1987} \code{8/4/1988} \code{8/4/1989} \code{8/5/1985} \code{8/5/1986} \code{8/5/1987} \code{8/5/1988} \code{8/5/1991} \code{8/6/1984} \code{8/6/1985} \code{8/6/1986} \code{8/6/1987} \code{8/6/1990} \code{8/6/1991} \code{8/7/1984} \code{8/7/1985} \code{8/7/1986} \code{8/7/1987} \code{8/7/1989} \code{8/7/1990} \code{8/7/1991} \code{8/8/1984} \code{8/8/1985} \code{8/8/1986} \code{8/8/1988} \code{8/8/1989} \code{8/8/1990} \code{8/8/1991} \code{8/9/1984} \code{8/9/1985} \code{8/9/1988} \code{8/9/1989} \code{8/9/1990} \code{8/9/1991} \code{9/1/1987} \code{9/1/1988} \code{9/1/1989} \code{9/10/1984} \code{9/10/1985} \code{9/10/1986} \code{9/10/1987} \code{9/10/1990} \code{9/10/1991} \code{9/11/1984} \code{9/11/1985} \code{9/11/1986} \code{9/11/1987} \code{9/11/1989} \code{9/11/1990} \code{9/11/1991} \code{9/12/1984} \code{9/12/1985} \code{9/12/1986} \code{9/12/1988} \code{9/12/1989} \code{9/12/1990} \code{9/12/1991} \code{9/13/1984} \code{9/13/1985} \code{9/13/1988} \code{9/13/1989} \code{9/13/1990} \code{9/13/1991} \code{9/14/1984} \code{9/14/1987} \code{9/14/1988} \code{9/14/1989} \code{9/14/1990} \code{9/15/1986} \code{9/15/1987} \code{9/15/1988} \code{9/15/1989} \code{9/16/1985} \code{9/16/1986} \code{9/16/1987} \code{9/16/1988} \code{9/16/1991} \code{9/17/1984} \code{9/17/1985} \code{9/17/1986} \code{9/17/1987} \code{9/17/1990} \code{9/17/1991} \code{9/18/1984} \code{9/18/1985} \code{9/18/1986} \code{9/18/1987} \code{9/18/1989} \code{9/18/1990} \code{9/18/1991} \code{9/19/1984} \code{9/19/1985} \code{9/19/1986} \code{9/19/1988} \code{9/19/1989} \code{9/19/1990} \code{9/19/1991} \code{9/2/1986} \code{9/2/1987} \code{9/2/1988} \code{9/20/1984} \code{9/20/1985} \code{9/20/1988} \code{9/20/1989} \code{9/20/1990} \code{9/20/1991} \code{9/21/1984} \code{9/21/1987} \code{9/21/1988} \code{9/21/1989} \code{9/21/1990} \code{9/22/1986} \code{9/22/1987} \code{9/22/1988} \code{9/22/1989} \code{9/23/1985} \code{9/23/1986} \code{9/23/1987} \code{9/23/1988} \code{9/23/1991} \code{9/24/1984} \code{9/24/1985} \code{9/24/1986} \code{9/24/1987} \code{9/24/1990} \code{9/24/1991} \code{9/25/1984} \code{9/25/1985} \code{9/25/1986} \code{9/25/1987} \code{9/25/1989} \code{9/25/1990} \code{9/25/1991} \code{9/26/1984} \code{9/26/1985} \code{9/26/1986} \code{9/26/1988} \code{9/26/1989} \code{9/26/1990} \code{9/26/1991} \code{9/27/1984} \code{9/27/1988} \code{9/27/1989} \code{9/27/1990} \code{9/27/1991} \code{9/28/1984} \code{9/28/1987} \code{9/28/1988} \code{9/28/1989} \code{9/28/1990} \code{9/29/1986} \code{9/29/1987} \code{9/29/1988} \code{9/29/1989} \code{9/3/1985} \code{9/3/1986} \code{9/3/1987} \code{9/3/1991} \code{9/30/1985} \code{9/30/1986} \code{9/30/1987} \code{9/30/1988} \code{9/30/1991} \code{9/4/1984} \code{9/4/1985} \code{9/4/1986} \code{9/4/1987} \code{9/4/1990} \code{9/4/1991} \code{9/5/1984} \code{9/5/1985} \code{9/5/1986} \code{9/5/1989} \code{9/5/1990} \code{9/5/1991} \code{9/6/1984} \code{9/6/1985} \code{9/6/1988} \code{9/6/1989} \code{9/6/1990} \code{9/6/1991} \code{9/7/1984} \code{9/7/1988} \code{9/7/1989} \code{9/7/1990} \code{9/8/1986} \code{9/8/1987} \code{9/8/1988} \code{9/8/1989} \code{9/9/1985} \code{9/9/1986} \code{9/9/1987} \code{9/9/1988} \code{9/9/1991}}
    \item{\code{HP}}{a numeric vector}
  }
}
\source{
S+ FinMetrics hp.s.
}
\references{
S+ FinMetrics hp.s.
}
\examples{
data(hp.s)
## maybe str(hp.s) ; plot(hp.s) ...
}
\keyword{datasets}
