% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{import}
\alias{import}
\title{Monthly Argentinian imports and import taxes from Januaray 1983 to December 1990}
\format{A \samp{xts} object on Jan 1983/Dec 1990 containing:
\describe{
  \item{taxes}{a numeric vector}
  \item{imports}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
import
}
\description{
A \samp{xts} object of monthly Argentinian imports and import taxes from Januaray 1983 to
December 1990.
}
\keyword{datasets}

