% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IP}
\alias{IP}
\title{Monthly observations on industrial production growth from January 1919 to November 2001}
\format{A \samp{xts} object on Jan 1919/Nov 2001 containing:
\describe{
  \item{IP}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
IP
}
\description{
A \samp{xts} object of monthly observations on industrial production growth from January 1919 to
November 2001.
}
\keyword{datasets}

