\name{johansen.danish}
\alias{johansen.danish}
\docType{data}
\title{
monthly Danish data used in Johansen (1995)
}
\description{
monthly Danish data used in Johansen (1995)
}
\usage{data(johansen.danish)}
\format{
  A data frame with 55 observations on the following 6 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{Apr 1974} \code{Apr 1975} \code{Apr 1976} \code{Apr 1977} \code{Apr 1978} \code{Apr 1979} \code{Apr 1980} \code{Apr 1981} \code{Apr 1982} \code{Apr 1983} \code{Apr 1984} \code{Apr 1985} \code{Apr 1986} \code{Feb 1974} \code{Feb 1975} \code{Feb 1976} \code{Feb 1977} \code{Feb 1978} \code{Feb 1979} \code{Feb 1980} \code{Feb 1981} \code{Feb 1982} \code{Feb 1983} \code{Feb 1984} \code{Feb 1985} \code{Feb 1986} \code{Feb 1987} \code{Jan 1974} \code{Jan 1975} \code{Jan 1976} \code{Jan 1977} \code{Jan 1978} \code{Jan 1979} \code{Jan 1980} \code{Jan 1981} \code{Jan 1982} \code{Jan 1983} \code{Jan 1984} \code{Jan 1985} \code{Jan 1986} \code{Jan 1987} \code{Mar 1974} \code{Mar 1975} \code{Mar 1976} \code{Mar 1977} \code{Mar 1978} \code{Mar 1979} \code{Mar 1980} \code{Mar 1981} \code{Mar 1982} \code{Mar 1983} \code{Mar 1984} \code{Mar 1985} \code{Mar 1986} \code{Mar 1987}}
    \item{\code{LRM}}{a numeric vector}
    \item{\code{LRY}}{a numeric vector}
    \item{\code{LPY}}{a numeric vector}
    \item{\code{IBO}}{a numeric vector}
    \item{\code{IDE}}{a numeric vector}
  }
}
\source{
S+ FinMetrics johansen.danish.
}
\references{
Johansen (1995)
}
\examples{
data(johansen.danish)
## maybe str(johansen.danish) ; plot(johansen.danish) ...
}
\keyword{datasets}
