% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{johansen}
\alias{johansen}
\title{Monthly returns on Danish stocks from January 1974 to March 1987 taken from Johansen (1995)}
\format{A \samp{xts} object on Jan 1974/Mar 1987 containing:
\describe{
  \item{IBO}{a numeric vector}
  \item{IDE}{a numeric vector}
  \item{LPY}{a numeric vector}
  \item{LRM}{a numeric vector}
  \item{LRY}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
johansen
}
\description{
A \samp{xts} object of monthly returns on Danish stocks from January 1974 to March 1987.
}
\keyword{datasets}

