\name{lexrates.dat}
\alias{lexrates.dat}
\docType{data}
\title{
the 30-day US and other 13 countries annually interest rate differential
}
\description{
the 30-day US and other 13 countries annually interest rate differential from Apr 1976 to Sep 1995  
}
\usage{data(lexrates.dat)}
\format{
  A data frame with 245 observations on the following 13 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{Apr 1976} \code{Apr 1977} \code{Apr 1978} \code{Apr 1979} \code{Apr 1980} \code{Apr 1981} \code{Apr 1982} \code{Apr 1983} \code{Apr 1984} \code{Apr 1985} \code{Apr 1986} \code{Apr 1987} \code{Apr 1988} \code{Apr 1989} \code{Apr 1990} \code{Apr 1991} \code{Apr 1992} \code{Apr 1993} \code{Apr 1994} \code{Apr 1995} \code{Apr 1996} \code{Aug 1976} \code{Aug 1977} \code{Aug 1978} \code{Aug 1979} \code{Aug 1980} \code{Aug 1981} \code{Aug 1982} \code{Aug 1983} \code{Aug 1984} \code{Aug 1985} \code{Aug 1986} \code{Aug 1987} \code{Aug 1988} \code{Aug 1989} \code{Aug 1990} \code{Aug 1991} \code{Aug 1992} \code{Aug 1993} \code{Aug 1994} \code{Aug 1995} \code{Dec 1976} \code{Dec 1977} \code{Dec 1978} \code{Dec 1979} \code{Dec 1980} \code{Dec 1981} \code{Dec 1982} \code{Dec 1983} \code{Dec 1984} \code{Dec 1985} \code{Dec 1986} \code{Dec 1987} \code{Dec 1988} \code{Dec 1989} \code{Dec 1990} \code{Dec 1991} \code{Dec 1992} \code{Dec 1993} \code{Dec 1994} \code{Dec 1995} \code{Feb 1976} \code{Feb 1977} \code{Feb 1978} \code{Feb 1979} \code{Feb 1980} \code{Feb 1981} \code{Feb 1982} \code{Feb 1983} \code{Feb 1984} \code{Feb 1985} \code{Feb 1986} \code{Feb 1987} \code{Feb 1988} \code{Feb 1989} \code{Feb 1990} \code{Feb 1991} \code{Feb 1992} \code{Feb 1993} \code{Feb 1994} \code{Feb 1995} \code{Feb 1996} \code{Jan 1977} \code{Jan 1978} \code{Jan 1979} \code{Jan 1980} \code{Jan 1981} \code{Jan 1982} \code{Jan 1983} \code{Jan 1984} \code{Jan 1985} \code{Jan 1986} \code{Jan 1987} \code{Jan 1988} \code{Jan 1989} \code{Jan 1990} \code{Jan 1991} \code{Jan 1992} \code{Jan 1993} \code{Jan 1994} \code{Jan 1995} \code{Jan 1996} \code{Jul 1976} \code{Jul 1977} \code{Jul 1978} \code{Jul 1979} \code{Jul 1980} \code{Jul 1981} \code{Jul 1982} \code{Jul 1983} \code{Jul 1984} \code{Jul 1985} \code{Jul 1986} \code{Jul 1987} \code{Jul 1988} \code{Jul 1989} \code{Jul 1990} \code{Jul 1991} \code{Jul 1992} \code{Jul 1993} \code{Jul 1994} \code{Jul 1995} \code{Jun 1976} \code{Jun 1977} \code{Jun 1978} \code{Jun 1979} \code{Jun 1980} \code{Jun 1981} \code{Jun 1982} \code{Jun 1983} \code{Jun 1984} \code{Jun 1985} \code{Jun 1986} \code{Jun 1987} \code{Jun 1988} \code{Jun 1989} \code{Jun 1990} \code{Jun 1991} \code{Jun 1992} \code{Jun 1993} \code{Jun 1994} \code{Jun 1995} \code{Jun 1996} \code{Mar 1976} \code{Mar 1977} \code{Mar 1978} \code{Mar 1979} \code{Mar 1980} \code{Mar 1981} \code{Mar 1982} \code{Mar 1983} \code{Mar 1984} \code{Mar 1985} \code{Mar 1986} \code{Mar 1987} \code{Mar 1988} \code{Mar 1989} \code{Mar 1990} \code{Mar 1991} \code{Mar 1992} \code{Mar 1993} \code{Mar 1994} \code{Mar 1995} \code{Mar 1996} \code{May 1976} \code{May 1977} \code{May 1978} \code{May 1979} \code{May 1980} \code{May 1981} \code{May 1982} \code{May 1983} \code{May 1984} \code{May 1985} \code{May 1986} \code{May 1987} \code{May 1988} \code{May 1989} \code{May 1990} \code{May 1991} \code{May 1992} \code{May 1993} \code{May 1994} \code{May 1995} \code{May 1996} \code{Nov 1976} \code{Nov 1977} \code{Nov 1978} \code{Nov 1979} \code{Nov 1980} \code{Nov 1981} \code{Nov 1982} \code{Nov 1983} \code{Nov 1984} \code{Nov 1985} \code{Nov 1986} \code{Nov 1987} \code{Nov 1988} \code{Nov 1989} \code{Nov 1990} \code{Nov 1991} \code{Nov 1992} \code{Nov 1993} \code{Nov 1994} \code{Nov 1995} \code{Oct 1976} \code{Oct 1977} \code{Oct 1978} \code{Oct 1979} \code{Oct 1980} \code{Oct 1981} \code{Oct 1982} \code{Oct 1983} \code{Oct 1984} \code{Oct 1985} \code{Oct 1986} \code{Oct 1987} \code{Oct 1988} \code{Oct 1989} \code{Oct 1990} \code{Oct 1991} \code{Oct 1992} \code{Oct 1993} \code{Oct 1994} \code{Oct 1995} \code{Sep 1976} \code{Sep 1977} \code{Sep 1978} \code{Sep 1979} \code{Sep 1980} \code{Sep 1981} \code{Sep 1982} \code{Sep 1983} \code{Sep 1984} \code{Sep 1985} \code{Sep 1986} \code{Sep 1987} \code{Sep 1988} \code{Sep 1989} \code{Sep 1990} \code{Sep 1991} \code{Sep 1992} \code{Sep 1993} \code{Sep 1994} \code{Sep 1995}}
    \item{\code{USCNS}}{a numeric vector}
    \item{\code{USCNF}}{a numeric vector}
    \item{\code{USDMS}}{a numeric vector}
    \item{\code{USDMF}}{a numeric vector}
    \item{\code{USFRS}}{a numeric vector}
    \item{\code{USFRF}}{a numeric vector}
    \item{\code{USILS}}{a numeric vector}
    \item{\code{USILF}}{a numeric vector}
    \item{\code{USJYS}}{a numeric vector}
    \item{\code{USJYF}}{a numeric vector}
    \item{\code{USUKS}}{a numeric vector}
    \item{\code{USUKF}}{a numeric vector}
  }
}
\source{
S+ FinMetrics lexrates.dat.
}
\references{
S+ FinMetrics lexrates.dat.
}
\examples{
data(lexrates.dat)
## maybe str(lexrates.dat) ; plot(lexrates.dat) ...
}
\keyword{datasets}
