% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexrates}
\alias{lexrates}
\title{Monthly annual interest rate differential between US and other countries from February 1976 and
June 1996}
\format{A \samp{xts} object on Jan 1974/Mar 1987 containing:
\describe{
  \item{USCNF}{a numeric vector}
  \item{USCNS}{a numeric vector}
  \item{USDMF}{a numeric vector}
  \item{USDMS}{a numeric vector}
  \item{USFRF}{a numeric vector}
  \item{USFRS}{a numeric vector}
  \item{USILF}{a numeric vector}
  \item{USILS}{a numeric vector}
  \item{USJYF}{a numeric vector}
  \item{USJYS}{a numeric vector}
  \item{USUKF}{a numeric vector}
  \item{USUKS}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
lexrates
}
\description{
A \samp{xts} object of monthly annual interest rate differential between US and other countries
from February 1976 and June 1996
}
\keyword{datasets}

