% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maturity}
\alias{maturity}
\title{Cooresponding maturity of U.S. zero coupon rates and forward rates}
\format{A \samp{data.frame} object of 55 oberservations of 1 variable on the cooresponding
maturity of U.S. zero coupon rates and forward rates:
\describe{
  \item{dat}{a numeric vector}
}}
\source{
S+ FinMetrics
}
\usage{
maturity
}
\description{
A \samp{xts} object of monthly forward rates of U.S. coupons from August 1985 to February 1991
computed by McCulloch and Kwon (1993).
}
\keyword{datasets}

