% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msft}
\alias{msft}
\title{Daily open, high, low, close and volume on Microsoft stock from September 2000 to September 2001}
\format{A \samp{xts} object on 2000-09-27/2001-09-27 containing:
\describe{
  \item{open}{a numeric vector}
  \item{high}{a numeric vector}
  \item{low}{a numeric vector}
  \item{close}{a numeric vector}
  \item{volume}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
msft
}
\description{
A \samp{xts} object of the daily open, high, low, close and volume on Microsoft stock from
September 2000 to September 2001
}
\keyword{datasets}

