\name{ndx.dat}
\alias{ndx.dat}
\docType{data}
\title{
the daily opening, highest, lowest and closing prices of NASDAQ-100 index from January 2, 1996 to October 12, 2001
}
\description{
the daily opening, highest, lowest and closing prices of NASDAQ-100 index from January 2, 1996 to October 12, 2001
}
\usage{data(ndx.dat)}
\format{
  A data frame with 1457 observations on the following 5 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{01/02/1996} \code{01/02/1997} \code{01/02/1998} \code{01/02/2001} \code{01/03/1996} \code{01/03/1997} \code{01/03/2000} \code{01/03/2001} \code{01/04/1996} \code{01/04/1999} \code{01/04/2000} \code{01/04/2001} \code{01/05/1996} \code{01/05/1998} \code{01/05/1999} \code{01/05/2000} \code{01/05/2001} \code{01/06/1997} \code{01/06/1998} \code{01/06/1999} \code{01/06/2000} \code{01/07/1997} \code{01/07/1998} \code{01/07/1999} \code{01/07/2000} \code{01/08/1996} \code{01/08/1997} \code{01/08/1998} \code{01/08/1999} \code{01/08/2001} \code{01/09/1996} \code{01/09/1997} \code{01/09/1998} \code{01/09/2001} \code{01/10/1996} \code{01/10/1997} \code{01/10/2000} \code{01/10/2001} \code{01/11/1996} \code{01/11/1999} \code{01/11/2000} \code{01/11/2001} \code{01/12/1996} \code{01/12/1998} \code{01/12/1999} \code{01/12/2000} \code{01/12/2001} \code{01/13/1997} \code{01/13/1998} \code{01/13/1999} \code{01/13/2000} \code{01/14/1997} \code{01/14/1998} \code{01/14/1999} \code{01/14/2000} \code{01/15/1996} \code{01/15/1997} \code{01/15/1998} \code{01/15/1999} \code{01/16/1996} \code{01/16/1997} \code{01/16/1998} \code{01/16/2001} \code{01/17/1996} \code{01/17/1997} \code{01/17/2001} \code{01/18/1996} \code{01/18/2000} \code{01/18/2001} \code{01/19/1996} \code{01/19/1999} \code{01/19/2000} \code{01/19/2001} \code{01/20/1997} \code{01/20/1998} \code{01/20/1999} \code{01/20/2000} \code{01/21/1997} \code{01/21/1998} \code{01/21/1999} \code{01/21/2000} \code{01/22/1996} \code{01/22/1997} \code{01/22/1998} \code{01/22/1999} \code{01/22/2001} \code{01/23/1996} \code{01/23/1997} \code{01/23/1998} \code{01/23/2001} \code{01/24/1996} \code{01/24/1997} \code{01/24/2000} \code{01/24/2001} \code{01/25/1996} \code{01/25/1999} \code{01/25/2000} \code{01/25/2001} \code{01/26/1996} \code{01/26/1998} \code{01/26/1999} \code{01/26/2000} \code{01/26/2001} \code{01/27/1997} \code{01/27/1998} \code{01/27/1999} \code{01/27/2000} \code{01/28/1997} \code{01/28/1998} \code{01/28/1999} \code{01/28/2000} \code{01/29/1996} \code{01/29/1997} \code{01/29/1998} \code{01/29/1999} \code{01/29/2001} \code{01/30/1996} \code{01/30/1997} \code{01/30/1998} \code{01/30/2001} \code{01/31/1996} \code{01/31/1997} \code{01/31/2000} \code{01/31/2001} \code{02/01/1996} \code{02/01/1999} \code{02/01/2000} \code{02/01/2001} \code{02/02/1996} \code{02/02/1998} \code{02/02/1999} \code{02/02/2000} \code{02/02/2001} \code{02/03/1997} \code{02/03/1998} \code{02/03/1999} \code{02/03/2000} \code{02/04/1997} \code{02/04/1998} \code{02/04/1999} \code{02/04/2000} \code{02/05/1996} \code{02/05/1997} \code{02/05/1998} \code{02/05/1999} \code{02/05/2001} \code{02/06/1996} \code{02/06/1997} \code{02/06/1998} \code{02/06/2001} \code{02/07/1996} \code{02/07/1997} \code{02/07/2000} \code{02/07/2001} \code{02/08/1996} \code{02/08/1999} \code{02/08/2000} \code{02/08/2001} \code{02/09/1996} \code{02/09/1998} \code{02/09/1999} \code{02/09/2000} \code{02/09/2001} \code{02/10/1997} \code{02/10/1998} \code{02/10/1999} \code{02/10/2000} \code{02/11/1997} \code{02/11/1998} \code{02/11/1999} \code{02/11/2000} \code{02/12/1996} \code{02/12/1997} \code{02/12/1998} \code{02/12/1999} \code{02/12/2001} \code{02/13/1996} \code{02/13/1997} \code{02/13/1998} \code{02/13/2001} \code{02/14/1996} \code{02/14/1997} \code{02/14/2000} \code{02/14/2001} \code{02/15/1996} \code{02/15/2000} \code{02/15/2001} \code{02/16/1996} \code{02/16/1999} \code{02/16/2000} \code{02/16/2001} \code{02/17/1998} \code{02/17/1999} \code{02/17/2000} \code{02/18/1997} \code{02/18/1998} \code{02/18/1999} \code{02/18/2000} \code{02/19/1997} \code{02/19/1998} \code{02/19/1999} \code{02/20/1996} \code{02/20/1997} \code{02/20/1998} \code{02/20/2001} \code{02/21/1996} \code{02/21/1997} \code{02/21/2001} \code{02/22/1996} \code{02/22/1999} \code{02/22/2000} \code{02/22/2001} \code{02/23/1996} \code{02/23/1998} \code{02/23/1999} \code{02/23/2000} \code{02/23/2001} \code{02/24/1997} \code{02/24/1998} \code{02/24/1999} \code{02/24/2000} \code{02/25/1997} \code{02/25/1998} \code{02/25/1999} \code{02/25/2000} \code{02/26/1996} \code{02/26/1997} \code{02/26/1998} \code{02/26/1999} \code{02/26/2001} \code{02/27/1996} \code{02/27/1997} \code{02/27/1998} \code{02/27/2001} \code{02/28/1996} \code{02/28/1997} \code{02/28/2000} \code{02/28/2001} \code{02/29/1996} \code{02/29/2000} \code{03/01/1996} \code{03/01/1999} \code{03/01/2000} \code{03/01/2001} \code{03/02/1998} \code{03/02/1999} \code{03/02/2000} \code{03/02/2001} \code{03/03/1997} \code{03/03/1998} \code{03/03/1999} \code{03/03/2000} \code{03/04/1996} \code{03/04/1997} \code{03/04/1998} \code{03/04/1999} \code{03/05/1996} \code{03/05/1997} \code{03/05/1998} \code{03/05/1999} \code{03/05/2001} \code{03/06/1996} \code{03/06/1997} \code{03/06/1998} \code{03/06/2000} \code{03/06/2001} \code{03/07/1996} \code{03/07/1997} \code{03/07/2000} \code{03/07/2001} \code{03/08/1996} \code{03/08/1999} \code{03/08/2000} \code{03/08/2001} \code{03/09/1998} \code{03/09/1999} \code{03/09/2000} \code{03/09/2001} \code{03/10/1997} \code{03/10/1998} \code{03/10/1999} \code{03/10/2000} \code{03/11/1996} \code{03/11/1997} \code{03/11/1998} \code{03/11/1999} \code{03/12/1996} \code{03/12/1997} \code{03/12/1998} \code{03/12/1999} \code{03/12/2001} \code{03/13/1996} \code{03/13/1997} \code{03/13/1998} \code{03/13/2000} \code{03/13/2001} \code{03/14/1996} \code{03/14/1997} \code{03/14/2000} \code{03/14/2001} \code{03/15/1996} \code{03/15/1999} \code{03/15/2000} \code{03/15/2001} \code{03/16/1998} \code{03/16/1999} \code{03/16/2000} \code{03/16/2001} \code{03/17/1997} \code{03/17/1998} \code{03/17/1999} \code{03/17/2000} \code{03/18/1996} \code{03/18/1997} \code{03/18/1998} \code{03/18/1999} \code{03/19/1996} \code{03/19/1997} \code{03/19/1998} \code{03/19/1999} \code{03/19/2001} \code{03/20/1996} \code{03/20/1997} \code{03/20/1998} \code{03/20/2000} \code{03/20/2001} \code{03/21/1996} \code{03/21/1997} \code{03/21/2000} \code{03/21/2001} \code{03/22/1996} \code{03/22/1999} \code{03/22/2000} \code{03/22/2001} \code{03/23/1998} \code{03/23/1999} \code{03/23/2000} \code{03/23/2001} \code{03/24/1997} \code{03/24/1998} \code{03/24/1999} \code{03/24/2000} \code{03/25/1996} \code{03/25/1997} \code{03/25/1998} \code{03/25/1999} \code{03/26/1996} \code{03/26/1997} \code{03/26/1998} \code{03/26/1999} \code{03/26/2001} \code{03/27/1996} \code{03/27/1997} \code{03/27/1998} \code{03/27/2000} \code{03/27/2001} \code{03/28/1996} \code{03/28/2000} \code{03/28/2001} \code{03/29/1996} \code{03/29/1999} \code{03/29/2000} \code{03/29/2001} \code{03/30/1998} \code{03/30/1999} \code{03/30/2000} \code{03/30/2001} \code{03/31/1997} \code{03/31/1998} \code{03/31/1999} \code{03/31/2000} \code{04/01/1996} \code{04/01/1997} \code{04/01/1998} \code{04/01/1999} \code{04/02/1996} \code{04/02/1997} \code{04/02/1998} \code{04/02/2001} \code{04/03/1996} \code{04/03/1997} \code{04/03/1998} \code{04/03/2000} \code{04/03/2001} \code{04/04/1996} \code{04/04/1997} \code{04/04/2000} \code{04/04/2001} \code{04/05/1999} \code{04/05/2000} \code{04/05/2001} \code{04/06/1998} \code{04/06/1999} \code{04/06/2000} \code{04/06/2001} \code{04/07/1997} \code{04/07/1998} \code{04/07/1999} \code{04/07/2000} \code{04/08/1996} \code{04/08/1997} \code{04/08/1998} \code{04/08/1999} \code{04/09/1996} \code{04/09/1997} \code{04/09/1998} \code{04/09/1999} \code{04/09/2001} \code{04/10/1996} \code{04/10/1997} \code{04/10/2000} \code{04/10/2001} \code{04/11/1996} \code{04/11/1997} \code{04/11/2000} \code{04/11/2001} \code{04/12/1996} \code{04/12/1999} \code{04/12/2000} \code{04/12/2001} \code{04/13/1998} \code{04/13/1999} \code{04/13/2000} \code{04/14/1997} \code{04/14/1998} \code{04/14/1999} \code{04/14/2000} \code{04/15/1996} \code{04/15/1997} \code{04/15/1998} \code{04/15/1999} \code{04/16/1996} \code{04/16/1997} \code{04/16/1998} \code{04/16/1999} \code{04/16/2001} \code{04/17/1996} \code{04/17/1997} \code{04/17/1998} \code{04/17/2000} \code{04/17/2001} \code{04/18/1996} \code{04/18/1997} \code{04/18/2000} \code{04/18/2001} \code{04/19/1996} \code{04/19/1999} \code{04/19/2000} \code{04/19/2001} \code{04/20/1998} \code{04/20/1999} \code{04/20/2000} \code{04/20/2001} \code{04/21/1997} \code{04/21/1998} \code{04/21/1999} \code{04/22/1996} \code{04/22/1997} \code{04/22/1998} \code{04/22/1999} \code{04/23/1996} \code{04/23/1997} \code{04/23/1998} \code{04/23/1999} \code{04/23/2001} \code{04/24/1996} \code{04/24/1997} \code{04/24/1998} \code{04/24/2000} \code{04/24/2001} \code{04/25/1996} \code{04/25/1997} \code{04/25/2000} \code{04/25/2001} \code{04/26/1996} \code{04/26/1999} \code{04/26/2000} \code{04/26/2001} \code{04/27/1998} \code{04/27/1999} \code{04/27/2000} \code{04/27/2001} \code{04/28/1997} \code{04/28/1998} \code{04/28/1999} \code{04/28/2000} \code{04/29/1996} \code{04/29/1997} \code{04/29/1998} \code{04/29/1999} \code{04/30/1996} \code{04/30/1997} \code{04/30/1998} \code{04/30/1999} \code{04/30/2001} \code{05/01/1996} \code{05/01/1997} \code{05/01/1998} \code{05/01/2000} \code{05/01/2001} \code{05/02/1996} \code{05/02/1997} \code{05/02/2000} \code{05/02/2001} \code{05/03/1996} \code{05/03/1999} \code{05/03/2000} \code{05/03/2001} \code{05/04/1998} \code{05/04/1999} \code{05/04/2000} \code{05/04/2001} \code{05/05/1997} \code{05/05/1998} \code{05/05/1999} \code{05/05/2000} \code{05/06/1996} \code{05/06/1997} \code{05/06/1998} \code{05/06/1999} \code{05/07/1996} \code{05/07/1997} \code{05/07/1998} \code{05/07/1999} \code{05/07/2001} \code{05/08/1996} \code{05/08/1997} \code{05/08/1998} \code{05/08/2000} \code{05/08/2001} \code{05/09/1996} \code{05/09/1997} \code{05/09/2000} \code{05/09/2001} \code{05/10/1996} \code{05/10/1999} \code{05/10/2000} \code{05/10/2001} \code{05/11/1998} \code{05/11/1999} \code{05/11/2000} \code{05/11/2001} \code{05/12/1997} \code{05/12/1998} \code{05/12/1999} \code{05/12/2000} \code{05/13/1996} \code{05/13/1997} \code{05/13/1998} \code{05/13/1999} \code{05/14/1996} \code{05/14/1997} \code{05/14/1998} \code{05/14/1999} \code{05/14/2001} \code{05/15/1996} \code{05/15/1997} \code{05/15/1998} \code{05/15/2000} \code{05/15/2001} \code{05/16/1996} \code{05/16/1997} \code{05/16/2000} \code{05/16/2001} \code{05/17/1996} \code{05/17/1999} \code{05/17/2000} \code{05/17/2001} \code{05/18/1998} \code{05/18/1999} \code{05/18/2000} \code{05/18/2001} \code{05/19/1997} \code{05/19/1998} \code{05/19/1999} \code{05/19/2000} \code{05/20/1996} \code{05/20/1997} \code{05/20/1998} \code{05/20/1999} \code{05/21/1996} \code{05/21/1997} \code{05/21/1998} \code{05/21/1999} \code{05/21/2001} \code{05/22/1996} \code{05/22/1997} \code{05/22/1998} \code{05/22/2000} \code{05/22/2001} \code{05/23/1996} \code{05/23/1997} \code{05/23/2000} \code{05/23/2001} \code{05/24/1996} \code{05/24/1999} \code{05/24/2000} \code{05/24/2001} \code{05/25/1999} \code{05/25/2000} \code{05/25/2001} \code{05/26/1998} \code{05/26/1999} \code{05/26/2000} \code{05/27/1997} \code{05/27/1998} \code{05/27/1999} \code{05/28/1996} \code{05/28/1997} \code{05/28/1998} \code{05/28/1999} \code{05/29/1996} \code{05/29/1997} \code{05/29/1998} \code{05/29/2001} \code{05/30/1996} \code{05/30/1997} \code{05/30/2000} \code{05/30/2001} \code{05/31/1996} \code{05/31/2000} \code{05/31/2001} \code{06/01/1998} \code{06/01/1999} \code{06/01/2000} \code{06/01/2001} \code{06/02/1997} \code{06/02/1998} \code{06/02/1999} \code{06/02/2000} \code{06/03/1996} \code{06/03/1997} \code{06/03/1998} \code{06/03/1999} \code{06/04/1996} \code{06/04/1997} \code{06/04/1998} \code{06/04/1999} \code{06/04/2001} \code{06/05/1996} \code{06/05/1997} \code{06/05/1998} \code{06/05/2000} \code{06/05/2001} \code{06/06/1996} \code{06/06/1997} \code{06/06/2000} \code{06/06/2001} \code{06/07/1996} \code{06/07/1999} \code{06/07/2000} \code{06/07/2001} \code{06/08/1998} \code{06/08/1999} \code{06/08/2000} \code{06/08/2001} \code{06/09/1997} \code{06/09/1998} \code{06/09/1999} \code{06/09/2000} \code{06/10/1996} \code{06/10/1997} \code{06/10/1998} \code{06/10/1999} \code{06/11/1996} \code{06/11/1997} \code{06/11/1998} \code{06/11/1999} \code{06/11/2001} \code{06/12/1996} \code{06/12/1997} \code{06/12/1998} \code{06/12/2000} \code{06/12/2001} \code{06/13/1996} \code{06/13/1997} \code{06/13/2000} \code{06/13/2001} \code{06/14/1996} \code{06/14/1999} \code{06/14/2000} \code{06/14/2001} \code{06/15/1998} \code{06/15/1999} \code{06/15/2000} \code{06/15/2001} \code{06/16/1997} \code{06/16/1998} \code{06/16/1999} \code{06/16/2000} \code{06/17/1996} \code{06/17/1997} \code{06/17/1998} \code{06/17/1999} \code{06/18/1996} \code{06/18/1997} \code{06/18/1998} \code{06/18/1999} \code{06/18/2001} \code{06/19/1996} \code{06/19/1997} \code{06/19/1998} \code{06/19/2000} \code{06/19/2001} \code{06/20/1996} \code{06/20/1997} \code{06/20/2000} \code{06/20/2001} \code{06/21/1996} \code{06/21/1999} \code{06/21/2000} \code{06/21/2001} \code{06/22/1998} \code{06/22/1999} \code{06/22/2000} \code{06/22/2001} \code{06/23/1997} \code{06/23/1998} \code{06/23/1999} \code{06/23/2000} \code{06/24/1996} \code{06/24/1997} \code{06/24/1998} \code{06/24/1999} \code{06/25/1996} \code{06/25/1997} \code{06/25/1998} \code{06/25/1999} \code{06/25/2001} \code{06/26/1996} \code{06/26/1997} \code{06/26/1998} \code{06/26/2000} \code{06/26/2001} \code{06/27/1996} \code{06/27/1997} \code{06/27/2000} \code{06/27/2001} \code{06/28/1996} \code{06/28/1999} \code{06/28/2000} \code{06/28/2001} \code{06/29/1998} \code{06/29/1999} \code{06/29/2000} \code{06/29/2001} \code{06/30/1997} \code{06/30/1998} \code{06/30/1999} \code{06/30/2000} \code{07/01/1996} \code{07/01/1997} \code{07/01/1998} \code{07/01/1999} \code{07/02/1996} \code{07/02/1997} \code{07/02/1998} \code{07/02/1999} \code{07/02/2001} \code{07/03/1996} \code{07/03/1997} \code{07/03/2000} \code{07/03/2001} \code{07/05/1996} \code{07/05/2000} \code{07/05/2001} \code{07/06/1998} \code{07/06/1999} \code{07/06/2000} \code{07/06/2001} \code{07/07/1997} \code{07/07/1998} \code{07/07/1999} \code{07/07/2000} \code{07/08/1996} \code{07/08/1997} \code{07/08/1998} \code{07/08/1999} \code{07/09/1996} \code{07/09/1997} \code{07/09/1998} \code{07/09/1999} \code{07/09/2001} \code{07/10/1996} \code{07/10/1997} \code{07/10/1998} \code{07/10/2000} \code{07/10/2001} \code{07/11/1996} \code{07/11/1997} \code{07/11/2000} \code{07/11/2001} \code{07/12/1996} \code{07/12/1999} \code{07/12/2000} \code{07/12/2001} \code{07/13/1998} \code{07/13/1999} \code{07/13/2000} \code{07/13/2001} \code{07/14/1997} \code{07/14/1998} \code{07/14/1999} \code{07/14/2000} \code{07/15/1996} \code{07/15/1997} \code{07/15/1998} \code{07/15/1999} \code{07/16/1996} \code{07/16/1997} \code{07/16/1998} \code{07/16/1999} \code{07/16/2001} \code{07/17/1996} \code{07/17/1997} \code{07/17/1998} \code{07/17/2000} \code{07/17/2001} \code{07/18/1996} \code{07/18/1997} \code{07/18/2000} \code{07/18/2001} \code{07/19/1996} \code{07/19/1999} \code{07/19/2000} \code{07/19/2001} \code{07/20/1998} \code{07/20/1999} \code{07/20/2000} \code{07/20/2001} \code{07/21/1997} \code{07/21/1998} \code{07/21/1999} \code{07/21/2000} \code{07/22/1996} \code{07/22/1997} \code{07/22/1998} \code{07/22/1999} \code{07/23/1996} \code{07/23/1997} \code{07/23/1998} \code{07/23/1999} \code{07/23/2001} \code{07/24/1996} \code{07/24/1997} \code{07/24/1998} \code{07/24/2000} \code{07/24/2001} \code{07/25/1996} \code{07/25/1997} \code{07/25/2000} \code{07/25/2001} \code{07/26/1996} \code{07/26/1999} \code{07/26/2000} \code{07/26/2001} \code{07/27/1998} \code{07/27/1999} \code{07/27/2000} \code{07/27/2001} \code{07/28/1997} \code{07/28/1998} \code{07/28/1999} \code{07/28/2000} \code{07/29/1996} \code{07/29/1997} \code{07/29/1998} \code{07/29/1999} \code{07/30/1996} \code{07/30/1997} \code{07/30/1998} \code{07/30/1999} \code{07/30/2001} \code{07/31/1996} \code{07/31/1997} \code{07/31/1998} \code{07/31/2000} \code{07/31/2001} \code{08/01/1996} \code{08/01/1997} \code{08/01/2000} \code{08/01/2001} \code{08/02/1996} \code{08/02/1999} \code{08/02/2000} \code{08/02/2001} \code{08/03/1998} \code{08/03/1999} \code{08/03/2000} \code{08/03/2001} \code{08/04/1997} \code{08/04/1998} \code{08/04/1999} \code{08/04/2000} \code{08/05/1996} \code{08/05/1997} \code{08/05/1998} \code{08/05/1999} \code{08/06/1996} \code{08/06/1997} \code{08/06/1998} \code{08/06/1999} \code{08/06/2001} \code{08/07/1996} \code{08/07/1997} \code{08/07/1998} \code{08/07/2000} \code{08/07/2001} \code{08/08/1996} \code{08/08/1997} \code{08/08/2000} \code{08/08/2001} \code{08/09/1996} \code{08/09/1999} \code{08/09/2000} \code{08/09/2001} \code{08/10/1998} \code{08/10/1999} \code{08/10/2000} \code{08/10/2001} \code{08/11/1997} \code{08/11/1998} \code{08/11/1999} \code{08/11/2000} \code{08/12/1996} \code{08/12/1997} \code{08/12/1998} \code{08/12/1999} \code{08/13/1996} \code{08/13/1997} \code{08/13/1998} \code{08/13/1999} \code{08/13/2001} \code{08/14/1996} \code{08/14/1997} \code{08/14/1998} \code{08/14/2000} \code{08/14/2001} \code{08/15/1996} \code{08/15/1997} \code{08/15/2000} \code{08/15/2001} \code{08/16/1996} \code{08/16/1999} \code{08/16/2000} \code{08/16/2001} \code{08/17/1998} \code{08/17/1999} \code{08/17/2000} \code{08/17/2001} \code{08/18/1997} \code{08/18/1998} \code{08/18/1999} \code{08/18/2000} \code{08/19/1996} \code{08/19/1997} \code{08/19/1998} \code{08/19/1999} \code{08/20/1996} \code{08/20/1997} \code{08/20/1998} \code{08/20/1999} \code{08/20/2001} \code{08/21/1996} \code{08/21/1997} \code{08/21/1998} \code{08/21/2000} \code{08/21/2001} \code{08/22/1996} \code{08/22/1997} \code{08/22/2000} \code{08/22/2001} \code{08/23/1996} \code{08/23/1999} \code{08/23/2000} \code{08/23/2001} \code{08/24/1998} \code{08/24/1999} \code{08/24/2000} \code{08/24/2001} \code{08/25/1997} \code{08/25/1998} \code{08/25/1999} \code{08/25/2000} \code{08/26/1996} \code{08/26/1997} \code{08/26/1998} \code{08/26/1999} \code{08/27/1996} \code{08/27/1997} \code{08/27/1998} \code{08/27/1999} \code{08/27/2001} \code{08/28/1996} \code{08/28/1997} \code{08/28/1998} \code{08/28/2000} \code{08/28/2001} \code{08/29/1996} \code{08/29/1997} \code{08/29/2000} \code{08/29/2001} \code{08/30/1996} \code{08/30/1999} \code{08/30/2000} \code{08/30/2001} \code{08/31/1998} \code{08/31/1999} \code{08/31/2000} \code{08/31/2001} \code{09/01/1998} \code{09/01/1999} \code{09/01/2000} \code{09/02/1997} \code{09/02/1998} \code{09/02/1999} \code{09/03/1996} \code{09/03/1997} \code{09/03/1998} \code{09/03/1999} \code{09/04/1996} \code{09/04/1997} \code{09/04/1998} \code{09/04/2001} \code{09/05/1996} \code{09/05/1997} \code{09/05/2000} \code{09/05/2001} \code{09/06/1996} \code{09/06/2000} \code{09/06/2001} \code{09/07/1999} \code{09/07/2000} \code{09/07/2001} \code{09/08/1997} \code{09/08/1998} \code{09/08/1999} \code{09/08/2000} \code{09/09/1996} \code{09/09/1997} \code{09/09/1998} \code{09/09/1999} \code{09/10/1996} \code{09/10/1997} \code{09/10/1998} \code{09/10/1999} \code{09/10/2001} \code{09/11/1996} \code{09/11/1997} \code{09/11/1998} \code{09/11/2000} \code{09/12/1996} \code{09/12/1997} \code{09/12/2000} \code{09/13/1996} \code{09/13/1999} \code{09/13/2000} \code{09/14/1998} \code{09/14/1999} \code{09/14/2000} \code{09/15/1997} \code{09/15/1998} \code{09/15/1999} \code{09/15/2000} \code{09/16/1996} \code{09/16/1997} \code{09/16/1998} \code{09/16/1999} \code{09/17/1996} \code{09/17/1997} \code{09/17/1998} \code{09/17/1999} \code{09/17/2001} \code{09/18/1996} \code{09/18/1997} \code{09/18/1998} \code{09/18/2000} \code{09/18/2001} \code{09/19/1996} \code{09/19/1997} \code{09/19/2000} \code{09/19/2001} \code{09/20/1996} \code{09/20/1999} \code{09/20/2000} \code{09/20/2001} \code{09/21/1998} \code{09/21/1999} \code{09/21/2000} \code{09/21/2001} \code{09/22/1997} \code{09/22/1998} \code{09/22/1999} \code{09/22/2000} \code{09/23/1996} \code{09/23/1997} \code{09/23/1998} \code{09/23/1999} \code{09/24/1996} \code{09/24/1997} \code{09/24/1998} \code{09/24/1999} \code{09/24/2001} \code{09/25/1996} \code{09/25/1997} \code{09/25/1998} \code{09/25/2000} \code{09/25/2001} \code{09/26/1996} \code{09/26/1997} \code{09/26/2000} \code{09/26/2001} \code{09/27/1996} \code{09/27/1999} \code{09/27/2000} \code{09/27/2001} \code{09/28/1998} \code{09/28/1999} \code{09/28/2000} \code{09/28/2001} \code{09/29/1997} \code{09/29/1998} \code{09/29/1999} \code{09/29/2000} \code{09/30/1996} \code{09/30/1997} \code{09/30/1998} \code{09/30/1999} \code{10/01/1996} \code{10/01/1997} \code{10/01/1998} \code{10/01/1999} \code{10/01/2001} \code{10/02/1996} \code{10/02/1997} \code{10/02/1998} \code{10/02/2000} \code{10/02/2001} \code{10/03/1996} \code{10/03/1997} \code{10/03/2000} \code{10/03/2001} \code{10/04/1996} \code{10/04/1999} \code{10/04/2000} \code{10/04/2001} \code{10/05/1998} \code{10/05/1999} \code{10/05/2000} \code{10/05/2001} \code{10/06/1997} \code{10/06/1998} \code{10/06/1999} \code{10/06/2000} \code{10/07/1996} \code{10/07/1997} \code{10/07/1998} \code{10/07/1999} \code{10/08/1996} \code{10/08/1997} \code{10/08/1998} \code{10/08/1999} \code{10/08/2001} \code{10/09/1996} \code{10/09/1997} \code{10/09/1998} \code{10/09/2000} \code{10/09/2001} \code{10/10/1996} \code{10/10/1997} \code{10/10/2000} \code{10/10/2001} \code{10/11/1996} \code{10/11/1999} \code{10/11/2000} \code{10/11/2001} \code{10/12/1998} \code{10/12/1999} \code{10/12/2000} \code{10/12/2001} \code{10/13/1997} \code{10/13/1998} \code{10/13/1999} \code{10/13/2000} \code{10/14/1996} \code{10/14/1997} \code{10/14/1998} \code{10/14/1999} \code{10/15/1996} \code{10/15/1997} \code{10/15/1998} \code{10/15/1999} \code{10/16/1996} \code{10/16/1997} \code{10/16/1998} \code{10/16/2000} \code{10/17/1996} \code{10/17/1997} \code{10/17/2000} \code{10/18/1996} \code{10/18/1999} \code{10/18/2000} \code{10/19/1998} \code{10/19/1999} \code{10/19/2000} \code{10/20/1997} \code{10/20/1998} \code{10/20/1999} \code{10/20/2000} \code{10/21/1996} \code{10/21/1997} \code{10/21/1998} \code{10/21/1999} \code{10/22/1996} \code{10/22/1997} \code{10/22/1998} \code{10/22/1999} \code{10/23/1996} \code{10/23/1997} \code{10/23/1998} \code{10/23/2000} \code{10/24/1996} \code{10/24/1997} \code{10/24/2000} \code{10/25/1996} \code{10/25/1999} \code{10/25/2000} \code{10/26/1998} \code{10/26/1999} \code{10/26/2000} \code{10/27/1997} \code{10/27/1998} \code{10/27/1999} \code{10/27/2000} \code{10/28/1996} \code{10/28/1997} \code{10/28/1998} \code{10/28/1999} \code{10/29/1996} \code{10/29/1997} \code{10/29/1998} \code{10/29/1999} \code{10/30/1996} \code{10/30/1997} \code{10/30/1998} \code{10/30/2000} \code{10/31/1996} \code{10/31/1997} \code{10/31/2000} \code{11/01/1996} \code{11/01/1999} \code{11/01/2000} \code{11/02/1998} \code{11/02/1999} \code{11/02/2000} \code{11/03/1997} \code{11/03/1998} \code{11/03/1999} \code{11/03/2000} \code{11/04/1996} \code{11/04/1997} \code{11/04/1998} \code{11/04/1999} \code{11/05/1996} \code{11/05/1997} \code{11/05/1998} \code{11/05/1999} \code{11/06/1996} \code{11/06/1997} \code{11/06/1998} \code{11/06/2000} \code{11/07/1996} \code{11/07/1997} \code{11/07/2000} \code{11/08/1996} \code{11/08/1999} \code{11/08/2000} \code{11/09/1998} \code{11/09/1999} \code{11/09/2000} \code{11/10/1997} \code{11/10/1998} \code{11/10/1999} \code{11/10/2000} \code{11/11/1996} \code{11/11/1997} \code{11/11/1998} \code{11/11/1999} \code{11/12/1996} \code{11/12/1997} \code{11/12/1998} \code{11/12/1999} \code{11/13/1996} \code{11/13/1997} \code{11/13/1998} \code{11/13/2000} \code{11/14/1996} \code{11/14/1997} \code{11/14/2000} \code{11/15/1996} \code{11/15/1999} \code{11/15/2000} \code{11/16/1998} \code{11/16/1999} \code{11/16/2000} \code{11/17/1997} \code{11/17/1998} \code{11/17/1999} \code{11/17/2000} \code{11/18/1996} \code{11/18/1997} \code{11/18/1998} \code{11/18/1999} \code{11/19/1996} \code{11/19/1997} \code{11/19/1998} \code{11/19/1999} \code{11/20/1996} \code{11/20/1997} \code{11/20/1998} \code{11/20/2000} \code{11/21/1996} \code{11/21/1997} \code{11/21/2000} \code{11/22/1996} \code{11/22/1999} \code{11/22/2000} \code{11/23/1998} \code{11/23/1999} \code{11/24/1997} \code{11/24/1998} \code{11/24/1999} \code{11/24/2000} \code{11/25/1996} \code{11/25/1997} \code{11/25/1998} \code{11/26/1996} \code{11/26/1997} \code{11/26/1999} \code{11/27/1996} \code{11/27/1998} \code{11/27/2000} \code{11/28/1997} \code{11/28/2000} \code{11/29/1996} \code{11/29/1999} \code{11/29/2000} \code{11/30/1998} \code{11/30/1999} \code{11/30/2000} \code{12/01/1997} \code{12/01/1998} \code{12/01/1999} \code{12/01/2000} \code{12/02/1996} \code{12/02/1997} \code{12/02/1998} \code{12/02/1999} \code{12/03/1996} \code{12/03/1997} \code{12/03/1998} \code{12/03/1999} \code{12/04/1996} \code{12/04/1997} \code{12/04/1998} \code{12/04/2000} \code{12/05/1996} \code{12/05/1997} \code{12/05/2000} \code{12/06/1996} \code{12/06/1999} \code{12/06/2000} \code{12/07/1998} \code{12/07/1999} \code{12/07/2000} \code{12/08/1997} \code{12/08/1998} \code{12/08/1999} \code{12/08/2000} \code{12/09/1996} \code{12/09/1997} \code{12/09/1998} \code{12/09/1999} \code{12/10/1996} \code{12/10/1997} \code{12/10/1998} \code{12/10/1999} \code{12/11/1996} \code{12/11/1997} \code{12/11/1998} \code{12/11/2000} \code{12/12/1996} \code{12/12/1997} \code{12/12/2000} \code{12/13/1996} \code{12/13/1999} \code{12/13/2000} \code{12/14/1998} \code{12/14/1999} \code{12/14/2000} \code{12/15/1997} \code{12/15/1998} \code{12/15/1999} \code{12/15/2000} \code{12/16/1996} \code{12/16/1997} \code{12/16/1998} \code{12/16/1999} \code{12/17/1996} \code{12/17/1997} \code{12/17/1998} \code{12/17/1999} \code{12/18/1996} \code{12/18/1997} \code{12/18/1998} \code{12/18/2000} \code{12/19/1996} \code{12/19/1997} \code{12/19/2000} \code{12/20/1996} \code{12/20/1999} \code{12/20/2000} \code{12/21/1998} \code{12/21/1999} \code{12/21/2000} \code{12/22/1997} \code{12/22/1998} \code{12/22/1999} \code{12/22/2000} \code{12/23/1996} \code{12/23/1997} \code{12/23/1998} \code{12/23/1999} \code{12/24/1996} \code{12/24/1997} \code{12/24/1998} \code{12/26/1996} \code{12/26/1997} \code{12/26/2000} \code{12/27/1996} \code{12/27/1999} \code{12/27/2000} \code{12/28/1998} \code{12/28/1999} \code{12/28/2000} \code{12/29/1997} \code{12/29/1998} \code{12/29/1999} \code{12/29/2000} \code{12/30/1996} \code{12/30/1997} \code{12/30/1998} \code{12/30/1999} \code{12/31/1996} \code{12/31/1997} \code{12/31/1998} \code{12/31/1999}}
    \item{\code{Open}}{a numeric vector}
    \item{\code{High}}{a numeric vector}
    \item{\code{Low}}{a numeric vector}
    \item{\code{Close}}{a numeric vector}
  }
}
\source{
S+ FinMetrics ndx.dat.
}
\references{
S+ FinMetrics ndx.dat.
}
\examples{
data(ndx.dat)
## maybe str(ndx.dat) ; plot(ndx.dat) ...
}
\keyword{datasets}
