% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndx}
\alias{ndx}
\title{Daily open, high, low, close and volume on the NASDAQ 100 from January 1996 to October 2001}
\format{A \samp{xts} object on 1996-01-02/2001-10-12 containing:
\describe{
  \item{open}{a numeric vector}
  \item{high}{a numeric vector}
  \item{low}{a numeric vector}
  \item{close}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
ndx
}
\description{
A \samp{xts} object of the daily open, high, low, close and volume on the NASDAQ 100 from January
1996 to October 2001.
}
\keyword{datasets}

