% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newtaxes}
\alias{newtaxes}
\title{newtaxes}
\format{A \samp{data.frame} object with 10 observations of 1 variable:
\describe{
  \item{taxes}{a numeric vector}
}}
\source{
S+ FinMetrics
}
\usage{
newtaxes
}
\description{
A \samp{data.frame} object of newtaxes.
}
\keyword{datasets}

