% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyse}
\alias{nyse}
\title{Daily returns on a value weighted New York Stock Exchange Index from February 1984 to December
1991}
\format{A \samp{xts} object on 1984-02-02/1991-12-31 containing:
\describe{
  \item{nyse}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
nyse
}
\description{
A \samp{xts} object of daily returns on a value weighted New York Stock Exchange Index from
February 1984 to December 1991.
}
\keyword{datasets}

