% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policy}
\alias{policy}
\title{Monthly data on U.S. real GDP and the Federal Funds rate from January 1959 to March 1998}
\format{A \samp{xts} object on Jan 1959/Mar 1998 containing:
\describe{
  \item{CP}{a numeric vector}
  \item{CPI}{a numeric vector}
  \item{FFR}{a numeric vector}
  \item{GDP}{a numeric vector}
  \item{M2}{a numeric vector}
  \item{U}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
policy
}
\description{
A \samp{xts} object of monthly data on U.S. real GDP and the Federal Funds rate from January 1959
to March 1998.
}
\keyword{datasets}

