% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pricing}
\alias{pricing}
\title{Monthly data on ten size-based portfolios from the Center for Research in Security Prices (CRSP)
database from February 1959 to November 1993}
\format{A \samp{xts} object on Feb 1959/Nov 1993 containing:
\describe{
  \item{CONS}{a numeric vector}
  \item{R1}{a numeric vector}
  \item{R2}{a numeric vector}
  \item{R3}{a numeric vector}
  \item{R4}{a numeric vector}
  \item{R5}{a numeric vector}
  \item{R6}{a numeric vector}
  \item{R7}{a numeric vector}
  \item{R8}{a numeric vector}
  \item{R9}{a numeric vector}
  \item{R10}{a numeric vector}
  \item{RF}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
pricing
}
\description{
A \samp{xts} object of monthly data on ten size-based portfolios from the Center for Research in
Security Prices (CRSP) database from February 1959 to November 1993. That is, portfolio 1
contains the monthly returns on the smallest 10 percent of firms (by market capitalization)
listed on the New York Stock Exchange, and portfolio 10 contains the returns on the largest 10
percent of firms. The risk-free asset is the monthly return on 3-month U.S. T-bills, and real
consumption is measured by total U.S. personal consumption expenditures on nondurables and
services.
}
\keyword{datasets}

