% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printewmaVol.R
\name{print.ewmaVol}
\alias{print.ewmaVol}
\title{print.ewmavol()}
\usage{
\method{print}{ewmaVol}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{additional arguments passed to \samp{print()}}
}
\value{
invisible(object)
}
\description{
print.ewmavol()
}
\examples{
\dontrun{
print.ewmavol(r)
}
}
\author{
Eric Zivot
}

