% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residualsewmaVol.R
\name{residuals.ewmaVol}
\alias{residuals.ewmaVol}
\title{residuals.ewmavol()}
\usage{
\method{residuals}{ewmaVol}(object, standardize = FALSE)
}
\arguments{
\item{object}{object}

\item{standardize}{standardize=FALSE}
}
\value{
object$returns/object$sigma
}
\description{
residuals.ewmavol()
}
\examples{
\dontrun{
residuals.ewmavol(r)
}
}
\author{
Eric Zivot
}

