% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf30}
\alias{rf30}
\title{Monthly data on the annual nominal interest rate of the 30 day T-bill from July 1926 to
December 2000}
\format{A \samp{xts} object on Jul 1926/Dec 2000 containing:
\describe{
  \item{rf}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
rf30
}
\description{
A \samp{xts} object of monthly data on the annual nominal interest rate of the 30 day T-bill from
July 1926 to December 2000.
}
\keyword{datasets}

