\name{shiller.annual}
\alias{shiller.annual}
\docType{data}
\title{
The annual S&P 500 dividend information over the period 1871 - 2000
}
\description{
The annual S&P 500 dividend information over the period 1871 - 2000
}
\usage{data(shiller.annual)}
\format{
  A data frame with 130 observations on the following 11 variables.
  \describe{
    \item{\code{RowNames}}{a numeric vector}
    \item{\code{price}}{a numeric vector}
    \item{\code{dividend}}{a numeric vector}
    \item{\code{earnings}}{a numeric vector}
    \item{\code{cpi}}{a numeric vector}
    \item{\code{real.price}}{a numeric vector}
    \item{\code{real.dividend}}{a numeric vector}
    \item{\code{real.earnings}}{a numeric vector}
    \item{\code{pe.10}}{a numeric vector}
    \item{\code{dp.ratio}}{a numeric vector}
    \item{\code{dp.yield}}{a numeric vector}
  }
}
\source{
S+ FinMetrics shiller.annual.
}
\references{
S+ FinMetrics shiller.annual.
}
\examples{
data(shiller.annual)
## maybe str(shiller.annual) ; plot(shiller.annual) ...
}
\keyword{datasets}
