% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shiller}
\alias{shiller}
\title{Monthly S&P 500 dividend information from 1871 to 2000}
\format{A \samp{xts} object on Jan 1871/Mar 2001 containing:
\describe{
  \item{cpi}{a numeric vector}
  \item{dividend}{a numeric vector}
  \item{earnings}{a numeric vector}
  \item{pe.10}{a numeric vector}
  \item{price}{a numeric vector}
  \item{real.dividend}{a numeric vector}
  \item{real.earnings}{a numeric vector}
  \item{real.price}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
shiller
}
\description{
A \samp{xts} object of monthly S&P 500 dividend information from 1871 to 2000.
}
\keyword{datasets}

