% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shillerAnnual}
\alias{shillerAnnual}
\title{Yearly S&P 500 dividend information from 1871 to 2000}
\format{A \samp{xts} object on Jan 1871/Jan 2000 containing:
\describe{
  \item{cpi}{a numeric vector}
  \item{dividend}{a numeric vector}
  \item{dp.ratio}{a numeric vector}
  \item{dp.yield}{a numeric vector}
  \item{earnings}{a numeric vector}
  \item{pe.10}{a numeric vector}
  \item{price}{a numeric vector}
  \item{real.dividend}{a numeric vector}
  \item{real.earnings}{a numeric vector}
  \item{real.price}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
shillerAnnual
}
\description{
A \samp{xts} object of yearly S&P 500 dividend information from 1871 to 2000.
}
\keyword{datasets}

