\name{singleIndex.dat}
\alias{singleIndex.dat}
\docType{data}
\title{
monthly closing price data on Microsoft and the S&P 500 index
}
\description{
monthly closing price data on Microsoft and the S&P 500 index from Apr 1990 to Sep 2000
}
\usage{data(singleIndex.dat)}
\format{
  A data frame with 133 observations on the following 3 variables.
  \describe{
    \item{\code{RowNames}}{a factor with levels \code{Apr 1990} \code{Apr 1991} \code{Apr 1992} \code{Apr 1993} \code{Apr 1994} \code{Apr 1995} \code{Apr 1996} \code{Apr 1997} \code{Apr 1998} \code{Apr 1999} \code{Apr 2000} \code{Aug 1990} \code{Aug 1991} \code{Aug 1992} \code{Aug 1993} \code{Aug 1994} \code{Aug 1995} \code{Aug 1996} \code{Aug 1997} \code{Aug 1998} \code{Aug 1999} \code{Aug 2000} \code{Dec 1990} \code{Dec 1991} \code{Dec 1992} \code{Dec 1993} \code{Dec 1994} \code{Dec 1995} \code{Dec 1996} \code{Dec 1997} \code{Dec 1998} \code{Dec 1999} \code{Dec 2000} \code{Feb 1990} \code{Feb 1991} \code{Feb 1992} \code{Feb 1993} \code{Feb 1994} \code{Feb 1995} \code{Feb 1996} \code{Feb 1997} \code{Feb 1998} \code{Feb 1999} \code{Feb 2000} \code{Jan 1990} \code{Jan 1991} \code{Jan 1992} \code{Jan 1993} \code{Jan 1994} \code{Jan 1995} \code{Jan 1996} \code{Jan 1997} \code{Jan 1998} \code{Jan 1999} \code{Jan 2000} \code{Jan 2001} \code{Jul 1990} \code{Jul 1991} \code{Jul 1992} \code{Jul 1993} \code{Jul 1994} \code{Jul 1995} \code{Jul 1996} \code{Jul 1997} \code{Jul 1998} \code{Jul 1999} \code{Jul 2000} \code{Jun 1990} \code{Jun 1991} \code{Jun 1992} \code{Jun 1993} \code{Jun 1994} \code{Jun 1995} \code{Jun 1996} \code{Jun 1997} \code{Jun 1998} \code{Jun 1999} \code{Jun 2000} \code{Mar 1990} \code{Mar 1991} \code{Mar 1992} \code{Mar 1993} \code{Mar 1994} \code{Mar 1995} \code{Mar 1996} \code{Mar 1997} \code{Mar 1998} \code{Mar 1999} \code{Mar 2000} \code{May 1990} \code{May 1991} \code{May 1992} \code{May 1993} \code{May 1994} \code{May 1995} \code{May 1996} \code{May 1997} \code{May 1998} \code{May 1999} \code{May 2000} \code{Nov 1990} \code{Nov 1991} \code{Nov 1992} \code{Nov 1993} \code{Nov 1994} \code{Nov 1995} \code{Nov 1996} \code{Nov 1997} \code{Nov 1998} \code{Nov 1999} \code{Nov 2000} \code{Oct 1990} \code{Oct 1991} \code{Oct 1992} \code{Oct 1993} \code{Oct 1994} \code{Oct 1995} \code{Oct 1996} \code{Oct 1997} \code{Oct 1998} \code{Oct 1999} \code{Oct 2000} \code{Sep 1990} \code{Sep 1991} \code{Sep 1992} \code{Sep 1993} \code{Sep 1994} \code{Sep 1995} \code{Sep 1996} \code{Sep 1997} \code{Sep 1998} \code{Sep 1999} \code{Sep 2000}}
    \item{\code{MSFT}}{a numeric vector}
    \item{\code{SP500}}{a numeric vector}
  }
}
\source{
S+ FinMetrics singleIndex.dat.
}
\references{
S+ FinMetrics singleIndex.dat.
}
\examples{
data(singleIndex.dat)
## maybe str(singleIndex.dat) ; plot(singleIndex.dat) ...
}
\keyword{datasets}
