% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{singleIndex}
\alias{singleIndex}
\title{Monthly closing price on Microsoft and the S&P 500 Index from January 1990 to January 2001}
\format{A \samp{xts} object on Jan 1990/Jan 2001 containing:
\describe{
  \item{msft}{a numeric vector}
  \item{sp500}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
singleIndex
}
\description{
A \samp{xts} object of monthly closing price on Microsoft and the S&P 500 Index from January 1990
to January 2001.
}
\keyword{datasets}

