% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp}
\alias{sp}
\title{Daily closing price of the S&P 500 Index from January 1960 to October 1987}
\format{A \samp{xts} object on 1960-01-04/1987-10-16 containing:
\describe{
  \item{close}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
sp
}
\description{
A \samp{xts} object of daily closing price of the S&P 500 Index from January 1960 to October
1987.
}
\keyword{datasets}

