% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Daily returns of the S&P 500 Index from January 1928 to August 1991}
\format{A \samp{xts} object on 1928-01-04/1991-08-30 containing:
\describe{
  \item{returns}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
sp500
}
\description{
A \samp{xts} object of daily returns of the S&P 500 Index from January 1928 to August 1991.
}
\keyword{datasets}

