\name{stack.loss}
\alias{stack.loss}
\docType{data}
\title{
stack.loss
}
\description{
stack.loss
}
\usage{data(stack.loss)}
\format{
  A data frame with 21 observations on the following variable.
  \describe{
    \item{\code{data}}{a numeric vector}
  }
}
\source{
S+ FinMetrics stack.loss.
}
\references{
S+ FinMetrics stack.loss.
}
\examples{
data(stack.loss)
## maybe str(stack.loss) ; plot(stack.loss) ...
}
\keyword{datasets}
