\name{stack.x}
\alias{stack.x}
\docType{data}
\title{
stack.x
}
\description{
stack.x
}
\usage{data(stack.x)}
\format{
  A data frame with 21 observations on the following 3 variables.
  \describe{
    \item{\code{Air.Flow}}{a numeric vector}
    \item{\code{Water.Temp}}{a numeric vector}
    \item{\code{Acid.Conc.}}{a numeric vector}
  }
}
\source{
S+ FinMetrics stack.x.
}
\references{
S+ FinMetrics stack.x.
}
\examples{
data(stack.x)
## maybe str(stack.x) ; plot(stack.x) ...
}
\keyword{datasets}
