% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stackLoss}
\alias{stackLoss}
\title{stackLoss}
\format{A \samp{data.frame} object with 21 observations of 1 variable:
\describe{
  \item{dat}{a numeric vector}
}}
\source{
S+ FinMetrics
}
\usage{
stackLoss
}
\description{
A \samp{data.frame} object of stackLoss.
}
\keyword{datasets}

