% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stackX}
\alias{stackX}
\title{stackX}
\format{A \samp{data.frame} object with 21 observations of 3 variables:
\describe{
  \item{airFlow}{a numeric vector}
  \item{waterTemp}{a numeric vector}
  \item{acidCon}{a numeric vector}
}}
\source{
S+ FinMetrics
}
\usage{
stackX
}
\description{
A \samp{data.frame} object of stackX.
}
\keyword{datasets}

