% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varex}
\alias{varex}
\title{Monthly data on real stock returns, real interest rates, inflation and real output growth from
July 1926 to December 2000}
\format{A \samp{xts} object on Jul 1926/Dec 2000 containing:
\describe{
  \item{INF}{a numeric vector}
  \item{IPG}{a numeric vector}
  \item{marketReal}{a numeric vector}
  \item{rfReal}{a numeric vector}
}
Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
varex
}
\description{
A \samp{xts} object of monthly data on real stock returns, real interest rates, inflation and
real output growth from July 1926 to December 2000.
}
\keyword{datasets}

