% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yahoo}
\alias{yahoo}
\title{Daily open, high, low, close and volume on Yahoo stock for the month of February 2002}
\format{A \samp{xts} object on 2002-02-01/2002-02-28 containing:
\describe{
  \item{open}{a numeric vector}
  \item{high}{a numeric vector}
  \item{low}{a numeric vector}
  \item{close}{a numeric vector}
  \item{volume}{a numeric vector}
}
Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
yahoo
}
\description{
A \samp{xts} object of the daily open, high, low, close and volume on Yahoo stock for the month
of February 2002.
}
\keyword{datasets}

