% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeroRate}
\alias{zeroRate}
\title{Monthly zero coupon rates from August 1985 to February 1991}
\format{A \samp{xts} object on Aug 1985/Feb 1991. Indexed by objects of class: [yearmon] TZ: UTC}
\source{
S+ FinMetrics
}
\usage{
zeroRate
}
\description{
A \samp{xts} object of monthly zero coupon rates from August 1985 to February 1991 computed by
McCulloch and Kwon (1993).
}
\keyword{datasets}

