\name{danish}
\alias{danish}
\docType{data}
\title{
Danish fire loss data
}
\description{
The data in danish consist of 2167 daily insurance claims for losses exceeding one million Danish Krone from January 3, 1980 through December 31,
1990.
}
\usage{data(danish)}
\format{
  A data frame with 2167 observations on the following 2 variables.
}

\source{
S+ FinMetrics danish file.
}
\references{
McNeil (1999)
}
\examples{
data(danish)
## maybe str(danish) ; plot(danish) ...
}
\keyword{datasets}
